==> 01-project.Rmd <==
---
output: 
#  word_document: #
#    reference_docx: Ecosistemas_template.docx
  pdf_document:
    latex_engine: xelatex
    #latex_engine: lualatex
    #keep_md: TRUE
    keep_tex: TRUE
    toc_depth: 4
    number_sections: TRUE
    fig_caption: yes
    # include:
    #   after_body: tabconsis.tex
#documentclass: article #this modifies everything to 220 mm × 280 mm Letter size
classoption: a4paper
header-includes:
- \usepackage{multirow}
- \usepackage{pdflscape}
- \usepackage{afterpage}
- \usepackage{capt-of}
- \usepackage{array}
- \usepackage{color,soul}
- \usepackage{geometry}
- \usepackage[none]{hyphenat}
bibliography: malaria.bib
csl: iso690-author-date-es.csl
#UNMSM: http://crai.ub.edu/es/pmf/iso-esp
#ISO-690: numeric OR author-date? https://www.zotero.org/styles?q=iso690
link-citations: yes
linkcolor: Blue
#citecolor: Maroon
urlcolor: Blue
biblio-style: apalike
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = FALSE, warning = FALSE, message = FALSE)
```

<!--
FROM:
https://es.sharelatex.com/learn/Spanish
- \usepackage[utf8]{inputenc}
- \usepackage[spanish, es-tabla]{babel} #Cuadro (default)

BIBLIOGRAFY: (In:) to (En:) problem
https://tex.stackexchange.com/questions/321972/spanish-bibliography
https://tex.stackexchange.com/questions/238157/reference-in-spanish-and-english-language-in-bibtex
- \usepackage[english,main=spanish]{babel}

OR:
\renewcommand{\listtablename}{Índice de tablas}
\renewcommand{\figurename}{Figura}
\renewcommand{\figureautorefname}{Figure}
\renewcommand{\equationname}{Ecuación}
\renewcommand{\equationautorefname}{Ecuación}
-->
\renewcommand{\contentsname}{Índice General} 
\renewcommand{\tablename}{Tabla}
\renewcommand{\tableautorefname}{Tabla}

\pagenumbering{gobble}

```{r, child="titlepage.Rmd"}
```

\newpage
\tableofcontents

\newpage
\pagenumbering{arabic}

# RESUMEN {-}

<!-- 
## Descripción del proyecto {#resumen}
RESUMEN: CORRIGELO AL FINAL 
-->

> _Plasmodium vivax_ es responsable del 80% de la malaria en el Perú.
Casos de enfermedad severa por mono-infecciones de _P. vivax_ han sido reportados 
tanto en el noreste amazónico como en norte costero; sin embargo, dicha condición aún está
<!-- subestimada en la Norma Técnica a nivel nacional, incrementando el riesgo -->
<!-- de posibles diagnósticos tardíos o inapropiados por el personal de salud. -->
ausente en el actual Plan de Control y Eliminación, incrementando el riesgo 
de afectar la morbilidad y mortalidad por malaria en zonas endémicas.
Con el fin de 
describir la respuesta inmune e 
identificar biomarcadores de relevancia clínica contra la malaria 
<!--vivax-->severa por _P. vivax_, el presente proyecto de tesis propone
comparar la respuesta de anticuerpos ante la infección 
en pacientes severos y no-severos provenientes de un estudio prospectivo caso-control
ejecutado en la ciudad de Iquitos, Loreto - Perú, empleando microarreglos de proteínas.
Los resultados permitirán proponer marcadores serológicos <!--discriminantes--> 
de severidad<!-- o episodios previos--> 
con el potencial de implementarse en programas de 
serovigilancia en miras al control y eliminación de la malaria por _P. vivax_ en la región.
<!-- complementar las actuales medidas de control "MALARIA CERO 2017" -->

# PLANTEAMIENTO DEL PROBLEMA

## Formulación del problema {#intro}

<!-- 
## Fundamentación del problema {#intro}

INTRODUCCIÓN
Escribir la introducción en tres párrafos:
¿Qué se sabe del tema?
¿Qué no se sabe del tema?
¿Qué se va a investigar? 

JUSTIFICACIÓN
¿Qué es lo que se va a hacer con los resultados de la investigación?
-->

La malaria es una enfermedad parasitaria 
<!-- de importancia mundial, -->
causada por protozoarios del género _Plasmodium_ 
y transmitida por mosquitos del género _Anopheles_. 
En el 
<!-- año  -->
2019 
se estimaron 229 millones de casos 
en 87 países endémicos y 
409 mil muertes 
atribuidas a esta infección a nivel mundial [@who2020world]. 
<!-- Aunque  -->
<!--Si bien--> 
La especie _P. falciparum_ representa el 
97% de los casos y 
99% de las muertes,
pero
<!-- se estimó que  -->
en los continentes fuera de África
_P. vivax_ es responsable 
del 41% y 86%, respectivamente [@WHO2016world].
Más aún, la región de las Américas tiene la mayor proporción 
de casos por _P. vivax_ (72%)
donde Perú es el cuarto país con más 
casos estimados
<!-- reportes -->
<!--de malaria--> 
(n=45,729), 
<!-- [33,000-64,000] -->
<!--19%-->
detrás de Venezuela, Brasil y Colombia [@who2020world], 
causados en un 80% por _P. vivax_ [@rosas2016peru; @dgemalaria2020].

Reportes 
recientes
de malaria severa y fatal causados por _P. vivax_ han desafiado 
su tradicional condición de enfermedad benigna [@baird2009; @quispe2014; @TovarAcero2021].
<!--Se ha propuesto que, en parte,-->
Una hipótesis plantea que
el incremento de
estos casos pueden ser consecuencia de 
un retraso en la adquisición de inmunidad en la población 
<!-- por  -->
debido a
una reducción en la intensidad de transmisión [@reyburn2015];
<!--o exposición a la infección, ES externalidad negativa-->
<!-- En contraste con dicha hipótesis,  -->
<!-- Con respecto a dicha hipótesis, -->
sin embargo, se dispone de limitados estudios
que describan la respuesta inmune en esta población.
<!-- Por ejemplo,  -->
Un estudio en
<!-- pacientes de  -->
la Amazonía Peruana
comparó la respuesta humoral contra PvMSP1\textsubscript{19},
un marcador serológico tradicional,
en pacientes con malaria por _P. vivax_ severa 
y no complicada mostrando un grado similar de exposición previa
en ambos grupos [@baldevi2013].
No obstante, 
<!-- recientes  -->
estudios serológicos con múltiples antígenos
han desafiado la validez 
de MSP1\textsubscript{19} como indicador de 
inmunidad o exposición [@crompton2010; @Helb2015exposure].
Adicionalmente, la malaria severa por _P. vivax_ está definida clínicamente en la 
actual norma técnica [@norma2015]
<!-- [@norma2007]  -->
pero está ausente
<!-- Por ello, su  -->
<!-- ausencia -->
<!-- actual  -->
<!-- subestimación  -->
en el plan de control y eliminación [@cero2017].
El desconocimiento de la respuesta inmune de la 
malaria severa por _P. vivax_,
<!-- la falta de  -->
marcadores 
serológicos para su vigilancia
<!-- o moleculares de  -->
<!-- protección contra la -->
<!-- malaria severa  -->
<!-- en las actuales estrategias de eliminación  -->
[@accelerate2016] 
y su ausencia en planes de control
podrían 
<!-- implicar el -->
<!-- serio problema de Salud Pública  -->
afectar la morbilidad y mortalidad por malaria 
en zonas endémicas
al largo plazo.

Por esta razón,
el presente estudio tiene como objetivo comparar la respuesta de anticuerpos
ante la infección con _P. vivax_ 
contra más de 500 antígenos del parásito
<!-- expresados en el estadio eritrocítico de la infección. -->
en pacientes severos y no-severos
<!--provenientes-->de <!--un estudio prospectivo conducido en--> la 
ciudad de Iquitos<!--entre marzo del 2012 y junio del 2013-->,
empleando microarreglos de proteínas. 
Nuestra hipótesis es que los no-severos poseen mayor reactividad serológica 
contra antígenos de exposición, invasión o adhesión celular, 
con respecto a los severos.
Primero, se identificarán los antígenos con reactividad<!-- serológica--> diferenciada 
entre ambos grupos.
Luego, 
<!-- se determinarán los antígenos de respuesta secundaria en pacientes -->
<!-- con episodios previos de esta infección. -->
se realizará la misma comparación entre pacientes 
con y sin episodios previos de esta infección
para identificar posibles antígenos de respuesta secundaria.
Finalmente, se describirán sus características proteicas junto 
a los antígenos con mayor reactividad en toda la muestra 
con el propósito de proponerlos como candidatos a vigilancia seroepidemiológica.

<!-- ## Preguntas de investigación -->

<!--
Enuncie la pregunta de investigación que se va tratar en este estudio en una oración corta 
y directa.
La pregunta de investigación propuesta debe estar completa. Asegúrese de mencionar los 
siguientes elementos en la pregunta de investigación:

Mencionar los componentes del acrónimo PECO:
Población: ¿En qué población se quiere hacer el estudio?
Exposición: ¿Cuál es la intervención/exposición de interés?
Comparador: ¿Cuál es el grupo de comparación?
Outcome: ¿Cuál es el resultado o desenlace de interés?
-->

<!-- 1. ¿Cuáles son los antígenos con reactividad serológica diferenciada -->
<!-- ante la infección por _P. vivax_ entre pacientes severos y no-severos -->
<!--, de la ciudad de Iquitos (Loreto-Perú)-->
<!-- ? -->

<!-- 2. ¿Cuáles son los antígenos con reactividad serológica diferenciada -->
<!-- ante la infección por _P. vivax_ entre pacientes con y sin episodios previos -->
<!--, de la ciudad de Iquitos (Loreto-Perú)-->
<!-- ? -->

<!-- 3. ¿Cuáles son las características proteicas de los antígenos -->
<!-- con reactividad serológica diferenciada o predominante -->
<!-- en pacientes con malaria por _P. vivax_? -->

# JUSTIFICACIÓN DE LA INVESTIGACIÓN

## Justificación {#justif}

<!-- e importancia -->

Ante la reemergencia repetitiva de la malaria en la Amazonía Peruana 
[@griffing2013history; @rosas2016peru], 
es prioritario
implementar vigilancias serológicas programáticas en nuestro país [@hotspots2015]. 
En zonas con baja transmisión, estos ensayos poseen una mayor sensibilidad y representan
un menor gasto económico en comparación a las estrategias convencionales 
de monitoreo y control [@elliott2014]. 
Por esta razón, el presente estudio se justifica en 
el descubrimiento de antígenos potencialmente
discriminantes de condiciones clínicas como severidad o exposición. 
Estos permitirán optimizar y acelerar la ejecución de vigilancias serológicas 
en los programas de salud pública contra
la malaria en el Perú, en miras a su control y posterior eliminación [@accelerate2016].

<!-- 
Históricamente, reemergencias repetitivas de esta infección en la amazonía peruana 
han demostrado que las estrategias convensionales de monitoreo y control 
ya no son sostenibles [@rosas2016peru].

Una solución a este problema está en el uso de anticuerpos 
especie-específicos como biomarcadores para el monitoreo activo
de la exposición e inmunidad. 

En zonas de baja transmisión, estos ensayos poseen una 
mayor sensibilidad y menor gasto económico en comparación a las
estimaciones entomológicas o prevalencias por parasitemia [@elliott2014].

Dada su transmisión baja y estacional, 

AGREGA IMPORTANCIA DE: 
- BAJA transmisión y consecuente importancia de
- vigilar EXPOSICIÓN y 
- CARACTERIZAR respuesta de pacientes SEVEROS!!  INMUNIDAD? 

En este sentido, el análisis de respuestas serológicas ha surgido como una alternativa con 
el potencial de 
identificar _hotspots_ de transmisión [@hotspots2015] y así
direccionalizar la ejecución de intervenciones, en combinación con estudios geoespaciales. 
Sin embargo, aún existen brechas técnicas y de conocimiento que limitan el desarrollo 
y optimización de ensayos serológicos para el desarrollo de vigilancias programáticas. 
[@rosas2016peru]

Una de ellas es la selección de antígenos de _P. vivax_ en base a su rendimiento, 
tomando en cuenta parámetros como inmunogenicidad, polimorfismo, 
magnitud de respuesta, duración, especificidad por especie, sensibilidad y razón de 
adquicisión [@elliott2014].

La identificación de biomarcadores de exposición y severidad, junto a 
la implementación de un flujo de trabajo reproducible para el análisis de 
la respuesta serológica a larga escala, permitirá acortar las brechas mencionadas y
acelerar el descubrimiento de nuevas proteínas que puedan ser usadas como 
futuros marcadores serológicos óptimos en la ejecución de vigilancia seroepidemiológica
y evaluación de políticas de salud contra la malaria vivax.

Por ello, el Proyecto para la Eliminación de Malaria en el Perú tiene como objetivo 
optimizar ensayos serológicos complementarios a test de diagnóstico rápido con el fin de 
identificar a portadores asintomáticos [@accelerate2016]

## Consecuencia de la investigación {#consec}

La principal consecuencia de esta investigación estará en la identificación
de nuevos candidatos a marcadores serológicos en programas de vigilancia contra la malaria 
vivax. 
En contraste a lo identificado en estudios seroepidemiológicos a nivel poblacional 
[@cutts2014meta], 
los estudios de perfiles de anticuerpos a larga escala han reportado antígenos 
asociados a exposición o inmunidad distintos a los tradicionales
(PvCSP, PvMSP-1~19~, PvMSP-1~NT~ y PvAMA-1)
y con mayor resolución en la discriminación de pacientes [@crompton2010].

Una segunda consecuencia estará en la metodología de análisis empleando el software libre 
de computación estadística R. 
Este procedimiento permitirá que el flujo de trabajo ejecutado sea reproducible y, 
potencialmente, una plantilla estándar para el análisis de próximos trabajos 
con nuevos diseños experimentales y a su vez flexible para la generación de nuevas 
hipótesis, 
en comparación a recientes servidores web [@gmine2016].


Una vacuna efectiva
permitirá disminuir los índices de infección y mortalidad en niños de
la región amazónica a causa de esta infección parasitaria. Una vez
conocida cómo es la respuesta inmune y qué anticuerpos son adquiridos
naturalmente, se podrá evaluar nuevos candidatos a diagnóstico, dianas
para drogas y vacunas, tanto en modelos animales como en estudios
epidemiológicos.
-->

## Limitaciones {#limit}

<!-- Resumimos tres limitaciones del estudio. -->

Primero, las muestras a evaluar provienen de pacientes con infecciones no sincronizadas. 
<!-- El muestreo se ejecutó  -->
Las muestras fueron colectadas
en la fase sintomática, momento en el que 
los pacientes
acuden al Hospital, con 
un desconocimiento del inicio de la infección por paciente; sin embargo, infecciones 
experimentales han estimado que dicha fase ocurre normalmente entre los 11 y 13 días de 
infección [@arevalo2014].

Segundo, el estudio posee debilidades propias del diseño experimental a emplear.
El diseño de tipo caso control 
<!-- es susceptible a errores sistemáticos de  -->
<!-- selección y clasificación,  -->
posee dificultad para establecer relaciones causa-efecto 
y a través de él no se pueden calcular prevalencias o incidencias.
Además, la ausencia de un seguimiento activo impide el registro de covariables 
relevantes para la caracterización de la enfermedad.
<!--
severidad, e.g.: la exposición de los pacientes a malaria durante el último año
--
Sin embargo, las historias clínicas registraron el número de eventos previos con malaria 
reportados por el paciente
-->

Tercero, los microarreglos de proteínas poseen limitantes propias a su fabricación
[@vigil2010]. Cada paso (amplificación, clonamiento, expresión de genes a larga escala 
e impresión del arreglo) posee una eficiencia límite que afectará la calidad final de 
las proteínas. 
Además, el plegamiento proteico<!-- y formación de complejos multiméricos--> no será 
posible de verificar a dicha escala. Por último, la identificación de antígenos con 
modificaciones postranscriptacionales, particularmente relevantes en _Plasmodium_ 
[@leroch2009postmod],<!-- y antígenos no-protéicos, como polisacáridos y glucolípidos,--> 
no serán posibles de reproducir en su integridad en el sistema de expresión procarionte. 
A pesar de ello, se evaluará la validez<!-- y reproducibilidad--> del ensayo con controles 
internos. 
<!-- a detallar en la [sección 4.3.2.2](#validez). -->

<!-- Es por ello que parámetros 
de eficiencia en la producción del arreglo serán reportados 
(ver la sección de [validez y confiabilidad del instrumento](#validez)).  
- amplificación,
- clonamiento,
- expresión,
-->

<!--
SUGERENCIA DE NUEVOS DISEÑOS!
Con respecto a la exposición, variables claves como el tiempo 
desde la última infección y, 
con respecto a la severidad, manifestaciones clínicas en los episodios previos
-->

## Viabilidad

El proyecto es viable por la factibilidad de su ejecución, 
el interés de sus resultados para el área, la novedad de su enfoque y 
la relevancia de su método para la investigación biomédica.

Primero, su ejecución es factible por la disponibilidad inmediata de muestras
provenientes de un estudio prospectivo ya realizado, su adecuado tamaño muestral (n=60),
la objetividad de la pregunta planteada y la experiencia técnica del investigador en la
herramienta de análisis R [@R2016] y Bioconductor [@bioconductor2004].
Segundo, el problema es de interés para la comunidad por la posibilidad de que
sus resultados generen un cambio en la práctica médica.
Tercero, la novedad de su enfoque a larga escala permitirá 
<!-- corroborar, refutar o  -->
extender el estado del conocimiento y brindar mayor evidencia con respecto a la 
controversia mencionada.<!-- actual / generada por el enfoque tradicional -->
Cuarto, la relevancia de su método de análisis<!-- de datos--> reproducible con software libre
permitirá transparentar la generación de resultados y promover su puesta en práctica 
en futuras investigaciones.
<!-- de esta u otras áreas. -->
<!--
Finalmente, un aspecto novedoso del proyecto será el uso del lenguaje de programación y 
software libre R [@R2016] para el análisis de datos. Esta plataforma facilitará la 
reproducibilidad del estudio, promoviendo su reutilización y contribución por otros grupos
[@CienciaReproducible2016].
Este enfoque de análisis permitirá acelerar el progreso en esta área de investigación, 
automatizar procesos de análisis y reducir las brechas técnicas que continúan limitando 
el descubrimiento de biomarcadores relevantes para la aplicación de vigilancias 
seroepidemiológicas en nuestro país. -->
<!--
facilitando el proceso de muestreo y 
limitándolo al planteamiento del problema y análisis de los datos. 
microarreglos de proteínas requieren de pequeño volumen de plasma por paciente (~2$\mu$L), 
se contará con un adecuado tamaño muestral (n=60) el cual permitirá poner a prueba las 
hipótesis formuladas con suficiente poder estadístico. -->
<!-- 
Esto debido a la elección de un diseño Caso-Control, ideal para el estudio de casos 
infrecuentes, la evaluación de multiples factores para el mismo resultado y eficiente en 
recursos por tiempo. 
Cabe agregar , condición que posibilita escalar su aplicación a tamaño muestral 
epidemiológico. 
 -->


# FORMULACIÓN DE OBJETIVOS

## Objetivos

<!-- 
Objetivos:
  ¿Cuáles son los objetivos que piensas explorar en este estudio? 
  Usualmente se definen después de tener clara la pregunta de investigación y de haber 
  elaborado el marco teórico. Deben ser pocos y enfocarse sólo en el tema de interés.

Hipótesis:
  Describa la respuesta esperada a la pregunta de investigación.
-->
  
### General
  
- Identificar un subconjunto de antígenos con reactividad serológica 
diferenciada
<!-- discriminante de condiciones clínicas relevantes  -->
ante la infección por _P. vivax_
según severidad y episodios previos.

<!-- 
- Identificar antígenos de _P. vivax_ con reactividad serológica 
diferenciada ante la infección con malaria vivax entre pacientes 
según su severidad y episodios previos.

un subconjunto de
contra la malaria vivax severa 
-->
  
### Específicos
<!--  
- Identificar antígenos con reactividad diferenciada según la
severidad y episodios previos de los pacientes con malaria vivax.
- Identificar los antígenos con reactividad predominante en los pacientes 
con malaria vivax.
sintomática
-->
- Identificar antígenos de _P. vivax_ con reactividad serológica 
diferenciada ante la infección entre pacientes 
severos y no-severos.

- Identificar antígenos de _P. vivax_ con reactividad serológica 
diferenciada ante la infección entre pacientes 
con y sin episodios previos.

- Describir las características proteicas de los antígenos con reactividad 
diferenciada o predominante<!-- según parámetros disponibles en la base de datos 
PlasmoDB-->.


<!-- ### Exploratorio -->

<!-- - Comparar la amplitud e intensidad de respuesta de anticuerpos según  -->
<!-- la -->
<!-- severidad, episodios previos y--> 
<!-- edad de los pacientes con malaria por _P. vivax_. -->

<!--
- Comparar la amplitud e intensidad de respuesta de anticuerpos
según la edad de los pacientes con malaria vivax.
- Comparar la amplitud e intensidad de respuesta de anticuerpos según 
la severidad y episodios previos por rangos de edad de los pacientes con malaria vivax.
-->

# HIPÓTESIS Y VARIABLES DE INVESTIGACIÓN

## Hipótesis

<!-- 
Hipótesis:
Describa la respuesta esperada a la pregunta de investigación.
-->

<!--
### General

- Los pacientes con malaria vivax reconocen 
un subconjunto de antígenos con reactividad serológica 
y relevancia clínica en _P. vivax_.

Los pacientes con malaria vivax no-severa reconocen 
un subconjunto de antígenos con reactividad serológica 
y relevancia clínica contra la malaria vivax severa.
-->

<!-- ### De diferencia entre grupos -->
### Alternativa

1. Los pacientes con malaria no-severa poseen 
mayor reactividad serológica contra<!-- un subconjunto de--> antígenos de _P. vivax_
asociados a exposición, invasión o adhesión celular
<!--ante la infección -->con respecto a los pacientes<!-- con malatia vivax--> severos.

2. Los pacientes con episodios previos de malaria poseen
mayor reactividad serológica contra<!-- un subconjunto de--> antígenos de _P. vivax_
asociados a exposición
<!--ante la infección -->con respecto a los pacientes sin episodios previos.

<!-- ### Descriptiva -->

3. Los antígenos con reactividad diferenciada o predominante
<!--en pacientes con malaria vivax-->
se caracterizan por poseer una localización extracelular 
y estar bajo presión selectiva por el sistema inmune.<!--, entre ellas la presencia de 
dominios transmembrana, péptidos señal y 
razón de mutaciones no-sinónimas sobre sinónimas mayor a 1.-->

### Nula

1. Los pacientes con malaria no-severa 
no 
poseen 
mayor reactividad serológica contra<!-- un subconjunto de--> antígenos de _P. vivax_
asociados a exposición, invasión o adhesión celular
<!--ante la infección -->con respecto a los pacientes<!-- con malatia vivax--> severos.

2. Los pacientes con episodios previos de malaria 
no 
poseen
mayor reactividad serológica contra<!-- un subconjunto de--> antígenos de _P. vivax_
asociados a exposición
<!--ante la infección -->con respecto a los pacientes sin episodios previos.

<!-- ### Descriptiva -->

3. Los antígenos con reactividad diferenciada o predominante
<!--en pacientes con malaria vivax-->
no
se caracterizan por poseer una localización extracelular 
y estar bajo presión selectiva por el sistema inmune.<!--, entre ellas la presencia de 
dominios transmembrana, péptidos señal y 
razón de mutaciones no-sinónimas sobre sinónimas mayor a 1.-->


<!-- 4 HIPOTESIS
1. Los pacientes con malaria vivax no-severa poseen una 
amplitud e intensidad de respuesta de anticuerpos mayor a 
los pacientes con malatia vivax severa.

2. Los pacientes con malaria vivax no-severa reconocen 
un subconjunto de antígenos con mayor reactividad que 
los pacientes con malatia vivax severa.

3. Los pacientes con eventos previos reportados de malaria poseen una 
amplitud e intensidad de respuesta de anticuerpos mayor a 
los pacientes sin eventos previos reportados.

4. Los pacientes con eventos previos reportados de malaria reconocen 
un subconjunto de antígenos con mayor reactividad que
los pacientes sin eventos previos reportados.
-->

<!-- 6 HIPOTESIS ESTADÍSTICAS
1. Los pacientes con malaria vivax no-severa poseen una 
amplitud de respuesta de anticuerpos mayor a 
los pacientes con malatia vivax severa.

1. Los pacientes con malaria vivax no-severa poseen una 
intensidad de respuesta de anticuerpos mayor a 
los pacientes con malatia vivax severa.

2. Los pacientes con malaria vivax no-severa reconocen 
un subconjunto de antígenos con mayor reactividad serológica que 
los pacientes con malatia vivax severa.


3. Los pacientes con eventos previos reportados de malaria poseen una 
amplitud de respuesta de anticuerpos mayor a 
los pacientes sin eventos previos reportados.

3. Los pacientes con eventos previos reportados de malaria poseen una 
intensidad de respuesta de anticuerpos mayor a 
los pacientes sin eventos previos reportados.

4. Los pacientes con eventos previos reportados de malaria reconocen 
un subconjunto de antígenos con mayor reactividad serológica que
los pacientes sin eventos previos reportados.
-->

## Variables

<!--
### Identificación y clasificación
  **Variables**                      **Clasificación**
  --------------------------------- -------------------
  Reactividad serológica            Cuantitativa
  Episodio previo                   Cualitativa
  Severidad                         Cualitativa
-->

El presente estudio busca contrastar la respuesta de anticuerpos ante la infección con 
_P. vivax_ entre pacientes clasificados según su severidad y episodios previos. 
Para ello se medirá la variable de reactividad serológica de antígenos de _P. vivax_ luego 
de sondear plasma de pacientes en un microarreglo de proteínas. Además se clasificará a los 
pacientes como severos o no-severos de acuerdo a diagnóstico clínico y bioquímico, y como 
con-episodios o sin-episodios previos de acuerdo a lo reportado en la encuesta con el 
médico tratante.

Primero, la reactividad serológica cuantificará la especificidad de los anticuerpos de 
respuesta contra los antígenos del patógeno _P. vivax_. Esta variable medirá indirectamente 
la interacción entre antígenos y anticuerpos, a través de la lectura de la intensidad 
fluorescente producto de esta reacción.

Segundo, los pacientes severos se definirán por la presencia de 
manifestaciones clínicas severas o complicaciones sistémicas. Esta variable será asignada 
por la presencia de uno o más criterios clínicos y de laboratorio para la malaria severa,
según la clasificación recomendada por la Organización Mundial de la Salud (OMS) [@WHO2014severe].

Tercero, la clasificación por episodios previos reportados representará la presencia de 
infecciones previas con malaria y, consecuentemente, la presencia de una respuesta inmune 
humoral secundaria. Esta variable será reportada por el paciente como el número de eventos 
previos a la actual infección. Dada la incertidumbre de dicho dato, la variable será 
dicotomizada con el fin de facilitar su inclusión.

### Operacionalización de variables

Ver \autoref{tab:opera}.

<!-- 
VARIABLE INTERVINIENTE con OBJETIVO EXPLORATORIO: EDAD (hipotesis sin poder estadístico)

**Variable**, **Dimensión**, **Indicador**, **Clasificación**, **Escala de medición**
-->

\begin{table}[ht]
        \captionof{table}{Operacionalización de variables}
        \label{tab:opera}
        \vspace{-1mm}
\begin{center}
\hspace*{-1cm}
\begin{tabular}{>{\centering}m{2.4cm} m{2.2cm}m{2.2cm}m{1.8cm}m{2cm}m{1.7cm}m{1.5cm}m{1.6cm} @{}m{0pt}@{} }
  
  \hline
  \multirow{2}{*}{Variable}
  & 
  \multicolumn{2}{c}{Definición} 
  %&
  %\begin{minipage}{2.2cm}
  %Definición\\conceptual
  %\end{minipage}
  %&
  %\begin{minipage}{2.2cm}
  %Definición\\operacional
  %\end{minipage}
  & 
  \multirow{2}{*}{
  \begin{minipage}{1.8cm}
  Instrumento\\de medición
  \end{minipage}
  }
  &
  \multirow{2}{*}{
  \begin{minipage}{2cm}
  Criterios\\de medición
  \end{minipage}
  }
  &
  \multirow{2}{*}{
  \begin{minipage}{1.7cm}
  Tipo de\\variable
  \end{minipage}
  }
  &
  \multirow{2}{*}{
  \begin{minipage}{1.5cm}
  Escala de \\medición
  \end{minipage}
  }
  &
  \multirow{2}{*}{
  Fuente
  } &\\[0ex]
  %\hline
  \cline{2-3}
  
  &
  Conceptual
  &
  Operacional
  & 
  &
  &
  & &\\[1ex]
  \hline
  
  \textbf{Dependiente} Reactividad serológica
  & 
  % esCONCEPTUAL: 
  \begin{minipage}{2.2cm} 
  Especificidad \\de anticuerpos \\de respuesta contra un antígeno
  \end{minipage} 
  &
  % aOPERACIONAL: 
  \begin{minipage}{2.2cm} 
  Medida \\indirecta de \\la reacción antígeno-anticuerpo
  \end{minipage} 
  % aDETALLES: medida indirecta de la reacción antígeno-anticuerpo 
  % mediante la lectura de la reacción fluorescente entre 
  % anticuerpo secundario y fluoroforo por spot
  & 
  \begin{minipage}{2.2cm} 
  %Lector\\ óptico de\\
  Microarreglo\\de proteínas
  \end{minipage}
  & 
  \begin{minipage}{2cm} 
  \textbf{0-6000} MFI o intensidad\\
  fluorescente \\promedio.
  \end{minipage} 
  &
  Numérica contínua
  & 
  Razón
  &
  Plasma sanguíneo &\\[13ex]
  \hline

  \textbf{Independiente} Severidad
  & 
  % aCONCEPTUAL: 
  Presencia de manifestaciones clínicas severas o complicaciones sistémicas
  &
  % aOPERACIONAL:
  Número de criterios de la clasificación OMS para malaria severa
  & 
  \begin{minipage}{2.2cm} 
  Diagnóstico \\clínico \\y pruebas \\bioquímicas 
  \end{minipage}
  & 
  \begin{minipage}{2cm} 
  \textbf{No-severa:} 0 criterios\\
  \textbf{Severa:} 1 o más criterios.
  \end{minipage}
  &
  Categórica dicotómica
  & 
  Nominal
  &
  Historia clínica y muestra de sangre &\\[15ex]
  \hline
  
  \textbf{Independiente} Episodios previos
  & 
  % aCONCEPTUAL: 
  Exposición a la infección de malaria en el pasado
  &
  % aOPERACIONAL:
  Número de episodios previos reportados 
  & 
  Encuesta
  & 
  \begin{minipage}{2.1cm} 
  \textbf{Sin:} 0 episodios\\
  \textbf{Con:} 1 o más episodios
  \end{minipage}
  &
  Categórica dicotómica
  & 
  Nominal
  &
  Historia clínica &\\[10ex]
  \hline

%  \textbf{Interviniente} Edad %Confusora
%  & 
%  % aCONCEPTUAL: 
%  Edad del paciente
%  &
%  % aOPERACIONAL:
%  Años de vida reportados
%  & 
%  Encuesta
%  & 
%  \begin{minipage}{2.2cm} 
%  \textbf{0-90} años.
%  \end{minipage}
%  &
%  Numérica discreta
%  & 
%  Razón
%  &
%  Historia clínica &\\[10ex]
%  \hline


  % etc. ...
\end{tabular}
\hspace*{-1cm}
\end{center}
\end{table}

# ANTECEDENTES

<!--# MARCO TEÓRICO-->

<!-- 
teoría/conceptos
¿hallazgos/resultados?
-->

## Antecedentes de la investigación

<!-- a. En África -->
    
En África
    un hito en la aplicación de microarreglos de proteínas para el estudio 
    de la respuesta humoral a escala epidemiológica fue la publicación de 
    @crompton2010. 
    Este estudio comparó la respuesta de anticuerpos contra el 23% del proteoma 
    de _P. falciparum_ antes y después de la temporada de malaria en 220 
    individuos de Mali, en dos grupos poblacionales: 2-10 y 18-25 años.
    Dentro del grupo de niños entre 8 y 10 años se identificaron 49 proteínas con 
    mayor reactividad serológica en el grupo de infectados asintomáticos, en 
    comparación a los sintomáticos. Cinco de los principales candidatos a vacuna
    (CSP, LSA-3, MSP1, MSP2, AMA1) no lograron discriminar ambos grupos. Sin embargo,
    cuatro candidatos secundarios (STARP, LSA-1, RESA, antígeno 332), con expresión en 
    diferentes estadios del ciclo biológico, sí lograron tal discriminación.
    <!--Por último, una alta respuesta contra protéinas intracelulares fue sugerida como un
    marcador de mayor muerte parasitaria.-->
    
Un segundo hito de interés 
    en África 
    <!-- representa  -->
    se describe en
    el trabajo de 
    @Helb2015exposure 
    <!-- Ellos  -->
    donde
    reportaron una estrategia para 
    identificar combinaciones de respuestas de anticuerpos
    contra más de un antígeno que maximicen la información de la exposición reciente
    al nivel de individuos. 
    Para ello emplearon modelos basados en el aprendizaje automático o 
    _machine learning_ para el análisis de las respuestas contra
    865 antígenos de _P. falciparum_ en 186 niños (3-6 años) de Uganda
    en base al registro activo y pasivo de sus historias clínicas a los largo de un año.
    En contraste a los marcadores tradicionalmente empleados para evaluar exposición
    a nivel poblacional (CSP, MSP1, MSP2, AMA1) [@elliott2014], se identificaron
    marcadores más informativos como hyp2, GEXP18, EMP1, ETRAMP4, HSP40-II y PF70.
    La validación de este método permitirá optimizar la selección tradicional de marcadores.
    <!--seleccionar mejores marcadores para la 
    vigilancia seroepidemiológica, relevante para el guiado y evaluación de los 
    programas de control y eliminación de malaria a nivel mundial.-->

<!-- b. En Perú -->
    
En Perú
    el 
    primer estudio en publicarse fue de
    @Torres2014asymptomatic 
    donde reportaron marcadores de inmunidad clínica -no esterilizante- de adquisición 
    natural a la malaria por _P. falciparum_.
    Además de ello, el resultado con mayor implicancias fue que
    estos marcadores presentaron un enriquecimiento de polimorfismos no-sinónimos, 
    indicativo de una presión de selección positiva por parte del sistema inmune, a pesar
    de provenir de una zona con baja transmisión.
    Las 51 proteínas con mayor reactividad en 14 pacientes infectados y asintomáticos
    se obtuvieron al comparar
    las respuestas con 24 paciente sintomáticos provenientes de la Amazonía Peruana 
    (Departamento de Loreto, Provincia de Maynas y Requena)
    contra 824 fragmentos (699 proteínas) de _P. falciparum_.
    
El segundo y último estudio 
    sobre malaria y microarreglo de proteínas 
    en Perú
    en ser publicado ha sido el de
    @chuquiyauri2015vivax. 
    Aquí compararon la respuesta contra _P. vivax_ de pacientes
    con relapsos y reinfección, sin encontrar diferencia alguna entre ambos grupos. Sin 
    embargo, al igual que el anterior estudio, identificaron un enriquecimiento de 
    proteínas con polimorfismos no-sinónimos en el grupo de antígenos reactivos. Además,
    dentro del grupo con mayor reactividad en toda la muestra, resaltaron a 
    PvMSP-10 como potencial candidato a vacuna al presentar la expresión más consistente 
    y validar lo reportado por dos estudios previos con enfoque tradicional 
    donde emplean ambos sistemas de expresión: eucariota y procariota. 
    <!-- El estudio empleó -->
    Aquí emplearon
    un arreglo con 2233 fragmentos (1936 proteínas) en 106 individuos de la ciudad de 
    Maynas, Loreto - Perú.

<!--
a. mundial

    Trieu et al. 2011 [@Trieu2011], inmunización experimental contra _P. falciparum_
    
    Finney et al. 2014 [@Finney2014], mono- y co-infecciones 
    de _P. falciparum_ y _P. vivax_ 
    con anticuerpos de adquisición natural en Papua Nueva Guinea

c. En Latinoamérica

    Molina et al. 2012 [@molina2012] en _P. vivax_
    con anticuerpos de adquisición natural en Colombia
    
    Arévalo-Herrera et al. 2016 [@arevalo2016] inmunización experimental con _P. vivax_ 
-->

## Bases teóricas

<!-- 
SEVERIDAD
Por un lado, modelos experimentales de malaria falciparum severa 
han sugerido un rol de la inmunidad en la reducción de la severidad [@Moreno2013],
la cual se adquiere por una constante exposición a la infección [@Stanisic2015].

.-.-..-.-.-..-.-.-.-.-.-..

INMUNOGENICIDAD
Ante una infección parasitaria, el sistema inmune responde a la infección
mediante el reconocimiento de inmunógenos del patógeno por parte de los linfocitos B.
Esta acción gatilla su diferenciación y la consecuente producción de 
células plasmáticas encargadas de la secreción de anticuerpos 
con afinidad o reactividad específica
contra los antígenos desencadenantes [@abbas2012].

INMUNÓGENOS
Generalmente, se reconoce que los antígenos protéicos poseen 
características generales que permite seleccionarlos de acuerdo a 
sus propiedades físicas, función y localización celular.
Este subconjunto suele estar estar directamente relacionado 
a los niveles de expresión protéica [@vigil2010],
más aún en patógenos con ciclos de vida complejos, como _P. vivax_,
quien presentan perfiles de expresión únicos por cada estadio de 
infección [@bozdech2008; @westenberger2010; @bozdech2016].

INMUNIDAD
titulo/nivel de anticuerpos vs edad/exposición -> INMUNIDAD PROTECTORA [@Stanisic2015]

.-.-..-.-.-..-.-.-.-.-.-..

CROSS-REACTIVITY
Existe la probabilidad que los pacientes generen anticuerpos contra
proteínas con alta homología entre *P. vivax* y *P. falciparum*
o especie-específicas (sin ortología con otras especies).
-->

    

### Malaria por _Plasmodium vivax_

a. Epidemiología
    
    <!-- i.  **A nivel mundial** -->
    
    A nivel mundial
        *P. vivax* y *P. falciparum* son los principales responsables de los 
        casos de malaria en humanos. Ambas especies exponen aproximadamente a
        2.5 mil millones de personas en riesgo de infección [@howes2016global];
        sin embargo, *P. vivax* es el parásito dominante en las regiones fuera 
        del África subsahariana, en su mayoría densamente pobladas y empobrecidas.
        Entre ellas, Etiopía, India, Indonesia y Pakistán acumularon el 78% de casos de 
        _P. vivax_ a nivel mundial. A su vez, la región de las Américas tuvo la 
        mayor proporción de estos, con un 69% [@WHO2016world]. A pesar de ello,
        hasta el momento la mayoría de la investigación y financiamiento está 
        destinado a la prevención, tratamiento y control de _P. falciparum_ [@path2011].
    
    <!-- ii. __En el Perú__ -->
    
    Con respecto a Perú,
        en 
        el 2015 
        este 
        fue el tercer país con más casos reportados en 
        Latinoamérica (19%), detrás de Brasil (24%) y Venezuela (30%) [@WHO2016world].
        El 80% fueron causados por _P. vivax_ (63,153 en total), en regiones con 
        endemismo y transmisión heterogénea [@rosas2016peru]. El 95% 
        pertenecieron al noreste amazónico (con una razón Pv/Pf de 4/1) 
        y el resto al norte costero 
        y la región minera del Suroeste. Notablemente, comunidades en 
        Madre de Dios no han registrado caso por _P. falciparum_ en 
        una década (2001-2012) [@rosas2016peru]. Con respecto a factores 
        ambientales, en el año 1998 
        durante el fenómeno El Niño-Oscilación Sur (ENSO) se produjo 
        el mayor pico de casos anuales en la historia (200,000 casos), donde la región
        del norte costero representó el 48% de los casos a nivel nacional [@gagnon2002enso]. 
        <!--Los casos anuales 
        provenientes de Tumbes, Piura y Lambayeque a nivel nacional pasaron de
        representar menos del 7% en 1996, a un 48% entre 1998 y 1999.-->

b. Biología
    
    i. __Ciclo de vida__
    
        La ecología natural del parásito de la malaria involucra a dos hospederos, 
        el humano y el mosquito, y tres ciclos de vida:
        1. El _ciclo hepático o exo-eritrocítico_
        inicia con la picadura del mosquito infectado, la liberación de los esporozoitos 
        al fluido sanguíneo del humano y su ingreso a las células hepáticas. 
        Aquí se da su desarrollo asexual
        donde se multiplican hasta formar esquizontes, los cuales egresan al torrente 
        sanguíneo en la forma de merozoitos.
        2. El _ciclo eritrocítico_ se inicia con la invasión de
        glóbulos rojos (RBC) y el desarrollo consecutivo de trofozoitos 
        inmaduros <!--(estadio "anillo")-->, maduros y esquizontes, los cuales a su
        ruptura liberan nuevos merozoitos que reinfectan a más RBC. 
        <!--La rapidez de este ciclo determina la razón de multiplicación parasitaria.-->
        Por motivos aún no determinados, 
        a partir de los trofozoitos inmaduros se inicia el desarrollo de gametocitos
        diferenciándose sexualmente dentro del torrente sanguíneo del humano. 
        3. El _ciclo esporogónico_ o fase sexual se inicia en el mosquito hembra mediante la
        ingestión accidental de gametocitos en la alimentación sanguínea, 
        con la intención inicial de proveer de nutrientes a sus huevos.
        En su tracto digestivo se generan cigotos mótiles
        que invaden las paredes intestinales para formar un oocisto
        donde los esporozoitos se desarrollan y replican hasta su ruptura y liberación.
        Finalmente migran a las glándulas salivares para continuar
        la transmisión en la siguiente 
        ingesta de sangre del mosquito hospedero.
    
    ii. __Particularidades__
        
        _P. vivax_ posee importantes variantes biológicas que caracterizan su 
        epidemiología y dinámica de infección [@howes2016global]. Tres de las 
        más importantes son:
        1. Presencia de relapsos por la activación de 
        hipnozoitos, el cual es un estado de latencia o dormacia en el _ciclo hepático_, 
        condición que puede generar tanto una liberación prolongada y lenta de merozoitos 
        como reservorios de infección que prolongan 
        su transmisión; 
        2. Tropismo hacia reticulocitos o RBC inmaduros
        (1-2% de RBC circulantes) en el _ciclo eritrocítico_, 
        condición que genera bajas parasitemias en comparación a _P. falciparum_; 
        y 
        3. Dependencia del antígeno _Duffy_ para la invasión de RBC, cuya
        ausencia en individuos con ancestría africana 
        ha sido considerada como un factor de resistencia a _P. vivax_.
        <!--tiene una alta frecuencia en las poblaciones-->

### Malaria Severa

a. Definición
    
    Ante la ausencia de una descripción especie-específica, la malaria
    severa por _P. vivax_ está definida por los criterios para _P. falciparum_ otorgados 
    por la OMS en el 2014 [@WHO2014severe], el cual incluye una o más de
    las siguientes (todas registradas en mono-infecciones 
    por _P. vivax_):
    1. Condición neurológica: coma, mareo, pérdida de conciencia;
    2. Condición hematológica: anemia/trombocitopenia severa;
    3. Síntomas sistémicos: shock circulatorio; y
    4. Fallo de órganos vitales:
    disfución respiratoria, estrés respiratorio agudo, 
    daño renal agudo, ruptura esplénica, 
    disfunción hepática e ictericia (hiperbilirrubina).

b. Epidemiología
    
    <!-- i. __A nivel mundial__ -->
        
    A nivel mundial,
        entre 
        el 2005 y 2015, 1-3% de casos no-complicados fueron asumidos 
        como malaria severa. Esta causó la muerte de 429,000 personas [@WHO2016world],
        en su mayoría (90%) niños menores de 5 años en África [@wassmer2015].
        La vulnerabilidad a esta manifestación ha sido asociada con la intensidad de 
        transmisión y el desarrollo de la inmunidad dependiente a la edad [@reyburn2015].
        1. En zonas de _alta transmisión_ como en el África subsahariana, 
        las poblaciones más vulnerables son: niños menores de 5 años con un desarrollo 
        incompleto de inmunidad parcial contra la malaria [@Stanisic2015], 
        mujeres embarazadas en parte a la adhesión placentaria de 
        glóbulos rojos infectados (iRBC) [@rogerson2007preg], y viajeros o migrantes sin 
        inmunidad provenientes de áreas con baja o ninguna transmisión de malaria. 
        2. En zonas de _baja transmisión_ como en Asia y América Latina, 
        al haber una menor exposición a la infección, la mayoría de la población 
        llega a la adultez sin haber desarrollado una inmunidad protectora. 
        Como consecuencia, <!--tanto en la Amazonía Peruana como en el norte costero -->la 
        población adolescente y adulta joven es la más susceptible a desarrollar esta 
        patología [@llanoschea2015], comúnmente al iniciar trabajos a campo 
        abierto, e.g. actividades madereras o mineras, en zonas de alto riesgo de 
        contacto con mosquitos infectados [@factores2001].
        
    <!--Cabe agregar que la presencia de comorbilidades, coinfecciones y malnutrición 
        son reconocidos como factores contribuyentes a la Malaria Severa, los cuales deben
        ser adecuadamente registrados. Sin embargo, se ha sugerido que la coinfección de 
        _Plasmodium_ con Dengue no está asociada con una peor enfermedad,
        asemejándose a una mono-infección con Dengue tanto en frecuencia de síntomas como 
        en el nivel de marcadores inmunológicos [@baldevi2016].-->
        
    <!-- i. __En el Perú__ -->
        
    Con respecto a Perú,
        en 
        el norte costero @quispe2014 mediante un estudio 
        retrospectivo (2008-2009) identificaron 81/6502 casos de malaria severa por _P. vivax_
        con anemia severa (57%), 
        shock circulatorio (25%), hiperbilirrubina (25%), daño pulmonar (21%), 
        daño renal agudo (14%) y Glasgow $\le$ 9/14 (11%).
        Comparados con los pacientes no complicados, los severos fueron mayores 
        (38 vs 26 años, P<0.001).
        
    Adicionalmente, 
    en el noreste amazónico un reciente estudio prospectivo dirigido por 
        El Centro de Enfermedades Tropicales de la Marina de los Estados Unidos 
        NAMRU-6 (2012-2013) [@smith2013] identificó 55/164 casos de 
        malaria severa por _P. vivax_ con
        shock circulatorio (27.6%), daño pulmonar (24.5%)
        hiperbilirubina (18.1%), daño renal agudo (2%), Glasgow $\le$ 9/14 (2%) 
        y anemia severa (1.7%). 
        No se encontraron diferencias
        con respecto a la edad entre no-severos y severos (33 vs 29 años, P=0.497).
        Sin embargo, sí se identificó una mayor proporción de malaria 
        severa en mujeres (52 vs 35%, P=0.039).

<!--
b. Epidemiología
    
    i. __A nivel mundial__
        
        Entre el 2005 y 2015, 1-3% de casos no-complicados fueron asumidos 
        como malaria severa. Esta causó la muerte de 429,000 personas [@WHO2016world],
        en su mayoría (90%) niños menores de 5 años en África [@wassmer2015].
        <!--De las estrategias preventivas desarrolladas hasta el momento
        (control vectorial, quimioprevención, vacunas e indicadores)
        la única vacuna en completar un ensayo de fase 3, RTS,S/AS01, solo redujo 
        en un 39% la incidencia de enfermedad y en 31.5% la de malaria severa en niños
        entre 5-17 meses [@rts2015].
        
        Con respecto a la malaria severa por _P. vivax_, un reciente meta-análisis con 
        46,411 casos registrados desde el año 1900 identificó patrones geográficos 
        en la prevalencia de manifestaciones clínicas [@rahimi2014meta].
        Concluyeron que en las zonas de alta transmisión (e.g., Papúa Nueva Guinea)
        la anemia severa fue el síntoma con mayor prevalencia en niños jóvenes. 
        <!--Por ejemplo, en India Kochar et al. 2009 [@kochar2009] en un estudio 
        prospectivo reportó 
        40/456 casos de malaria vivax severa, entre 18-60 años de edad, con 
        hiperbilirrubina (58%), daño renal agudo (45%),
        anemia severa (33%), malaria cerebral (13%), daño pulmonar (10%), shock 
        circulatorio (8%) e hipoglicemia (3%).
        Por otro lado, en zonas con baja transmisión (e.g., la región de las 
        Américas) los adultos eran los más susceptibles a la severidad con una 
        mayor variedad de disfunciones orgánicas.
        Por ejemplo, en Brasil un estudio retrospectivo
        identifico 17/11,251 casos de malaria vivax severa, entre 28-80 años de edad, 
        con hiperbilirrubina (59%), 
        anemia severa (29%), daño renal agudo (12%), daño pulmonar (12%) y shock 
        circulatorio (6%) [@alexandre2010].
-->    


c. Biología
    
    i. __Patogénesis__
        
        Se han propuesto mecanismos
        en base a lo observado en los casos<!-- de malaria cerebral --> por _P. falciparum_,
        donde los procesos de invasión y adhesión parasitaria son relevantes.
        La respuesta celular contra la parasitemia, <!--dependiente de su tasa de multiplicación,-->
        sumada a la citoadherencia 
        de iRBC en estadio esquizonte a endotelio o RBC no infectados (proceso llamado "rosetamiento"), 
        desencadena una obstrucción microvascular que activa a las células endoteliales. 
        En respuesta, estas liberan una mayor cantidad de citoquinas proinflamatorias, 
        provocando 
        la pérdida de perfusión y una consecuente disfunción microvascular que progresa 
        por retroalimentación positiva.
    
    ii. __Comparación__
    
        Históricamente, la malaria por _P. vivax_
        ha sido considerada como "benigna" en comparación a _P. falciparum_ 
        debido a su: (1) baja invasión parasitaria, sesgada 
        a reticulocitos y rutas alternas de menor efectividad; y (2) pobre 
        citoadhesión de sus iRBC, dada por la ausencia de 
        protrusiones abastonadas o _knob protrusions_,
        codificadas por genes _var_. 
        Sin embargo, la identificación de genes homólogos a _Pf var_ (_Pv vir_)
        <!--la baja proporción de esquizontes tardíos -->y 
        evindencias _post mortem_ de iRBC en capilares pulmonares
        han sugerido la posibilidad de secuestramiento tisular por _P. vivax_ [@wassmer2015].
        En línea con esta evidencia, 
        recientemente se ha demostrado que en los casos severos por _P. vivax_ 
        la parasitemia periférica subestima la biomasa parasitaria total,
        siendo la fracción oculta la mayor contribuyente de citoquinas proinflamatorias, 
        semejante a los casos por _P. falciparum_ [@barber2015]. 
        <!--En este estudio, la parasitemia oculta fue la 
        mayor contribuyente de citoquinas proinflamatorias, semejante a lo
        observado con _P. falciparum_, pero no de la activación endotelial, 
        sugiriendo una posible acumulación parasitaria en partes de órganos sin endotelio.-->

### Respuesta de anticuerpos

La concentración de anticuerpos contra antígenos de _Plasmodium_ es un
biomarcador sensible a la exposición de malaria<!-- a nivel individual--> [@elliott2014];
sin embargo, las evidencias a favor de dianas para _P. vivax_ continúan estando 
limitadas por la ausencia de cultivos _in vitro_ y modelos animales de infección.
Este problema se ve reflejado en su escaso número de candidatos a vacuna (PvDBP y PvCSP) 
<!--(PvDBP en ensayos clínicos fase 1 y PvCSP en estudios pre-clínicos) -->
en comparación a los 23 de _P. falciparum_ [@rainbow2016].
Por ello, la principal fuente de evidencia a favor de biomarcadores para _P. vivax_ está en la
identificación de antígenos con respuestas de anticuerpos asociadas a la adquisición 
natural de inmunidad en poblaciones endémicas.

<!--
a. Relevancia
    
    La concentración de anticuerpos contra antígenos de _Plasmodium_ es un
    marcador sensible a la exposición de malaria a nivel poblacional [@elliott2014].
    Con métodos analíticos aplicables a escala epidemiológica, es posible inferir su 
    desarrollo e identificar focos o _hotspots_ de transmisión [@hotspots2015].
    Además, su aplicación en salud pública tiene el potencial de determinar
    el impacto de intervenciones, confirmar los estados de control o eliminación, y 
    monitorear reemergencias [@sepulveda2015].
    
    Sin embargo, las evidencias a favor de marcadores en _P. vivax_ continúan estando 
    limitadas por la ausencia de cultivos _in vitro_ y modelos animales de infección.
    Esto se ve reflejado en sus dos candidatos a vacuna (PvDBP en ensayos 
    clínicos fase 1 y PvCSP en estudios pre-clínicos) contra los 23 de _P. falciparum_
    [@rainbow2016].
    Por ello, la principal fuente de evidencia en aquella especie está en la
    identificación de antígenos con respuestas asociadas a la adquisición 
    natural de inmunidad en poblaciones endémicas.
-->
    
a. Marcadores
    
    i. __de exposición__
    	
    	<!--Su presencia es un paso previo a la adquisición de inmunidad.-->
    	Estudios longitudinales han sugerido que, en un inicio, la intensidad de la 
    	respuesta de anticuerpos actúa como marcador de exposición la cual, luego de una 
    	constante exposición, incrementa hasta superar un umbral de protección
    	[@Stanisic2015].
	    En este sentido, estudios seroepidemiológicos en _P. vivax_ han logrado 
	    asociar marcadores
	    con exposición (PvCSP, PvMSP-1~19~, PvMSP-9~RIRII~ y PvAMA-1) 
	    e inmunidad (PvMSP-1~19~, PvMSP-1~NT~, PvMSP-3$\alpha$ y PvMSP-9~NT~) 
	    [@cutts2014meta].<!-- Sin embargo, su principal limitante está en la
	    carencia de evaluaciones en su cinética de adquisición, desarrollo y 
	    mantenimiento en cohortes longitudinales.-->
	    <!--
	    Actualmente, el cálculo de la razón de seroconversiónn (SCR) es el método más 
	    estandarizado para la evaluación de la exposición y caracterización
	    de la transmisión por adquisición de anticuerpos a nivel poblacional [@King2015FOC].
	    -->
    
	ii. __de inmunidad__
	    
	    En áreas de alta transmisión, los niños adquieren resistencia a la 
	    _manifestación severa_ de la malaria a la edad de cinco años, aproximadamente. 
	    Sin embargo, continúan siendo susceptibles a episodios no-complicados hasta la adolescencia,
	    donde adquieren un estado resistente a la _malaria sintomática_. A pesar de ello,
	    no se ha demostrado que esta adquisición natural de inmunidad por la exposición 
	    acumulada a malaria otorgue una _protección esterilizante_ o de resistencia a la 
	    infección [@crompton2014rev].
	    Los detalles de los procesos involucrados en cada una de estas tres fases no están
	    completamente entendidos, pero se tiene detalles de algunos componentes:
    	
	    1. __contra la severidad__
        	
	        <!--Se conoce que la respuesta contra esta manifestación está guiada por una
	        excesiva respuesta inflamatoria que incluye la producción de las citoquinas 
	        proinflamatorias TNF-$\alpha$, INF-$\alpha$, IL-1$\beta$ e IL-6 [@baird2013]
	        junto a una 
	        reducción en la producción de la citoquina antiinflamatoria IL-10, un regulador
	        central en la respuesta contra la malaria [@jagannathan2014].
	        Es por ello que esta primera fase implicaría la adquisición de
	        inmunidad a componentes parasitarios que desencadenan una respuesta inflamatoria 
	        innata. Una hipótesis está en anticuerpos que interfieran la unión de 
	        glucoproteínas proinflamatorias con receptores Toll 
	        [@schofield2006toll; @coban2005toll].-->
        	
	        La adquisición de anticuerpos que bloqueen la adhesión de iRBC 
	        o invasión parasitaria son candidatos a otorgar una protección a este
	        nivel [@wassmer2015].
	        Antígenos referenciales son (i) las proteínas VIR en _P. vivax_, homólogas a VAR 
	        en _P. falciparum_,
	        posiblemente secretadas en iRBC y causantes de adhesividad a endotelio o 
	        RBC no infectados, condición que desencadena el rosetamiento y 
	        posterior obstrucción microvascular [@portillo2001vir], y (ii)
	        las proteínas de la familia PvDBP y PvRBP
	        (_Duffy-_ y _Reticulocyte- binding proteins_), responsables
	        de la invasión a reticulocitos por la ruta tradicional y alternativa, respectivamente
	        [@galinski1992rbp].
    	
	    2. __contra la enfermedad__
	        
	        <!--Los antígenos del estadio eritrocítico (merozoitos) fueron los primeros 
	        marcadores en ser propuestos como importantes para la protección a este nivel,
	        tal como lo sugirió en 1961 la terapia efectiva por transferencia pasiva de 
	        anticuerpos donados por adultos con inmunidad adquirida de forma natural
	        [@cohen1961]. <!-- [@sabchareon1991] -->
	        
	        En el caso de _P. vivax_, su adquisición ocurre con una mayor rapidez, 
	        posiblemente facilitada por sus particularidades biológicas [@mueller2013].
	        Los antígenos propuestos hasta el momento para esta especie son proteínas del 
	        micronema de merozoitos como DBP y AMA1, y proteínas de superficie como 
	        MSP1, MSP1-P (región C-terminal), MSP3 (PvMSP-3$\alpha$, PvMSP-3$\beta$) 
	        y MSP9 [@lopez2017].
	        <!--Desarrollo de respuestas inmunoreguladoras que atenuen la respuesta 
	        a productos 
	        parasitarios como citoquinas antiinflamatorias (IL-10), expanción de células T 
	        regulatorias, y subtipos específicos de células B y T con alta eficiencia en la
	        generación de respuestas humorales.-->
        	
	    3. __contra la parasitemia__
        	
	        La adquisición de una inmunidad esterilizante solo ha sido comprobada 
	        mediante la inmunización con esporozoitos atenuados
	        por radiación. Tanto en _P. falciparum_ como en _P. vivax_ se ha confirmado
	        esta respuesta. Un reciente ensayo clínico de fase 2 en Colombia con _P. vivax_ 
	        encontró
	        que luego de siete inmunizaciones a lo largo de 56 semanas y una reexposición 
	        con esporozoitos 
	        infecciosos, 5/12 voluntarios obtuvieron dicha protección y 
	        la respuesta de anticuerpos 
	        IgG1 anti-PvCSP estuvo asociada a ella [@arevalo2016spz].
	        <!-- Estudios en modelos murinos han señalado a 
	        las respuestas de células T CD8^+^ como importantes mediadores de este tipo de 
	        inmunidad.-->
        	
b. Enfoques a larga escala
    
    El parásito de la malaria, al tener un ciclo de vida complejo con múltiples estadios
    de desarrollo en el humano, posee también un transcriptoma, proteoma y, por lo tanto, 
    un _inmunoma_ estadio-específico que deriva de un total de ~5300 proteínas putativas. 
    Por este motivo, tanto los esfuerzos de vacunas por subunidades 
    como por organismos completos atenuados han fracasado,
    en contraste a su eficacia en otras infecciones como 
    hepatitis B (HBsAg) y tuberculosis (BCG), respectivamente [@immunomics2016].
    <!--Incluso, ensayos 
    clínicos de fase 3 han demostrado que los actuales candidatos poseen una pobre 
    inmunogenicidad y protección contra la malaria (e.g., RTS,S) [@rts2015].-->
    
    <!--La era genómica permitió un cambio de perspectiva en la investigación inmunológica:
    se pasó de una enfocada en plataformas de vacunación a una 
    en antígenos de vacunación.
    Es por ello que, actualmente, el objetivo de mayor importancia está en la
    identificación de subgrupos de antígenos capaces de inducir respuestas inmunes
    protectoras [@Davies2015Large].-->
    
    i. __Inmunómica__
        
        
        En respuesta a las limitaciones de la _vacunología reversa_, basada en el
    uso inicial de la información genómica de patógenos, el enfoque _inmunómico_
    hace uso de la información de ambos agentes interactuantes: patógeno y hospedero.
    Su estrategia selecciona inmunógenos tanto por (1) _métodos empíricos_, usando data
    transcriptómica y proteómica de patógenos dinámicos más la información clínica 
    del hospedero, como por (2) _métodos teóricos_, empleando algoritmos computacionales predictivos 
    que tomen en cuenta la afinidad entre péptidos del patógeno y moléculas 
    del complejo principal de histocompatibilidad (MHC) del hospedero.
    
        Particularmente, el primer método
    ha permitido obtener datos empíricos sobre 
    diferencias en la amplitud, intensidad, cinética y longevidad de respuestas inmunes
    inducidas por patógenos. Además de mayor información sobre aspectos clave como el
    porcentaje del proteoma reconocido por el hospedero, la capacidad predictiva
    de anticuerpos sobre el estado de la enfermedad o de las
    secuencias de aminoácidos sobre su inmunogenicidad [@Davies2015Large].
    <!--Dos plataformas representativas son los microarreglos de proteínas
    y péptidos, este último con mayor resolución en el mapeo de epítopes 
    [@carmona2015peptide].-->
    
<!--
    i. __Vacunología reversa__
    
        El primer enfoque en respuesta a esta problemática
        hizo uso de la información genómica de los patógenos. 
        Su estrategia se basó en
        la predicción bioinformática de 
        antígenos candidatos a vacuna a partir de características protéticas asociadas a 
        inmunogenicidad y efectividad en vacunación. La principal limitación de este 
        enfoque 
        radica en el que el genoma es estático en el tiempo, no aplicable en organismos
        con ciclos de vida complejos [@immunomics2016].    
    
    ii. __Inmunómica__
        
        Un reciente enfoque
        hace uso de la información de ambos agentes interactuantes: patógeno y hospedero.
        Su estrategia se basa en
        la selección de inmunógenos por (i) _métodos empíricos_ usando data
        transcriptómica y proteómica de patógenos dinámicos más la información clínica 
        del hospedero, o (ii) _métodos teóricos_ empleando algoritmos computacionales predictivos 
        que tomen en cuenta la afinidad entre 
        péptidos del patógeno y moléculas 
        del complejo principal de histocompatibilidad (MHC) del hospedero.
        La principal ventaja del primer método
        está en la capacidad de priorizar la selección de antígenos en
        base a criterios clínicamente relevantes, como el de individuos con inmunidad por 
        adquisición natural [@immunomics2016].
        
        Este método ha permitido obtener datos empíricos sobre 
        diferencias en la amplitud, intensidad, cinética y longevidad de respuestas inmunes
        inducidas por patógenos. Además de mayor información sobre aspectos clave como el
        porcentaje del proteoma reconocido por el hospedero, la capacidad predictiva
        de anticuerpos sobre el estado de la enfermedad o de las
        secuencias de aminoácidos sobre su inmunogenicidad [@Davies2015Large].
        Dos plataformas representativas son los microarreglos de proteínas
        y péptidos, este último con mayor resolución en el mapeo de epítopes 
        [@carmona2015peptide].
-->

<!-- 
han alimentado a los actuales algoritmos de predicción computacional

Infecciones experimentales de _P. vivax_ en Colombia con seguimiento longitudinal 
por 145 días han demostrado que la dinámica de la respuesta humoral 
diverge entre individuos sin exposición previa (_naive_) y semi-inmunes, 
en base estos dos parámetros [@arevalo2016].
-->

### Microarreglos de proteínas

Esta es una técnica empleada en la medición a larga escala de las interacciones o 
actividades de proteínas capaz de reconstruir redes biológicas entre 
diferentes clases de moléculas y estados celulares [@uzoma2013interactome].
Su empleo en el contexto _inmunómico_, con la cuantificación de las interacciones antígeno-anticuerpo,
<!--ha identificado biomarcadores de exposición o protección ante distintos patógenos
mediante perfiles de respuesta humoral en individuos infectados o vacunados.-->
posee características analíticas similares a los microarreglos de ADN [@sundaresh2006],
lo que ha facilitado la adaptación de los componentes estandarizados para 
el diseño de experimentos en esta tecnología [@allison2006].
<!--además de parámetros de relevancia 
exclusiva para el estudio de la respuesta humoral.-->
<!--Ambas plataformas buscan resolver el mismo problema: 
    realizar un análisis diferencial de señales de luz 
    derivadas de eventos de unión molecular.-->
    <!-- Finalmente, se detallarán las 
    ventajas del uso de software libre para su ejecución, con respecto a los privativos
    y servidores web.-->

a. Componentes
    
    i. __Diseño__
        
        El desarrollo de un plan de experimentación tiene el objetivo de maximizar la calidad
        y cantidad de la información obtenida.
        Una de las secciones clave dentro del MIAME o "información mínima sobre experimentos basados 
            en microarreglos" [@brazma2001] es el
            detalle de la información de los elementos a incluir 
            (e.g., secuencias de moléculas correspondiente a genes específicos).
        El primer diseño de microarreglos para _P. vivax_ y _P. falciparum_ fue estandarizado 
        por Finney et al. 2014 [@Finney2014] en base a la evidencia transcriptómica, proteómica
        y características inmunogénicas de proteínas predichas en ambos genomas.
        Posteriormente, en el 2015 el Centro Internacional de Excelencia para la Investigación de la 
        Malaria (ICEMR) realizó una subselección empírica de este microarreglo 
        con 142 muestras de 10 paises con malaria endémica, entre ellos Perú [@King2015FOC]. 
        
        <!--El primer diseño de microarreglos para _P. vivax_ y _P. falciparum_ fue estandarizado 
        por Finney et al. 2014 [@Finney2014]
        Seis características fueron empleadas: (i) 
        proteínas predichas para el estadio eritrocítico con capacidad de ser 
        expuestas a superficie, (ii) evidencia en microarreglos, proteomas, EST, 
        o secretoma predicho para _P. falciparum_, (iii) proteínas únicas y ortólogas con
        presencia de péptidos señal (SP) o dominios transmembrana (TMD) para _P. vivax_, (iv)
        proteínas putativas citoplasmáticas carentes de SP o TMD en base a filtros
        de punto isoeléctrico por especie, (v) exclusión de proteínas con alta variabilidad 
        antigénica (e.g., PfEMP1, RIFIN, surfin, stevors; Pv _var_ y pseudogenes), 
        e (vi) inclusión de genes simples y multi-exónicos.-->
        
        <!--Posteriormente, en el 2015 el Centro Internacional de Excelencia para la Investigación de la 
        Malaria ICEMR realizó una subselección empírica del microarreglo anteriormente
        detallado [@King2015FOC]. 
        Para ello sondearon los microarreglos de _P. falciparum_ con muestras 
        de Papúa Nueva Guinea, Kenya y Mali, mientras que
        los de _P. vivax_ con muestras de Papúa Nueva Guinea, China, Perú y Tailandia,
        ambos con controles de individuos norteamericanos sin exposición a malaria.
        A partir de ello<!-- se seleccionó a todos los antígenos seroreactivos por 
        país, cumpliendo la condición de ser mayores a dos veces la desviación estándar
        de la media de la reactividad serológica en los controles. Finalmente, 
        se seleccionó el top 500 de antígenos para ambas especies empleando un filtrado jerárquico,
        dándole prioridad a los elementos con seroreactividad en todos los países<!-- y
        en las posiciones restantes los antígenos en orden descendiente a su reactividad
        promedio.<!--Este diseño se encuentra en la base de datos GEO
        con el código [GPL18316](https://www.ncbi.nlm.nih.gov/geo/query/acc.cgi?acc=GPL18316).-->
        
        <!--El tamaño y complejidad de los genomas de _Plasmodium_ desafían el diseño de
        microarreglos a una escala de proteoma completo.
        Por ejemplo, _P. falciparum_ posee más de 3kb de longitud, ~5300 genes, genes 
        multi-exónicos complejos y segmentos largos de secuencias repetitivas por el bajo 
        contenido de GC.
        Por esta razón, se requiere de procesos de selección en base
        a características proteicas y conocimiento empírico. <!--De esta manera se ha logrado 
        desarrollar plataformas costo-efectivas y rápidas de emplear, capaces de facilitar el 
        descubrimiento de antígenos para vacunas o serovigilancia con el menor trabajo en 
        laboratorio e independiente de algoritmos predictivos.-->
        
        <!--i. __Primera generación__
        
        Finney et al. 2014 [@Finney2014] diseñaron microarreglos de _P. vivax_ y 
        _P. falciparum_ en base
        a (i) proteínas predichas para el estadio eritrocítico con capacidad de ser 
        expuestas a superficie, (ii) evidencia en microarreglos, proteomas, EST, 
        o secretoma predicho para _P. falciparum_, (iii) proteínas únicas y ortólogas con
        presencia de péptidos señal (SP) o dominios transmembrana (TMD) para _P. vivax_, (iv)
        proteínas putativas citoplasmáticas carentes de SP o TMD en base a filtros
        de punto isoeléctrico por especie, (v) exclusión de proteínas con alta variabilidad 
        antigénica (e.g., PfEMP1, RIFIN, surfin, stevors; Pv _var_ y pseudogenes), 
        e (vi) inclusión de genes simples y multi-exónicos.
        
        ii. __Segunda generación__
        
        En el 2015, el Centro Internacional de Excelencia para la Investigación de la 
        Malaria ICEMR realizó una subselección empírica del microarreglo anteriormente
        detallado [@King2015FOC]. 
        Para ello sondearon los microarreglos de _P. falciparum_ con 20 muestras 
        de Papúa Nueva Guinea, 20 de Kenya, 20 de Mali y 10 controles norteamericanos, y 
        los de _P. vivax_ con 15 de Papúa Nueva Guinea, 15 de China, 22 de Perú, 10 de 
        Tailandia y 10 controles. Luego se seleccionó a los antígenos seroreactivos por 
        país, cumpliendo la condición de ser mayores a dos veces la desviación estándar
        de la media de la reactividad serológica en los controles. Finalmente, 
        se seleccionó el top 500 para ambas especies empleando un filtrado jerárquico,
        dándole prioridad a los antígenos con seroreactividad en todos los países y
        en las posiciones restantes los antígenos en orden descendiente a su reactividad
        promedio. Este diseño ha sido depositado en la base de datos GEO
        con el código [GPL18316](https://www.ncbi.nlm.nih.gov/geo/query/acc.cgi?acc=GPL18316).-->
        
        <!--1. __Diseño__
            
            De acuerdo al MIAME o información mínima sobre experimentos basados 
            en microarreglos [@brazma2001], se requiere definir 6 secciones a 
            ser detalladas en el [capítulo 4](#meto): 
            - Diseño experimental:
            descripción del conjunto total de experimentos a ejecutar;
            - Diseño de arreglos: 
            detalle de la información de los elementos a incluir en los _spots_; 
            - Muestras:
            descripción de la fuente de la muestra y el criterio para su clasificación;
            - Hibridación:
            descripción de las condiciones de laboratorio bajo las cuales se realizará;
            - Mediciones:
            descripción del progreso de las imágenes escaneadas a la matriz de datos; y
            - Controles de normalización:
            descripción de los elementos conocidos o invariantes.-->
            
    ii. __Pre procesamiento__ 
            
        Con el objetivo de remover variaciones sistemáticas en el experimento, 
        se ejecutan en forma consecutiva los tres siguientes pasos:
        <!--Incluye la normalización, transformación y filtrado de la matriz de datos.-->
            <!--También a los criterios bajo los cuales se controlará la calidad de las
            mediciones del experimento.-->
        1. Normalización: Permite homogeneizar la variabilidad entre arreglos
            en base a controles, los cuales se asume que son invariantes entre muestras. 
            Este procedimiento es ejecutado tradicionalmente a través de la 
            sustracción de controles o _fold over control_ (FOC),
            obteniendo el cociente de las lecturas objetivo con respecto 
            a la mediana de los controles negativos por muestra [@King2015FOC];<!--[@arevalo2016]-->
            <!--
            De acuerdo al diseño de arreglos, este procedimiento es 
            ejecutado de tres formas:
            - sustracción de controles o _fold over control_ (FOC),
            aplicado en microarreglos de proteínas con respecto a la mediana del control negativo
            por muestra [@King2015FOC]
            - estabilización de varianzas o VSN, 
            aplicado en microarreglos de DNA con respecto a controles de 
            intensidad invariante entre muestras, e.g. genes _housekeeping_ [@huber2002vsn]; y
            - modelo robusto lineal o RLM,
            con la capacidad de controlar la variabilidad entre muestras
            observada en los diseños experimentales de microarreglos de proteínas 
            [@sboner2009rlm].
            -->
        2. Transformación: Reduce la
            heterocedasticidad o heterogeneidad de varianzas entre las observaciones 
            por errores aleatorios en el proceso de hibridación [@kreil2005bullet].
            Este fenómeno se evidencia en el incremento de la varianza <!--de las medidas -->
            directamente proporcional al incremento de su intensidad [@brown2001image].
            Dependiendo de la dispersión de los datos, se asumen modelos de error
            multiplicativo o aditivo, donde la transformación logarítmica o
            arcoseno hiperbólico es ejecutada sobre los valores normalizados.
            <!--
            Para ello, dos principales transformaciones han sido definidas:
            - Logarítmica (log),
            la cual asume un error multiplicativo que corrige las varianzas de 
            alta intensidad pero capaz dispersar a las de baja intensidad; y
            - Arcoseno hiperbólico (asinh),
            la cual asume un error multiplicativo y aditivo que corrige esta dispersión, 
            relevante en contextos donde los valores normalizados menores a la unidad 
            tienen significado biológico.
            -->
        3. Filtrado: En experimentos a larga escala permite incrementar
            el poder de detección de elementos con expresión diferenciada [@bourgon2010filter].
            En estos, las estrategias de corrección posteriores a la comparación de múltiples
            hipótesis son sensibles a esta cantidad.
            En este sentido, este procedimiento retira de forma preliminar a los elementos 
            con reducidas probabilidades de expresarse diferencialmente.
            <!-- sin la 
            necesidad de realizar la prueba de hipótesis.
            -->
        
    iii. __Inferencia__ 
    
            Su objetivo es poner a prueba hipótesis estadísticas relacionadas 
        a la expresión o reactividad diferencial de elementos entre grupos.
        Sin embargo, a esta escala se posee dos principales problemas: (1)
        presencia de varianzas artificialmente altas o bajas producto del 
        bajo número de réplicas entre arreglos o muestras [@baldi2001cybert] y (2) un
        amplio número de hipótesis puestas a prueba en forma simultánea, donde los 
        test estadísticos tradicionales pueden generar un largo número de 
        falsos positivos [@kayala2012cyber].
        
            <!--
        (i) poner a prueba la hipótesis nula de igualdad de medias en las expresión
        de elementos por grupo y (ii) ordernar o ranquear a los elemento según la
        evidencia contra la hipótesis nula, incluso luego de fallar en el rechazo
        de dicha hipótesis [@smyth2004ebayes]. 
        -->
        
            1. Test-t moderado: 
            Un enfoque bayesiano empírico permite reducir (o moderar) las varianzas de 
            todas las lecturas. Bajo el supuesto que las diferencias entre grupos 
            generan cambios en la intensidad promedio del gen mas no en su varianza,
            se estima una varianza general como probabilidad previa para actualizar (corregir)
            todas las varianzas observadas en el experimento [@smyth2004ebayes]
            <!--
            Inicialmente, cada varianza observada por gen se corregía con respecto a
            la varianza de genes vecinos con medias de expresión 
            similar [@baldi2001cybert].
            Posteriormente, esta metodología se extendió y optimizó mediante la estimación de 
            una varianza general previa para la actualización de todas las varianzas 
            observadas [@smyth2004ebayes].
            -->
            <!--
            A diferencia de un enfoque bayesiano puro, donde se asume una 
            probabilidad previa para su actualización dado un evento condicionante,
            en un enfoque empírico esta información es obtenida de un _pool_ de
            elementos observados en el mismo experimento.
            -->
            Con ello, <!--la ejecución de un test-t con varianzas moderadas 
            (o test-t moderado)-->es posible obtener inferencias más estables que el 
            test-t ordinario bajo un contexto limitado de réplicas [@kayala2012cyber].
            2. Razón de falsos descubrimientos (FDR):
            Los métodos de corrección o ajuste de valores P permiten que el 
            investigador pueda controlar la razón de falsos descubrimientos con 
            respecto al total de hipótesis positivas [@brazma2001].
            El método Benjamini-Hochberg
            determina un valor crítico dependiente del total 
            de hipótesis puestas a prueba y el FDR que se desee tolerar,
            comúnmente del 5% [@benjamini1995fdr].
            <!-- EXPLAIN STOREY's METHOD! -->
            <!--
            Dos principales enfoques han sido propuestos. El enfoque
            probabilístico emplea modelos mixtos sobre la distribución de valores p
            para diferenciar a las poblaciones de hipótesis nulas y alternativas
            [@allison2002mmm].
            El enfoque frecuentista emplea el método Benjamini-Hochberg basado en
            la determinación de un valor crítico de valores P dependiente del total 
            de hipótesis puestas a prueba y el valor de FDR deseado [@benjamini1995fdr].
            -->
            
        
    iv. __Clasificación__
            
        <!--Una aplicación importante de los microarreglos de proteínas está en la
            identificación de un conjunto de antígenos de serodiagnóstico que 
            puedan detectar patrones y clasificar perfiles de respuesta humoral
            [@sundaresh2007].
            Por ello, el esquema de ranqueo de elementos comúnmente ha sido complementado 
            con técnicas de clasificación.-->
        Este proceso involucra la asignación de objetos (e.g., genes) en categorías
            pre-existentes (clasificación supervisada), o el desarrollo progresivo de 
            un conjunto de categorías por características de los objetos.
            (clasificación no supervisada) [@allison2006].
            Ejemplos de esta técnica son
            _support vector machines_ (SVM) con validación
            por _leave-one-out cross-validation_ (LOOCV) y
            el agrupamiento jerárquico o _hierarchical clustering_
            de acuerdo a la distancia Euclidiana entre elementos.
        
        <!--
            La clasificación supervisada o predicción de clase, define _a priori_ 
            el número de categorías, el conjunto de entrenamiento y el de prueba. 
            Requiere del
            uso de procedimientos de validación cruzada y tomar en cuenta sesgos de
            selección. Por otro lado, su precisión predictiva es afectada por la 
            sobreparametrización, más aún con bajo número de muestras y amplia 
            cantidad de elementos u objetos. Ejemplos de esta técnica son
            _support vector machines_ (SVM), _k-nearest neighbors_ (k-NN) y validación
            por _leave-one-out cross-validation_ (LOOCV).
            
            La clasificación no-supervisada o descubrimiento de clase, agrupa objetos 
            en base a métricas de similaridad de los objetos a lo largo de las muestras.
            Para ello requiere definir 2 tipos de algoritmos: de distancia y agrupamiento.
            Ejemplos del primero son la distancia Euclidiana, en la que se prioriza la 
            magnitud sobre la dispersión de señales, y la correlación de 
            Pearson, en la que se prioriza la dispersión sobre la magnitud.
            Ejemplo del segundo es el agrupamiento jerárquico o _hierarchical clustering_,
            en la que los racimos o _clusters_ se forman iterativamente al calcular 
            la distancia entre elementos y racimos en formación. Otros son el 
            _k-mean clustering_, _self-organizing maps_ (SOM) y 
            _principal component analysis_ (PCA).
            Las principales desventajas de la técnica están en el requerimiento 
            de decisiones entre varios algoritmos sin consenso y su escasa reproducibilidad
            entre experimentos.
            -->
        
        <!-- 4. __Validación__ ||| PARA LA DISCUSIÓN o RECOMENDACIONES!
            (i)Estrategia de 3 fases para la identificación de biomarcadores!!!
            [@song2009novel; @uzoma2013interactome]
            I Selección (baja escala) -> 
            II Validación (cohorte, descarte de comorbilidad, inespecificidad, 
            reactividad cruzada con otras enfermedades o patologías) -> 
            III Performance (ELISA u gold-standard)
            (ii) Validación constructiva (probar Ho usando diferentes metodologías)
            (iii) Validación Operacional x 2 fuentes de error
            I error aleatorio de medición, sesgan resultados hacia la Ho (Falsos Negativos)
            -> usar medicion más precisa o mayor tamaño muestral
            II error aleatorio de muestreo, por la probabilidad de encontrar dos 
            poblaciones con reactividad diferenciada para un elemento (Falsos Positivos)
            -> Usar nuevos casos con mismo diseño experimental-->
            
    iv. __Parámetros adicionales__
        
        Además de la reactividad serológica, dos parámetros han demostrado ser
        de vital importancia para la caracterización de la respuesta humoral: 
        la amplitud e intensidad de respuesta 
        [@crompton2010; @Helb2015exposure; @King2015FOC].
        El primero se obtiene calculando el número de antígenos 
        reactivos por individuo
        y el segundo por el promedio de sus intensidades<!-- por individuo-->.
        Particularmente, la amplitud ha demostrado ser tan relevante como la
        reactividad de antígenos individuales en el desarrollo de inmunidad contra
        la malaria [@crompton2010].
        <!--Los perfiles de anticuerpos a larga escala han demostrado la amplitud y
        heterogeneidad de las respuestas humorales [@crompton2010].
        __Intensidad de respuesta:__
        Promedio de intensidades de antígenos reactivos por individuo
        __Amplitud de respuesta:__
        Número de antígenos reactivos por individuo.
        -->

<!--b. Plataformas de análisis
    
    <!--i. __Privativo__
        
        La amplia variedad de herramientas desarrolladas para la solución de 
        cada uno de los componentes mencionados ha generado una alta 
        heterogeneidad de metodologías observadas en la literatura.
        Este problema limita la reproducibilidad de estos estudios debido al 
        reporte excesivo o escaso de parámetros empleados, 
        además de desacelerar el progreso del área. Ejemplos de ello son 
        Molina et al. 2012 (Cyber-T y GraphPad Prism) [@molina2012],
        Baum et al. 2013, 2015 (SAM, MeV, JMP9.0 y Systat11) [@baum2013; @baum2015],
        Lu et al. 2014
        (GraphPad, SigmaPlot10.0, Cluster3.0, TreeView y pearsoncorrelation.com)
        [@lu2014rama], y
        Chen et al. 2015 (SAS, MULTTEST y TIGR MeV) [@chen2015immunomics].
        <!-- PRE-PRINT of TIPS TP MAKE YOUR RESEARCH IRREPRODUCIBLIBLE (2015)
        
    ii. __Libre__
    
    La amplia variedad de herramientas desarrolladas para la solución de 
        cada uno de los componentes mencionados ha generado una alta 
        heterogeneidad de metodologías observadas en la literatura.
    En respuesta a esta problemática, el uso de lenguajes de programación
        libre como R [@R2016] junto a ambientes de análisis como Bioconductor 
        [@bioconductor2004] ha
        homogenizado protocolos,
        reducido las barreras de acceso a la investigación interdisciplinaria
        y facilitado el análisis transparente de datos moleculares de larga escala.
        Su flexibilidad a la
        creación de _scripts_ de análisis o _workflows_ permite combinar diferentes
        tipos de datos (e.g., epidemiológicos y genómicos), 
        herramientas estadísticas y visualizaciones [@Biobase].
        Finalmente, el uso de reportes dinámicos [@knitr]
        también ha permitido el
        registro explícito de ediciones a data cruda y parámetros
        empleados en el análisis, siguiendo principios de 
        Ciencia Reproducible [@CienciaReproducible2016].

<!--
        3. __Servidores web__
            
            Desventajas otorgadas a los software privativos por sus
            limitados métodos estadísticos y capacidades visuales, 
            y a las alternativas libres por las
            habilidades computacionales requeridas para su uso, han estimulado el
            desarrollo de servidores web como GMine [@gmine2016]
            para el análisis integral y exploratorio.
            Sin embargo, la falta de reporte o registro de los de parámetros empleados
            nuevamente cae en una escasa o dilatada reproducibilidad no automatizada.
-->

## Definiciones conceptuales

a. Acrónimos
    
    __IVTT:__
    (Proteína) transcrita/traducida _in vitro_.
    
    __MFI:__
    Unidad de lectura cruda expresada como
    intensidad fluorescente promedio
    de todos los píxeles de cada _spot_, 
    normalizada localmente mediante la
    sustracción de la intensidad de fondo presente 
    a su alrededor.
    
    __RTS:__
    Sistema rápido de traducción de proteínas recombinantes libre de células.
    
b. Términos
    
    <!-- 
    **Malaria.-** Enfermedad transmitida por mosquitos, 
    predominantemente _Anopheles darlingi_ en América del Sur, 
    causada por un parásito del género _Plasmodium_.
    
    **Malaria severa.-** Clasificación de los pacientes con malaria 
    según criterios clínicos y de laboratorio.
    -->
    
    <!--i. Generales
        
        **Antígeno.-** Molécula que se une con productos de la respuesta inmune, 
        como anticuerpos o receptores de linfocitos T o B [@abbas2012].
        
        **Inmunógeno.-** Un antígeno que induce una respuesta inmunitaria [@abbas2012].
        
        **Inmunoma.-** Conjunto de todos los inmunógenos que interaccionan con 
        el sistema inmune de determinado hospedero [@immunomics2016; @sette2005].
        
        **Inmunómica.-** Estudio del inmunoma [@immunomics2016].
        
        **Anticuerpo.-** Tipo de molécula glucoproteica, también llamada 
        inmunoglobulina (Ig), producida por los linfocitos B, que se une a antígenos 
        con un grado alto de especificidad y afinidad [@abbas2012].
        
        **Anticuerpo secundario.-** Anticuerpo de unión específica a la fracción constante 
        de los anticuerpos de un hospedero, comúnmente conjugado con biotina.
        
        **Fluoróforo.-** Componente de una molécula que brinda la cualidad 
        de fluorescencia, comúnmente conjugada con estreptavidina para la detección de 
        moléculas biotiniladas.
        
    ii. Específicos-->
        
    __Antígeno-IVTT:__
        Antígeno objetivo o 
        _spot_ con proteína IVTT a partir de un plásmido con ADN insertado 
        correspondiente al polipéptido, segmento o exón de interés.
        
    __Control-IVTT:__
        Control negativo o 
        _spot_ con mix de expresión RTS y plásmido sin ADN insertado,
        representante de la intensidad de fondo específica del paciente.
        
    __Proteína purificada:__
        Control de comparación o 
        _spot_ con proteína de antigenicidad conocida expresada dentro de célula.
        
    __Intensidad de antígeno:__
        Lectura normalizada de cada antígeno-IVTT entre individuos. 
        También llamada "reactividad serológica".
        
    __Antígeno reactivo:__
        Lectura transformada de cada antígeno-IVTT mayor o igual a 
        dos veces la mediana de los control-IVTT por individuo.
        
    <!--__Frecuencia del antígeno:__
        Número de individuos con antígeno reactivo por antígeno-IVTT. -->
        
        
# MATERIAL Y MÉTODOS {#meto}

## Material biológico (Población y muestra)

### Población

<!--El universo teórico está representado por los pacientes con malaria vivax 
de la cuenca amazónica del Perú.-->
<!--El marco muestral-->

La población está delimitada por pacientes diagnosticados con malaria por _P. vivax_ 
<!--y clasificados como severos y no-severos-->de la ciudad de Iquitos, Loreto - Perú, 
entre enero del 2012 y junio del 2013 a lo largo de un estudio prospectivo 
ejecutado en dos hospitales de referencia: Hospital de Apoyo y Hospital Regional.

### Criterios de inclusión y exclusión

a. Criterios de inclusión
    
    En el estudio prospectivo mencionado, 
    164 pacientes con malaria por _P. vivax_ fueron enrolados mediante vigilancia pasiva.
    Según la presencia de uno o más criterios clínicos y de laboratorio de la
    clasificación OMS para la 
    malaria severa 
    [@WHO2014severe], los pacientes fueron 
    estratificados como severos o no-severos.
    <!-- OJO: REDEFINIR LUEGO DE REVISAR sevcov!! -->
    
    Los criterios clínicos de la clasificación OMS son:
    Shock circulatorio (presión sanguínea sistólica < 80 mmHg), 
    Deterioro del nivel de conciencia (puntaje Glasgow $\le$ 9/14),
    Daño del sistema nerviosos central (convulsión),
    Daño pulmonar (síndrome de dificultad respiratoria aguda o edema pulmonar).
    
    Los criterios de laboratorio de la clasificación OMS son:
    Hipoglicemia (glucosa < 40 mg/dL), 
    Anemia severa (hemoglobina < 7g/dL), 
    Daño renal (creatinina > 3mg/dl), e
    Hiperbilirrubina (bilirrubina sérica > 2.5 mg/dL).
    <!-- exluidos de Malaria Severa 2014: sangrado, acidosis, hiperparasitemia -->
    
b. Criterios de exclusión
    
    Dos criterios de exclusión fueron empleados: (i) Presencia de coinfecciones con especies de 
    _Plasmodium_ distintas a _P. vivax_ determinadas por PCR, y (ii)
    presencia de coinfecciones como leptospirosis, dengue u otra arbovirosis determinada por 
    técnicas de aislamiento viral e inmunofluorescencia.
    <!-- Auscencia de mono-infecciones de _P. vivax_ determinadas por PCR y -->
    
### Selección de participantes

A partir de esta estratificación, se ejecutará 
una selección arbitraria de 30 pacientes representativos de malaria severa
con uno o más criterios de la clasificación OMS para el grupo de casos y 
una selección aleatoria simple de 30 no-severos para el grupo control.

- Tipo de muestra: No probabilística para los casos y probabilística para los controles. 
Los elementos del grupo caso provendrán de una selección arbitraria,
mientras que el grupo control será seleccionado al azar, 
i.e. con probabilidad de selección conocida.


## Material de laboratorio

### Instrumento de medición 
<!--y recolección de datos-->

<!-- 
**Microarreglo de proteínas.-** 
1. Técnica de larga escala para rastrear las interacciones o  actividades de proteínas.
2. Matriz con _spots_ de polipéptidos IVTT, 
impresos en una lámina portaobjetos y 
organizados en 8 bloques paralelos, 
cada uno con 4 arreglos compuestos por cuadrículas de 17x17 _spots_.

El microarreglo de proteínas es una herramienta que permite medir
a larga escala los anticuerpos reactivos a determinados antígenos de 
un patógeno presentes en plasma.

ADN 300-3000 aa
entre 100 y 1000 aminoácidos de longitud, 
En este caso, la matriz posee 515 fragmentos protéicos,
representantes de 446 proteínas, aproximadamente el 8% del genoma de _P. vivax_ (Sal1).
8.21% del genoma Pvivax Sal1 -->

El presente estudio hará uso de 
microarreglos diseñados con 1014 fragmentos proteicos recombinantes 
(498 de _P. falciparum_ y 516 de _P. vivax_, expresado como Pf498/Pv516 o PfPv500) 
representando 873 proteínas predichas (427 de _P. falciparum_ y 446 de _P. vivax_, 
~8% del total predicho para el genoma de _P. vivax_ Sal1) 
seleccionadas luego de una extensiva evaluación serológica con uno 
de mayor escala (Pf2208/Pv2233) [@Finney2014].
<!--Las proteínas son transcritas/traducidas _in vitro_ 
empleando un sistema de expresión de _Escherichia coli_ libre de células e 
impresas en bloques de nitrocelulosa sobre una lámina portaobjetos 
con 8 bloques paralelos, 
cada uno con 4 arreglos paralelos compuestos por cuadrículas de 17x17 _spots_.-->
Este diseño ha sido validado mediante el sondeo con plasma 
colectado de pacientes con malaria y controles sanos alrededor del mundo [@King2015FOC]
y depositado en la base de datos GEO
([GPL18316](https://www.ncbi.nlm.nih.gov/geo/query/acc.cgi?acc=GPL18316)).

<!--
a utilizar fue desarrollado
por el Dr. Phil Felgner y el Dr. Huw Davis de la Universidad de
California Irvine. 
Brevemente, 
se utilizará la tecnología de alto
rendimiento de un sistema de traducción para expresar 500 proteínas de
P. falciparum y 500 proteínas de P. vivax, que representan aprox. 10%
de proteoma del parásito. Las proteínas fueron impresos en un
portaobjetos y se probaron con plasma diluido 1/200 después de
bloquear con tampón de bloqueo (Whatman).

FABRICACIÓN INCLUIDA EN LIMITACIONES!! -> VA PARA LA TESIS

-->

#### Aplicación
<!--de recolección de datos
Sobre el procedimiento para procesar muestras para microarreglos de proteínas!

EM SERIO 2uL POR MUESTRA????!!!!!!


PAPER METHODS!!

Las muestras de plasma sanguíneo (2uL de plasma por microarreglo) se
depositarán en las lámina con microarreglos, una muestra por bloque. 

Los valores de fluorescencia serán medidos a través de un lector de
microarreglo con láser confocal, la cual otorgará valores para cada
uno de los arreglos de interacción proteína-anticuerpo, obteniéndose
una matriz final con valores numéricos.

-->

La aplicación del instrumento consiste en tres pasos: sondeo, escaneo y análisis, 
tal como ha sido publicado previamente [@Driguez2015].
Primero, previo al sondeo los microarreglos se hidratan con buffer de bloqueo dentro de 
cámaras de incubación. Paralelamente, el plasma de los pacientes se diluye con aquel 
buffer en 1:100 y se pre-absorbe con lisado de _E. coli_ en 10%(w/v), con el objetivo de 
reducir ruido de fondo en las mediciones tanto por uniones inespecíficas con el sustrato 
como por antígenos bacterianos del sistema de expresión RTS. Segundo, aspirado el buffer de 
bloqueo del microarreglo, se procede con el sondeo al agregar el plasma pre-absorbido e 
incubar _overnight_ en cámara húmeda a 4 °C con leve agitación. Tercero, con lavados y 
aspirados repetitivos antes y después de cada paso, se agrega la solución con anticuerpos 
secundarios biotinilados y posteriormente la solución con fluoróforos conjugados con 
estreptavidina. Cuarto, luego de centrifugar las láminas, se procede al escaneo de las 
fluorescencias con lectores de microarreglos de láser confocal (e.g., Genepix 4300A). Las 
medidas crudas se obtienen al aplicar una normalización local mediante la sustracción de la 
intensidad de fondo presente alrededor de cada _spot_, ejecutada en el software del lector 
(Genepix Pro 7). 
<!-- Finalmente, se procede al análisis de datos detallado en la [sección 4.4](#anadata). -->

#### Validez {#validez}

<!--de medición y recolección
 
 DOS PUNTOS RELEVANTES: 
VALIDEZ (PRECISIÓN) -> PURIFIED vs IVTT
REPRODUCIBILIDAD (CONFIABILIDAD) -> DAY1 vs DAY2

 PREVIO AL SONDEO PRINCIPAL
Luego, se ejecuta un sondeo preliminar con anticuerpos dirigidos a extremos N- y C- 
terminales, con colas 10x HIS (histidina) y HA (hemoaglutinina) con el objetivo de realizar 
un control de la calidad de la expresión e impresión de proteínas.

Se agregarán diluciones de proteínas recombinantes expresadas en un
sistema celular de reactividad conocida, controles positivos para
anticuerpos secundarios, control positivo de anticuerpos específicos
anti-hospedero, antígenos de extracción del parásito, control negativo
sin plásmidos de ADN y control negativo con solo buffer. Se colectarán
individuos controles de origen norteamericano que no hayan tenido
contacto alguno con la infección.

estará incluida en la [sección 4.4](#anadata).

 -->

La validez o exactitud del experimento será evaluada mediante la correlación 
lineal entre las lecturas de los antígenos-IVTT y sus correspondientes proteínas 
purificadas por muestra, compuestas por proteínas de inmunogenicidad conocida y expresadas 
por un sistema dentro de célula [@crompton2010].
<!--Segundo, la confiabilidad o reproducibilidad será evaluada mediante la correlación lineal 
entre controles positivos agregados a las lecturas del primer y segundo día de sondeo de 
las muestras.-->

#### Adquisición

Todos los reactivos, incluidos los microarreglos, anticuerpos secundarios biotinilados 
y fluoróforos conjugados con estreptavidina serán adquiridos a la empresa 
Antigen Discovery, INC [@adiinc].

## Métodos:

<!--# METODOLOGÍA {#meto} -->

<!-- 
# ASPECTOS METODOLÓGICOS DE LA INVESTIGACIÓN

Población, muestra y muestreo

	Población

Describir el contexto del estudio y los sujetos a estudiar
Especificar el lugar/ámbito en que se hará el estudio.
Características de la población de estudio

Criterios de inclusión

Especificar los criterios de inclusión  del estudio.

Criterios de exclusión

Especificar los criterios de exclusión del estudio. No deben repetir lo establecido en los 
criterios de inclusión.

Selección de participantes

Describir cómo se hará la selección de los participantes para el estudio. Por ejemplo: Se 
tomaran aleatoriamente a un grupo de alumnos que estén matriculados en el curso de 
Patología.

Variables del estudio

Describir las variable principales del estudio (variable dependiente e independiente)
Especificar lo siguiente para cada variable:
Variable (indicar el nombre de la variable y si es la variable dependiente o independiente 
del estudio)
Definición operativa
Medición (cómo se medirá y/o cómo se recolectará el dato)
Valores posibles
Fuente de la variable

Luego de hacer una descripción detallada, es útil resumir las variables de estudio en una 
tabla.
-->

### Diseño

El diseño del estudio es de tipo caso-control.
Su aplicación consistirá en la selección de la población 
objetivo por presencia (casos) o ausencia (controles) del evento en estudio.
Además se fijará el número de eventos a estudiar, así como el número de sujetos 
sin evento que se incluirán en la población de comparación.

#### Tipo de investigación

- Por la finalidad: Analítico. A diferencia de un descriptivo, realizaremos comparaciones 
entre grupos y evaluaremos una posible relación causal entre el factor y el efecto.

- Por el control de la asignación: Observacional. A diferencia de un experimental, no 
controlaremos la asignación de los factores (severidad y episodios previos). Es decir,
procederemos a observar el fenómeno, ejecutar la medición y analizar los resultados.

- Por el seguimiento: Transversal. A diferencia de uno longitudinal, no ejecutaremos un 
seguimiento. Las variables se medirán una sola vez, en un mismo instante.

- Por la relación cronológica: Prospectivo. A diferencia de un retrospectivo, 
recolectaremos los datos después de planificado el estudio.

<!--__Por la dirección temporal: Hacia atrás/adelante.__ La población será selecionada por 
la variable Efecto/Exposición-->

<!--__Por la unidad de análisis: Basado en individuo.__ A diferencia de uno ecológico, la 
unidad de observación son individuos o pacientes.-->

<!--
**Por el tipo de conocimiento previo usado en la investigación:** Científica
**Por la naturaleza del objeto de estudio:** Factual Natural
**Por el tipo de pregunta planteada:** Teórica Descriptiva de Relaciones no causales
El estudio analizará las diferencias en una misma población 
con respecto a dos variables de interés.
**Por el método de contrastación de la hipótesis:** No experimental
Se describirá al evento bajo diferentes
variables para la organización de la información, mas no para
determinar una causalidad.
**Por el método de estudio de las variables:** Cuantitativa
Evidenciado en la medición de la
intensidad de fluorescencia en la reacción antígeno-anticuerpo
**Por el número de variables:** Univariada
El estudio, a pesar de poseer varias variables, 
no establecerá relación de efecto entre ellas.
**Por el ambiente en que se realizan:** De laboratorio
La data será generada en laboratorio. 
El presente trabajo se centrará en el análisis de la data.
**Por el enfoque utilitario predominante:** Teórica
Cuyos resultados permitirán conocer nuevas alternativas para
el uso de biomarcadores en vigilancia seroepidemiológica.
**Por la profundidad:** Exploratorio
Se validarán las hipótesis formuladas.
**Por el tiempo de aplicación de la variable:** Transversal
No se realizará una seguimiento de la población.
-->



### Recolección de los datos e instrumento
<!-- Muestreo -->

#### Técnica para la recolección de datos

<!--de muestreo
SOBRE CÓMO SE EXTRAJO LAS MUESTRAS DE LA MUESTRA

POSTER+PAPER METHODS
-->

Previo al tratamiento antimalárico del paciente, 
se extrajeron las muestras de sangre
tanto para el diagnóstico de malaria por _P. vivax_ con la técnica de frotis
como para las pruebas bioquímicas.
Al momento del diagnóstico positivo, 
bajo consentimiento informado y de forma voluntaria, 
el plasma sanguíneo fue colectado y conservado a -80 °C hasta su uso.


### Codificación y creación del archivo de datos

Cada paciente del estudio estará identificado con un código de estructura `LIM####`, 
e.g.: `LIM1063`.
Los antígenos proteicos estarán identificados con el código asignado 
a sus genes en la base de datos PlasmoDB, e.g.: 
`PF3D7_0202500` o `PVX_091315`, 
ya sea un gen de _P. falciparum_ o _P. vivax_, respectivamente.
En caso los genes posean múltiples exónes, 
se amplificarán por separado y se extenderá el código de cada uno
con su número de orden y el total, e.g.: `_1o2` exón 1 de un gen con 2 exónes.
En caso los genes posean una longitud mayor a 3000 nucleótidos (nt),
se dividirán en segmentos sobrelapantes entre 300-3000nt y
se extenderá el código de cada uno con su respectivo número, 
e.g.: `_S1` para el primer segmento de un gen.

Los datos serán organizados en dos matrices: 
(i) archivo `samples.csv` con los códigos de los pacientes y sus covariables 
epidemiológicas y 
(ii) archivo `RawData.csv` con los códigos de las proteínas y sus lecturas crudas en MFI 
por código de paciente.


### Análisis de datos {#anadata}

Se hará uso del software de computación estadística R [@R2016],
complementado con funciones de distintos paquetes pertenecientes a
dos principales ambientes de análisis: 
Bioconductor [@bioconductor2004] 
y Tidyverse [@wickham2016r4ds].<!-- [@wickham2014tidy] -->
<!--
El informe final será generado por un formato que integrará texto, código y resultados, 
siguiendo principios de reproducibilidad [@CienciaReproducible2016].

Para ello se hará empleo de los siguientes paquetes:
`Rmarkdown`, `knitr`, `bookdown`, 
`tidyverse` (`broom`, `dplyr`, `forcats`, 
`ggplot2`, `haven`, `stringr`, `tibble`, `tidyr`),
`Hmisc`, `Biobase`, `limma`, 
`genefilter`, `NMF`, 
`PerformanceAnalytics`, `beeswarm`, `Rmisc`, `htmlTable` y `biobroom`.
-->

#### Control de calidad y pre procesamiento

Preliminarmente, se describirá la distribución y proporción de las
covariables epidemiológicas, clínicas y bioquímicas 
de los pacientes de la muestra.
Luego, con las lecturas crudas en MFI se evaluará la validez <!--y 
reproducibilidad--> del ensayo mediante un
test de asociación entre variables continuas
con la correlación de Pearson ($r$) o Spearman ($\rho$), de acuerdo a la distribución,
asumiendo un error del 5%.
Posteriormente, se procederá con la 
normalización entre muestras de cada antígeno-IVTT 
con respecto a la mediana de los control-IVTT por individuo,
transformación a escala logarítmica en base 2, y 
filtrado de todo antígeno que posea una frecuencia de reactividad menor al 10% 
de individuos en la muestra.
Para ello, un antígeno reactivo será definido <!--conceptualmente como el antígeno-IVTT 
mayor o igual a dos veces la mediana de los control-IVTT por individuo y -->
operacionalmente como el antígeno-IVTT con intensidad normalizada mayor o igual a 1.
Por último, se asociarán ambas matrices de datos en un `ExpressionSet`, 
a través de los códigos de pacientes, 
empleando el paquete `Biobase` [@Biobase].

<!-- 
__Transformación:__
Logaritmo en base dos de los MFI de antígenos-IVTT y control-IVTT.
~~y expresada como _Fold over control (FOC)_.~~

__Normalización:__
Sustracción de la mediana de los control-IVTT a cada antígeno-IVTT por individuo.

__Filtrado:__
Remoción de todo antígeno que posea una frecuencia de reactividad 
menor al 10%.
~~con una intensidad de antígeno mayor o igual a uno y calculado como $log_2(FOC) \geq 1$~~

__Antígeno reactivo:__
Lectura normalizada de cada antígeno-IVTT mayor o igual a 
dos veces la mediana de los control-IVTT por individuo.
-->

#### Prueba de hipótesis
<!-- Inferencia, Clasificación y Descripción -->

Las dos hipótesis de diferencia entre grupos seguirán el siguiente protocolo.
Primero, se contrastará la amplitud e intensidad de respuesta con 
un test de diferencias entre variables continuas y no pareadas de dos grupos
usando t-Student o Mann-Whitney, dependiendo de la distribución,
con un 5% de error asumido.
Segundo, se realizará un test de reactividad diferenciada de anticuerpos entre dos grupos, 
usando el test-t moderado [@smyth2004ebayes] 
con corrección por comparación múltiple de la razón de falsos descubrimientos (FDR)
por el método de Benjamini-Hochberg, disponible en el paquete `limma` [@limma],
con un 5% de error asumido.
Tercero, se realizará un agrupamiento jerárquico o _hierarchical clustering_ de los
antígenos identificados en base a su distancia euclidiana, disponible en el paquete `NMF`
[@Gaujoux2010NMF].
Finalmente, se describirá la<!-- siguientes características proteicas -->
presencia de dominios transmembrana, péptido señal, 
número de ortólogos en _Plasmodium_, ontología génica, 
número de polimorfismos de nucleótido simple y
la razón de mutaciones no-sinónimas 
por sinónimas para el subgrupo de interés,
disponible en la base de datos PlasmoDB [@plasmodb].
<!-- ¿EN QUÉ MOMENTO HABLAS DE PLASMODB EN EL PROYECTO? ¿QUÉ ES PLASMODB? -->

#### Reporte de resultados

<!--De acuerdo a las guía SAMPL [@lang2013basic],-->
Las correlaciones y pruebas de hipótesis serán reportadas con el valor del 
estadístico de prueba y el valor P. Las distribuciones serán visualizadas con
diagramas de dispersión para la correlación entre variables continuas, 
diagramas de cajas y barras para la comparación de variables continuas y frecuencias, respectivamente.
Las comparaciones múltiples reportarán el valor P-ajustado y serán visualizadas con dos
gráficos: un diagrama tipo volcán o _volcano plot_ con el tamaño del efecto o diferencia de reactividad 
entre grupos en escala logarítmica contra sus respectivos valores P<!-- no-ajustados-->
y un mapa de calor o _heat map_ segmentado por racimos o _clusters_.
El informe final consistirá de un reporte en formato `R Notebook` con extensión `.Rmd` 
que integrará texto, código y resultados [@knitr], siguiendo principios de 
reproducibilidad [@CienciaReproducible2016].

<!--# ASPECTOS ADMINISTRATIVOS-->

<!-- \newpage -->

# CRONOGRAMA DE TRABAJO

El Proyecto de Tesis fue trabajado en el Departamento de Parasitología del 
Centro de Enfermedades Tropicales de la Marina de los Estados Unidos NAMRU-6,
del 01 de enero del 2016 al 31 de diciembre del 2016, 
de acuerdo al cronograma de la \autoref{tab:crono}.

<!--En la \autoref{tab:crono} se programan las actividades paralelas y posteriores 
a la redacción del Informe.-->

\begin{table}[ht]
        \captionof{table}{Cronograma}
        \label{tab:crono}
        \vspace{-1mm}
\begin{center}
%\hspace*{-1cm}
\begin{tabular}{lcccccccccccc}
  \hline
  \textbf{ACTIVIDAD} & 
  \textbf{2016} & & & & & & & & & & &\\
  \hline
  & 
  ene & feb & mar & abr & may & jun & jul & ago & set & oct & nov & dic\\
  \cline{2-13}
  Recolección de datos & 
  x & & & & & & & & & & &\\
  \cline{2-13}
  Análisis de datos & 
  & x & x & x & x & x & & & & & &\\
  \cline{2-13}
  Interpretación de resultados & 
  & & & & & x & x & x & & & &\\
  %\cline{2-13}
  %Redacción de proyecto & 
  %& & & & & & & & & & &\\
  %\cline{2-13}
  %Redacción del Informe & 
  %& & & & & & & & & & &\\
  \cline{2-13}
  Redacción de discusión & 
  & & & & & & & x & x & x & &\\
  \cline{2-13}
  Redacción de conclusiones & 
  & & & & & & & & & x & x & x\\
  %\cline{2-13}
  %Correcciones & 
  %& & & & & & & & & & &\\
  %\cline{2-13}
  %Sustentación & 
  %& & & & & & & & & & &\\
  %\hline
  %d1 & 
  %& & & & & & & & & & & & 
  %& & & & & & & & & & &\\
  \hline
  % etc. ...
\end{tabular}
%\hspace*{-1cm}
\end{center}
\end{table}


<!--
Dado que las muestras ya han sido colectadas, solo se procederá a
recolectar lo datos a través del instrumento de medición descrito.

  **ACTIVIDAD PROGRAMADA**
  --------------------------------- ----- ----- ----- ----- ----- -----
                                    __2016__                                __2017__
                                    **E**     **Feb-Oct**   **Nov-Dic**   **Ago-Sep**   **Oct-Nov**   **Dic**  
  Recolección de datos              X
  Análisis de datos                             X
  Interpretación de resultados                            X
  Redacción final                                                   X
  Correcciones                                                                X
  **Sustentación**                                                                      X
-->

# PRESUPUESTO GLOBAL

El presupuesto detallado en la \autoref{tab:presup}
fue ejecutado con financiamiento interno de la institución.
<!--y financiado por \textcolor{red}{\hl{a}}
The Global Emerging Infectious Surveillance and Response System 
(GEIS) section at the 
Armed Forces Health Surveillance Branch (AFHSC),
The Military Infectious Diseases Research Program (MIDRP) y el
NIH Training grant 2D43TW007393.-->

```{r, eval=TRUE, echo=FALSE}
a <- c(50,200,#25,
       60*150*3.29,# (muestras)*(precio_por_muestra)*(tipo_de_cambio_sol-dolar)
       2000
       #,25
       )
b <- format((a*100/sum(a)), digits = 1)
#sum(a)
#rbind(a,b)
#`r a[4]` do not works
# costo total proyecto PQ+SEV= S/101,990
# (200*150*3.29)+(2*500*3.29) 
```

  ---------------------------------------------------------------
  **DESCRIPCIÓN**                                  **MONTO (S/)**
  ----------------------------------------- ---------------------
  **Bienes**                             
  
  Papelería, útiles y material de oficina                   50.00
  
  Insumos e instrumental de laboratorio                    200.00
  
  **Servicios**

  Compra, sondeo y lectura                              29,610.00
  de microarreglos
  
  Gastos en el transporte de muestras                    2,000.00
  
  **TOTAL**                                             31,860.00
  ---------------------------------------------------------------

  Table: Presupuesto \label{tab:presup}


  <!-- ------------------------------------------------------------------------------------- -->
  <!-- **DESCRIPCIÓN**                            **MONTO (S/)**        **PORCENTAJE (%)**  -->
  <!-- ----------------------------------------- --------------------- --------------------- -->
  <!-- **Bienes**                              -->

  <!-- Papelería, útiles y material de oficina            50                    `r b[1]` -->

  <!-- Insumos e instrumental de laboratorio              200                   `r b[2]` -->

  <!-- **Servicios** -->

  <!-- Compra, sondeo y lectura                           29610                 `r b[3]` -->
  <!-- de microarreglos -->

  <!-- Gastos en el transporte de muestras                2000                  `r b[4]` -->

  <!-- **TOTAL**                                          31910                 100 -->
  <!-- ------------------------------------------------------------------------------------- -->

  <!-- Table: Presupuesto \label{tab:presup} -->

<!--   
  Productos químicos                                 25                    `r b[3]`
  Impresiones, encuadernación                        25                    `r b[6]`
  y empastado 
-->

# REFERENCIAS BIBLIOGRÁFICAS
<!--
1.  **Referida al tema**
    
    Andrade BB, et al. (2010). Severe Plasmodium vivax malaria exhibits
    marked inflammatory imbalance. Malar J 9: 13.
    
    Arevalo-Herrera M, et al. (2012). Malaria in selected non-Amazonian
    countries of Latin America. Acta Trop 121: 303–314.
    
    Bassat Q, et al. (2011). Defying malaria: fathoming severe
    Plasmodium vivax disease. Nat Med 17: 48–49.
    
    Genton B, et al. (2008). Plasmodium vivax and mixed infections are
    associated with severe malaria in children: a prospective cohort
    study from Papua New Guinea. PLoS Med 5: e127.
    
    Poespoprodjo JR, et al. (2009). Vivax malaria: a major cause of
    morbidity in early infancy. Clin Infect Dis 48: 1704–1712.
-->
<!--
# ANEXOS

##  Matriz de consistencia

Ver \autoref{tab:consis} al final del documento.
-->
