==> 03-sevrcov.Rmd <==
---
title: "COVARIATES (merged: nov2014.dta + sep2016.dta)"
author: "Andree Valle Campos"
date: '`r Sys.Date()`'
output:
  # html_notebook:
  html_document:
    fig_caption: yes
    toc: yes
    toc_depth: 5
    toc_float:
      collapsed: yes
    code_folding: "hide"
    #df_print: paged
bibliography: malaria.bib
biblio-style: apalike
link-citations: yes
#csl: american-medical-association.csl
editor_options: 
  chunk_output_type: console
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE, #fig.path = "01-",
                      warning = FALSE)
# knitr::opts_knit$set(root.dir = '../.')

options(width = 110) # expand limits of CONSOLE output

rm(list=ls())

require(Hmisc)
#mu <- markupSpecs$html   # markupSpecs is in Hmisc
# hidden command (<code>r mu$widescreen()</code>) to use an entire wide screen. # mu$widescreen()
#`r mu$widescreen()`
```

## Aim

Retrieve relevant variables from the *Severe Vivax Malaria study* 
to use them as covariates for the subset of samples provided to the 
*Protein microarray Ab profile study*.

## Done

- se reporta la:
    - __actualización__ de la base de datos del estudio prospectivo de 
    Malaria Severa Vivax. ([ver](#update-covariates))
        + __comparación__ de variables entre bases y 
        + __ejecución__ de un consenso entre ambas.
    - __creación__ de nuevas variables en base a los criterios OMS.
    - __clasificación__ de las muestras según estos criterios.
        - __reproducción__ de la selección de una sub-muestra con 30 casos 
        severos más representativos.
        - __estadística descriptiva__ de los criterios OMS de la muestra total ([ver](#prop-all)), 
        - __estadística inferencial__ de variables epidemiológicas relevantes
        con la muestra total ([ver](#sev-inf)).
    - __filtración__ de las observaciones (sub-muestra) empleadas en el estudio de microarreglos de proteínas.
        - __estadística descriptiva__ de los criterios OMS de la sub-muestra ([ver](#abnomic-prop)), 
        - __estadística descriptiva__ de variables epidemiológicas relevantes ([ver](#abnomic-descrip)).
- Ir a la presentación del [problema](#introduction)
- Ir al resumen de [resultados](#results)
- Ir al detalle del [análisis](#analysis).

## Help!

__PRINCIPALES__

- con respecto a los criterios OMS:
    - __`convulsión`__ no es un criterio explícito ¿cómo justificarlo? ([ver](#theory))
    - ¿por qué no tomar en cuenta __`trombocitopenia`__ 
    (con valores cuantitativos en `cont_plaquetas`, [ver](#additional-plots)) o __`bleeding`__?
- ¿hay alguna variable que me permita aplicar algún __criterio de exclusión__?
    + Presencia de __mono-infecciones__ de _P. vivax_ determinadas por PCR, y
    + Ausencia de __coinfecciones__ (leptospirosis, dengue u otra arbovirosis).
- ¿Cómo están __definidas__ `presenta_comorb` y `enf_cronica`?
    - ¿Es relevante que incluya `presenta_comorb` o `enf_crónica` 
    en la descripción de la muestra?
    - __OJO:__ Se identificó una __alta proporción de pacientes con comorbilidades__ en
    la [sub-muestra](#abnomic-descrip) de 60 pacientes con respecto a la [muestra total](#sev-inf).

__SECUNDARIAS__

- ¿Se puede asumir un valor de `episodio_previo` con la ayuda de las __covariables
asociadas a exposición__? ¿vale la pena corregirlo? ([ver](#abnomic))
- ¿Cómo validar la variable __`shock`__ ante la falta de valores de `pa_sistolica`? ([ver](#compare-who))
- ¿Qué hacer con pacientes con puntaje final de __`glasgow`__ y auscencia 
de puntaje para sus sub-componentes (motor, verbal, ojos)? ([ver](#compare-who))

## Introduction

__Problema:__

-------------------------------------------------------------
 Dataset   Pros                      Cons
---------- ------------------------- -------------------------
  nov2014   única con valores        __mayor cantidad de
            cuantitativos             NA's__
            originales,
            incluido los
            __criterios de
            laboratorio__:
            glucosa,
            hemoglobina,
            creatinina,
            bilirrubina,
            ~~trombocitopenia~~,
            parasitemia.

  sep2016   menor cantidad de NA's    difiere en algunas 
            o __mayor cantidad de     variables, incluido los
            _non-missing values___    __criterios clínicos__:
                                      edad,
                                      sexo,
                                      pa_sistolica (shock),
                                      edema, sdra (pulmonar),
                                      glasgow (coma),
                                      convulción.

-------------------------------------------------------------

Table: Pros and cons between datasets

__Pregunta:__

- ¿Las diferencias y mayor cantidad de _non-missing values_ entre 
`nov2014` y `sep2016` deben ser consideradas 
como __actualizaciones__ o __correcciones__?
    + __rpta:__ `sep2016` ha pasado por control de calidad y reingreso de fichas (Smith-Nuñez, correo 28oct2016)
    + __repregunta:__ ¿qué variables han pasado por QC?

__Solución:__

- __criterios de laboratorio__ extraídos de `nov2014`.
- variables epidemiológicas y __criterios clínicos__ actualizados en base a `sep2016`.
    + [comparar](#compare-covariates) la presencia de diferencias y 
    cantidad de NA's por variable,
    + [ejecutar consenso](#execute-consensus-covariates) y 
    otorgar prioridad a `sep2016`.

## Theory

__WHO case definition__

_Tachados están los criterios tomados en cuenta en el presente reporte_

- criterios citados en @WHO2014severe para malaria falciparum severa:
    1. ~~impaired consiousness~~
    2. ~~acidosis~~
    3. ~~hypoglicemia~~
    4. ~~severe malarial anemia~~
    5. ~~renal impairment (acute kidney injury)~~
    6. ~~jaundice~~
    7. ~~pulmonary edema~~
    8. significant bleeding
    9. ~~shock~~
    10. ~~hyperparasitemia~~

- criterios citados en @howes2016global para malaria vivax severa:
    1. condición neurológica: ~~coma (glasgow), mareo, pérdida conciencia~~;
    2. condición hematológica: ~~anemia severa~~, trombocitopenia severa, y hemoglobinuria;
    3. síntomas sistémicos: ~~shock circulatorio~~; y
    4. fallo de órganos vitales:
    ~~disfución respiratoria, estrés respiratorio agudo, 
    daño renal agudo~~, ruptura esplénica, 
    ~~disfunción hepática e ictericia (hiperbilirrubina)~~.

- criterios empleados en @smith2013 y @quispe2014 para malaria vivax severa:
    1. ~~shock~~,
    2. ~~daño pulmonar~~,
    3. ~~glasgow~~,
    4. ~~hiperbilirrubina~~,
    5. ~~anemia severa~~,
    6. ~~daño renal~~,
    7. ~~hipoglicemia~~.

## Results

__LUEGO DE LA ACTUALIZACIÓN__

- __wrt criterios OMS: GENERAL__
    + __incrementa el N (non-missing values)__ por criterio clínico (i.e., disminuyen NA's)
    + el orden y proporciones [obtenidas](#prop-all) son consistentes 
    con lo publicado en __poster ASTMH__ por @smith2013
    + las únicas variables __terminadas en `_corr`__ en empeladas fueron:
        - glasgow (corrige 2 datos)
        - convulsión (no corrige, pero agrega datos a NA's)
        - edema (no corrige, pero agrega datos a NA's)

- __wrt criterios OMS: ESPECÍFICO__
    - `postración` es retirado como criterio por ser una variable subjetiva, 
    carente de una definición clara en la literatura.
    - `coma` es retirada ya que es una variable definida por `glasgow`.
    - `shock`: solo 8 de 38 casos están definidos por `pa_sistolica`. 
    Sus datos se complementan en `shock_oms`.
    - `tiempo_local` [demuestra](#eval-cond) que la base `sep2016` corrige 
    valores erróneamente asignados en `nov2014`.

- __wrt variables epidemiológicas:__
    + mismas conclusiones estadísticas
    + incrementa el N (non-missing values) por variable epidemiológica

- __`nov2014` vs (`nov2014`+`sep2016`):__
    + no hay cambios en el orden de las proporciones por criterio
    + De la muestra total (n=224 a n=228):
        - casos severos: 14+
        - shock: 14+
        - pulmones: 3+
        - convulsión: 2+
        - con un criterio: 11+
        - con dos criterios: 1+
        - con tres criterios: 2+
    + De la sub-muestra del estudio de microarreglo de proteínas (n=60):
        - casos severos: 6+
        - shock: 4+
        - pulmonar: 2+
        - con un solo criterio: 6+

- __muestra total vs sub-muestra__
    - al igual que la muestra total, el orden de las proporciones 
    por criterio se mantienen. ([ver](#abnomic-prop))
        - en comparación con una selección de 37 casos severos representativos 
        ([ver](#reproducible-subsetting)), 
        la sub-muestra de pacientes definidos como severos (n=23):
            - __coincide__ en 10 pacientes: #5 (n=1), #3 (n=1), #2 (n=4), #1 (n=4).
            - __excluye__ a 08 pacientes con dos o más criterios, y
            - __excluye__ a 19 pacientes con un criterio y [condiciones propuestas](#reproducible-subsetting).
            - [ver](#reproducible-subsetting-comparison).
    - a diferencia de la muestra total, la variables epidemiológicas 
    de la [sub-muestra](#abnomic-descrip) poseen:
        - menor diferencia en la proporción de mujeres severas
        - menor diferencia en la proporción de casos severos con episodio previo
        - PROBLEMA: alta proporción de pacientes con comorbilidades (?)

## Analysis

- __¡ver en la tabla de contenidos de la izquierda!:__
    - __actualización__ de la base de datos del estudio prospectivo de 
    Malaria Severa Vivax.
        + __comparación__ de variables entre bases y 
        + __ejecución__ de un consenso entre ambas.
    - __creación__ de nuevas variables en base a los criterios OMS.
    - __clasificación__ de las muestras según estos criterios.
        - __reproducción__ de la selección de una sub-muestra con 30 casos 
        severos más representativos.
        - __estadística descriptiva__ de los criterios OMS de la muestra total, 
        - __estadística inferencial__ de variables epidemiológicas relevantes 
        con la muestra total.
    - __filtración__ de las observaciones (sub-muestra) empleadas en el estudio de microarreglos de proteínas.
        - __estadística descriptiva__ de los criterios OMS de la sub-muestra, 
        - __estadística descriptiva__ de variables epidemiológicas relevantes.

<!--### Dependencies -->
```{r, message=FALSE}
##essential
library("tidyverse")  # load dplyr, tidyr, tibble, readr, purr, ggplot2
library("haven")      # import STATA .dta files
library("Hmisc")      # Data dictionary + descriptive / stratified statistics
##extra
library("forcats")    # factor vectors
library("stringr")    # for strings
library("knitr")      # To display nice tables
library("Rmisc")      # multiplot ggplots
library("viridis")    # color visualization
library("compareGroups")
library("naniar")
```

### Read datasets

- Todas las covariables por base están listadas en el [anexo](#glimpse-of-all-covariates)

#### Input

- __`nov2014`__
    - Read `malaria severa 28Nov2014_Ed.dta` file. Assigned to __`sevdb_pre`__: 

```{r}
sevdb_pre <- haven::read_dta("data-raw/malaria severa 28Nov2014_Ed.dta") #%>% haven::as_factor() #ERROR
```

- __`sep2016`__
    - Read `severe_malaria_vivax SEP2016.dta` file. Assigned to __`sevdb_pos`__: 

```{r}
sevdb_pos <- haven::read_dta("data-raw/severe_malaria_vivax SEP2016.dta") #%>% haven::as_factor() #ERROR
```

#### Tidy data

- por cada base de datos se:
    + __selecciona__ variables relevantes
    + __separa__ la variable `parterial` en dos nuevas variables: `pa_sistolica` y `pa_diastolica`
    + __renombra__ a 
        - la variable que alude al tiempo viviendo en zona endémica
        - mismas variables con distinto nombre en ambas bases
        - variables terminadas en `_corr`
    + __reemplaza__ el inicio de los códigos y valores 98/99/NaN por NA's
    + __redefine__ variables como numéricas

##### sevdb_pre

```{r}
sevdb_pre_1 <- sevdb_pre %>% 
  dplyr::select(codigo,
         cas_con_admin, temp_local,
         edad, sexo,
         admin_uci:peso,
         -numconvul, -dias_uci, #-starts_with("glas_"),
         -hiperpirexia, -temperatura, -fcardiaca, frespiratoria,
         malaria_pasado_num, #malaria_pasado_num:tx_malaria,
         presenta_comorb:enf_otra_esp, -ends_with("_esp"), -contains("otra"),enf_otra,
         parasitemia, 
         glucosa:casecontrol
         ) %>% 
  tidyr::separate(parterial, c("pa_sistolica", "pa_diastolica")) %>% 
  dplyr::rename(tiempo_local= temp_local) %>% 
  mutate(codigo= str_replace(codigo,"LIM0","LIM"),
         pa_sistolica= str_replace(pa_sistolica,"NaN|99",replacement = NA_character_),
         pa_diastolica= str_replace(pa_diastolica,"NaN|99",replacement = NA_character_),
         shock= str_replace(shock,"NaN|99",replacement = NA_character_),
         sdra= str_replace(sdra,"NaN|99",replacement = NA_character_),
         postracion= str_replace(postracion,"NaN|99",replacement = NA_character_),
         confusion= str_replace(confusion,"NaN|99",replacement = NA_character_),
         coma= str_replace(coma,"NaN|99",replacement = NA_character_),
         convulsion= str_replace(convulsion,"NaN|99",replacement = NA_character_),
         snc_normal= str_replace(snc_normal,"NaN|99",replacement = NA_character_),
         disnea= str_replace(disnea,"NaN|99",replacement = NA_character_),
         resp_normal= str_replace(resp_normal,"NaN|99",replacement = NA_character_),
         glic_normal= str_replace(glic_normal,"NaN|99",replacement = NA_character_),
         hemoglobinuria= str_replace(hemoglobinuria,"NaN|99",replacement = NA_character_),
         anemia_severa= str_replace(anemia_severa,"NaN|99",replacement = NA_character_),
         renal_normal= str_replace(renal_normal,"NaN|99",replacement = NA_character_),
         hiperbil= str_replace(hiperbil,"NaN|99",replacement = NA_character_),
         hepato_normal= str_replace(hepato_normal,"NaN|99",replacement = NA_character_),
         ictericia= str_replace(ictericia,"NaN|99",replacement = NA_character_),
         malaria_pasado_num=  str_replace(malaria_pasado_num,"98|99",replacement = NA_character_),
         #fiebre_hoy_corr=  str_replace(fiebre_hoy_corr,"99",replacement = NA_character_),
         presenta_comorb= str_replace(presenta_comorb,"NaN|99",replacement = NA_character_), 
         sin_enf_cronica= str_replace(sin_enf_cronica,"NaN|99",replacement = NA_character_), 
         diabetes= str_replace(diabetes,"NaN|99",replacement = NA_character_), 
         enf_cardiac= str_replace(enf_cardiac,"NaN|99",replacement = NA_character_), 
         enf_pulm= str_replace(enf_pulm,"NaN|99",replacement = NA_character_), 
         enf_renal= str_replace(enf_renal,"NaN|99",replacement = NA_character_), 
         enf_snc= str_replace(enf_snc,"NaN|99",replacement = NA_character_),
         gingivorragia= str_replace(gingivorragia,"NaN|99",replacement = NA_character_), 
         epistaxis= str_replace(epistaxis,"NaN|99",replacement = NA_character_),
         coag_normal= str_replace(coag_normal,"NaN|99",replacement = NA_character_),
         sang_intest= str_replace(sang_intest,"NaN|99",replacement = NA_character_),
         cas_con_admin=  str_replace(cas_con_admin,"99",replacement = NA_character_)
         ) %>% 
  mutate(shock= as.double(shock),
         postracion=as.numeric(postracion), 
         confusion=as.numeric(confusion), 
         coma=as.numeric(coma), 
         convulsion=as.numeric(convulsion), 
         #snc_normal=as.numeric(snc_normal),
         sdra= as.double(sdra),
         disnea=as.numeric(disnea), 
         taquipnea=as.numeric(taquipnea), 
         infiltracion=as.numeric(infiltracion), 
         edema_pulm=as.numeric(edema_pulm), 
         #resp_normal=as.numeric(resp_normal),
         hiperbil= as.double(hiperbil),
         malaria_pasado_num= as.numeric(malaria_pasado_num),
         pa_sistolica= as.numeric(pa_sistolica),
         pa_diastolica= as.numeric(pa_diastolica),
         #fiebre_hoy_corr= as.numeric(fiebre_hoy_corr),
         presenta_comorb= as.numeric(presenta_comorb),
         sin_enf_cronica= as.numeric(sin_enf_cronica),
         diabetes=as.numeric(diabetes),
         enf_snc=as.numeric(enf_snc),
         enf_pulm=as.numeric(enf_pulm),
         enf_cardiac=as.numeric(enf_cardiac),
         enf_renal=as.numeric(enf_renal),
         gingivorragia= as.numeric(gingivorragia),
         epistaxis= as.numeric(epistaxis),
         sang_intest= as.numeric(sang_intest),
         tiempo_local= as.numeric(tiempo_local),
         cas_con_admin= as.numeric(cas_con_admin)
         ) #%>% glimpse()
```

##### sevdb_pos

```{r}
sevdb_pos_1 <- sevdb_pos %>% 
  dplyr::select(codigo,
         cas_con_admin2, tiempoenlocal,
         edad, sexo,
         admin_uci_corr:peso,
         -num_convulsion, -dias_uci, #-starts_with("glas_"),
         -hiperpirexia, -temperatura, -fcardiaca, frespiratoria,
         malaria_pasado_num, #malaria_pasado_num:tx_malaria,
         presenta_comorb:enf_otra_esp,tx_par_intest, -ends_with("_esp"), -contains("otra"),enf_otra#,
         #parasitemia, 
         #glucosa:casecontrol
         ) %>% #glimpse()
  tidyr::separate(parterial, c("pa_sistolica", "pa_diastolica")) %>% #glimpse()
  dplyr::rename(cas_con_admin= cas_con_admin2,
                tiempo_local= tiempoenlocal,
                admin_uci= admin_uci_corr,
                postracion= postracion_corr,
                confusion= confusion_corr,
                coma= coma_corr,
                glasgow= glas_corr,
                convulsion= convulsion_corr,
                disnea= disnea_corr,
                taquipnea= taquipnea_corr,
                infiltracion= infiltracion_corr,
                edema_pulm= edema_pulm_corr,
                ictericia= ictericia_corr,
                gingivorragia= gingivorragia_corr,
                epistaxis= epistaxis_corr,
                sang_intest= sang_intest_corr,
                hemoglobinuria= hemoglobinuria_corr,
                oliguria= oliguria_corr,
                sangrado= bleeding) %>% 
  mutate(codigo= str_replace(codigo,"LIM0","LIM"),
         pa_sistolica= str_replace(pa_sistolica,"NaN|99",replacement = NA_character_),
         pa_diastolica= str_replace(pa_diastolica,"NaN|99",replacement = NA_character_),
         shock= str_replace(shock,"NaN|99",replacement = NA_character_),
         sdra= str_replace(sdra,"NaN|99",replacement = NA_character_),
         postracion= str_replace(postracion,"NaN|99",replacement = NA_character_),
         confusion= str_replace(confusion,"NaN|99",replacement = NA_character_),
         coma= str_replace(coma,"NaN|99",replacement = NA_character_),
         convulsion= str_replace(convulsion,"NaN|99",replacement = NA_character_),
         #snc_normal= str_replace(snc_normal,"NaN|99",replacement = NA_character_),
         disnea= str_replace(disnea,"NaN|99",replacement = NA_character_),
         #resp_normal= str_replace(resp_normal,"NaN|99",replacement = NA_character_),
         glic_normal= str_replace(glic_normal,"NaN|99",replacement = NA_character_),
         hemoglobinuria= str_replace(hemoglobinuria,"NaN|99",replacement = NA_character_),
         anemia_severa= str_replace(anemia_severa,"NaN|99",replacement = NA_character_),
         #renal_normal= str_replace(renal_normal,"NaN|99",replacement = NA_character_),
         hiperbil= str_replace(hiperbil,"NaN|99",replacement = NA_character_),
         #hepato_normal= str_replace(hepato_normal,"NaN|99",replacement = NA_character_),
         ictericia= str_replace(ictericia,"NaN|99",replacement = NA_character_),
         malaria_pasado_num=  str_replace(malaria_pasado_num,"98|99",replacement = NA_character_),
         #fiebre_hoy=  str_replace(fiebre_hoy,"99",replacement = NA_character_),
         presenta_comorb= str_replace(presenta_comorb,"NaN|99",replacement = NA_character_), 
         #sin_enf_cronica= str_replace(sin_enf_cronica,"NaN|99",replacement = NA_character_), 
         diabetes= str_replace(diabetes,"NaN|99",replacement = NA_character_), 
         enf_cardiac= str_replace(enf_cardiac,"NaN|99",replacement = NA_character_), 
         enf_pulm= str_replace(enf_pulm,"NaN|99",replacement = NA_character_), 
         enf_renal= str_replace(enf_renal,"NaN|99",replacement = NA_character_), 
         enf_snc= str_replace(enf_snc,"NaN|99",replacement = NA_character_),
         gingivorragia= str_replace(gingivorragia,"NaN|99",replacement = NA_character_), 
         epistaxis= str_replace(epistaxis,"NaN|99",replacement = NA_character_),
         sang_intest= str_replace(sang_intest,"NaN|99",replacement = NA_character_),
         #coag_normal= str_replace(coag_normal,"NaN|99",replacement = NA_character_),
         cas_con_admin=  str_replace(cas_con_admin,"99",replacement = NA_character_)
         ) %>% #glimpse()
  mutate(shock= as.double(shock),
         postracion=as.numeric(postracion), 
         confusion=as.numeric(confusion), 
         coma=as.numeric(coma), 
         convulsion=as.numeric(convulsion), 
         #snc_normal=as.numeric(snc_normal),
         sdra= as.double(sdra),
         disnea=as.numeric(disnea), 
         taquipnea=as.numeric(taquipnea), 
         infiltracion=as.numeric(infiltracion), 
         edema_pulm=as.numeric(edema_pulm), 
         #resp_normal=as.numeric(resp_normal),
         hiperbil= as.double(hiperbil),
         malaria_pasado_num= as.numeric(malaria_pasado_num),
         pa_sistolica= as.numeric(pa_sistolica),
         pa_diastolica= as.numeric(pa_diastolica),
         #fiebre_hoy= as.numeric(fiebre_hoy),
         presenta_comorb= as.numeric(presenta_comorb),
         #sin_enf_cronica= as.numeric(sin_enf_cronica),
         diabetes=as.numeric(diabetes),
         enf_snc=as.numeric(enf_snc),
         enf_pulm=as.numeric(enf_pulm),
         enf_cardiac=as.numeric(enf_cardiac),
         enf_renal=as.numeric(enf_renal),
         gingivorragia= as.numeric(gingivorragia),
         epistaxis= as.numeric(epistaxis),
         sang_intest= as.numeric(sang_intest),
         cas_con_admin= as.numeric(cas_con_admin),
         cont_plaquetas= as.numeric(cont_plaquetas),
         talla= as.numeric(talla),
         peso= as.numeric(peso),
         tiempo_local= as.numeric(tiempo_local),
         bilirrubina= as.numeric(bilirrubina)
         ) #%>% glimpse()
```

##### sample comparison

- ambas bases de datos comparten __222__ muestras

```{r, fig.align='center', fig.width=5, fig.height=5}
g1_id <- sevdb_pre %>% dplyr::select(codigo)      # 960
g2_id <- sevdb_pos %>% dplyr::select(codigo)     # 963

g1_idl= unlist(as.list(g1_id))
g2_idl= unlist(as.list(g2_id))

tmp <- gplots::venn(list(sevdb_pos=g2_idl, sevdb_pre=g1_idl))
```
```{r}
#intersect(sevdb_pre[,"codigo"], sevdb_pos[,"codigo"]) %>% dplyr::n_distinct()
setdiff(sevdb_pre[,"codigo"], sevdb_pos[,"codigo"])
setdiff(sevdb_pos[,"codigo"], sevdb_pre[,"codigo"])
```

##### full join

- se ejecuta:
    - __unión total__ de las observaciones y variables
    - dimensión final de 228 obs x 124 var
    - variables específicas a __`sevdb_pre` terminan en `_1`__
    - variables específicas a __`sevdb_pos` terminan en `_2`__

```{r}
sevdb_1 <- full_join(sevdb_pre_1, sevdb_pos_1, by="codigo", suffix= c("_1","_2")) #%>% glimpse()
#sevdb_1 %>% dplyr::select(codigo, temp_local, tiempoenlocal)
```
### Update covariates

#### Compare covariates

* Se identifican:
    - las __diferencias__ entre las variables __pre y pos actualización__
    - el número de elementos __no-NA's (non-missing values)__

* Evaluación preliminar a la creación de nuevas [variables consenso](#show-consensus-covariates) para:
    - __llenar__ los vacios en `nov2014`, y
    - __corregir__ la data en `nov2014` por la presente en `sep2016`.

* Ver tabla:

-------------------------------------------------------------------
variable              ¿hay datos             ¿`sep2016` tiene menos
                      diferentes entre       NA's con respecto a
                      `nov2014` y `sep2016`? `nov2014`?
--------------------- ---------------------- ----------------------
__general__

`edad`                x                      x

`sexo`                x                      x

__who criteria__

`shock`               x                      x

`pa_sistolica`                               x

`glasgow`             x                      x

`glas_ojos`                                  x

`glas_verbal`                                x

`glas_motor`                                 x

~~`coma`~~            x                      x

`convulsion`                                 x

`edema_pulm`                                 x

`sdra`                x                      x

__conditional__

`tiempo_local`        x                      x

`malaria_pasado_num`                         x

`presenta_comorb`     x                      x

`diabetes`                                   x

`enf_renal`           x                      x

`enf_cardiac`         x                      x

`enf_pulm`                                   x

`enf_snc`                                    x

-------------------------------------------------------------------

Table: Comparación entre bases de datos. (La respuesta afirmativa esta marcada con un aspa)

##### general

- **EDAD**
    + `edad_1` y `edad_2` tienen elementos diferentes
    + `edad_2` posee menos NA's

```{r}
sevdb_1 %>% 
  mutate(edad= edad_1==edad_2) %>% dplyr::count(edad) #%>% View()
sevdb_1 %>% 
  mutate(edad= edad_1==edad_2) %>%
  dplyr::select(edad, edad_1, edad_2) %>% summarise_all(funs(sum(!is.na(.))))
```

- **SEXO**
    + `sexo_1` y `sexo_2` tienen elementos diferentes
    + `sexo_2` posee menos NA's
    
```{r}
sevdb_1 %>% 
  mutate(sexo= sexo_1==sexo_2) %>% dplyr::count(sexo) #%>% View()
sevdb_1 %>% 
  mutate(sexo= sexo_1==sexo_2) %>%
  dplyr::select(sexo, sexo_1, sexo_2) %>% summarise_all(funs(sum(!is.na(.))))
```

##### who criteria {#compare-who}

- **SHOCK**
    + `shock_1` y `shock_2` tienen elementos diferentes
    + `shock_1` posee menos NA's
    + `pa_sistolica_1` y `pa_sistolica_2` no tienen elementos diferentes
    + `pa_sistolica_2` posee menos NA's
    + `shock` y `pa_sistolica` se complementan.

```{r}
sevdb_1 %>% 
  mutate(shock= shock_1==shock_2) %>% dplyr::count(shock) #%>% View()
sevdb_1 %>% 
  mutate(shock= shock_1==shock_2) %>%
  dplyr::select(shock, shock_1, shock_2) %>% summarise_all(funs(sum(!is.na(.))))
sevdb_1 %>% 
  mutate(pa_sistolica= pa_sistolica_1==pa_sistolica_2) %>% dplyr::count(pa_sistolica) #%>% View()
sevdb_1 %>% 
  mutate(pa_sistolica= pa_sistolica_1==pa_sistolica_2) %>%
  dplyr::select(pa_sistolica, pa_sistolica_1, pa_sistolica_2) %>% summarise_all(funs(sum(!is.na(.))))

sevdb_1 %>% 
  mutate(shock= shock_1==shock_2) %>%
  mutate(pa_sistolica= pa_sistolica_1==pa_sistolica_2) %>%
  group_by(shock, shock_1, shock_2,
           pa_sistolica, pa_sistolica_1, pa_sistolica_2) %>% 
  dplyr::count()
```

- **GLASGOW**
    + `glasgow_1` y `glasgow_2` tienen elementos diferentes
    + `glasgow_2` menos NA's
    + ningún sub-componente `_1` y `_2` tiene elementos diferentes
    + `glas_ojos_2`, `glas_verbal_2`, `glas_motor_2` tienen menos NA's
    + ¿diferencias entre `glas_total` y `glasgow_2`?
    + `coma_1` y `coma_2` tienen elementos diferentes
    + `coma_2` menos NA's
    + `coma` es una variable __erronea__, no consistente con la data de `glasgow`
    
```{r}
sevdb_1 %>% 
  mutate(glasgow= glasgow_1==glasgow_2) %>% dplyr::count(glasgow) #%>% View()
sevdb_1 %>% 
  mutate(glasgow= glasgow_1==glasgow_2) %>%
  dplyr::select(glasgow, glasgow_1, glasgow_2) %>% summarise_all(funs(sum(!is.na(.))))
sevdb_1 %>% 
  mutate(glasgow= glasgow_1==glasgow_2) %>%
  group_by(glasgow,
           glasgow_1, glas_verbal_1, glas_ojos_1, glas_motor_1,
           glasgow_2, glas_verbal_2, glas_ojos_2, glas_motor_2) %>% dplyr::count() %>% #View()
  dplyr::arrange(glasgow_2)

sevdb_1 %>% 
  mutate(glas_ojos= glas_ojos_1==glas_ojos_2) %>% dplyr::count(glas_ojos) #%>% View()
sevdb_1 %>% 
  mutate(glas_ojos= glas_ojos_1==glas_ojos_2) %>%
  dplyr::select(glas_ojos, glas_ojos_1, glas_ojos_2) %>% summarise_all(funs(sum(!is.na(.))))

sevdb_1 %>% 
  mutate(glas_verbal= glas_verbal_1==glas_verbal_2) %>% dplyr::count(glas_verbal) #%>% View()
sevdb_1 %>% 
  mutate(glas_verbal= glas_verbal_1==glas_verbal_2) %>%
  dplyr::select(glas_verbal, glas_verbal_1, glas_verbal_2) %>% summarise_all(funs(sum(!is.na(.))))

sevdb_1 %>% 
  mutate(glas_motor= glas_motor_1==glas_motor_2) %>% dplyr::count(glas_motor) #%>% View()
sevdb_1 %>% 
  mutate(glas_motor= glas_motor_1==glas_motor_2) %>%
  dplyr::select(glas_motor, glas_motor_1, glas_motor_2) %>% summarise_all(funs(sum(!is.na(.))))

sevdb_1 %>% 
  mutate(coma= coma_1==coma_2) %>% dplyr::count(coma) #%>% View()
sevdb_1 %>% 
  mutate(coma= coma_1==coma_2) %>%
  dplyr::select(coma, coma_1, coma_2) %>% summarise_all(funs(sum(!is.na(.))))

sevdb_1 %>% 
  mutate(glasgow= glasgow_1==glasgow_2) %>%
  mutate(coma= coma_1==coma_2) %>%
  group_by(glasgow,
           glasgow_1, #glas_verbal_1, glas_ojos_1, glas_motor_1,
           glasgow_2, #glas_verbal_2, glas_ojos_2, glas_motor_2,
           coma, 
           coma_1, coma_2) %>% dplyr::count() %>% #View()
  dplyr::arrange(glasgow_2)

```

- **CONVULSION**
    + `convulsion_1` y `convulsion_2` no tienen diferencias
    + `convulsion_2` tienen menos NA's
    
```{r}
#convulsion
sevdb_1 %>% 
  mutate(convulsion= convulsion_1==convulsion_2) %>% dplyr::count(convulsion) #%>% View()
sevdb_1 %>% 
  mutate(convulsion= convulsion_1==convulsion_2) %>%
  dplyr::select(convulsion, convulsion_1, convulsion_2) %>% summarise_all(funs(sum(!is.na(.))))
```

- **LUNG_INJURY**
    + `edema_pulm_1` y `edema_pulm_2` no tienen elementos diferentes
    + `edema_pulm_2` con menos NA's
    + `sdra_1` y `sdra_2` tienen elementos diferentes
    + `sdra_2` con menos NA's
    
```{r}
#edema_pulm y sdra
sevdb_1 %>% 
  mutate(edema_pulm= edema_pulm_1==edema_pulm_2) %>% dplyr::count(edema_pulm) #%>% View()
sevdb_1 %>% 
  mutate(edema_pulm= edema_pulm_1==edema_pulm_2) %>%
  dplyr::select(edema_pulm, edema_pulm_1, edema_pulm_2) %>% summarise_all(funs(sum(!is.na(.))))

sevdb_1 %>% 
  mutate(sdra= sdra_1==sdra_2) %>% dplyr::count(sdra) #%>% View()
sevdb_1 %>% 
  mutate(sdra= sdra_1==sdra_2) %>%
  dplyr::select(sdra, sdra_1, sdra_2) %>% summarise_all(funs(sum(!is.na(.))))
```

```{r, eval=FALSE}
#- **BLEEDING**
#    + `sangrado_2`, `gingivorragia_2`, `epistaxis_2`, `sang_intest_2` con correcciones y menos NA's

sevdb_1 %>% 
  mutate(sangrado= sangrado_1==sangrado_2) %>% dplyr::count(sangrado) #%>% View()
sevdb_1 %>% 
  mutate(sangrado= sangrado_1==sangrado_2) %>%
  dplyr::select(sangrado, sangrado_1, sangrado_2) %>% summarise_all(funs(sum(!is.na(.))))

sevdb_1 %>% 
  mutate(gingivorragia= gingivorragia_1==gingivorragia_2) %>% dplyr::count(gingivorragia) #%>% View()
sevdb_1 %>% 
  mutate(gingivorragia= gingivorragia_1==gingivorragia_2) %>%
  dplyr::select(gingivorragia, gingivorragia_1, gingivorragia_2) %>% summarise_all(funs(sum(!is.na(.))))

sevdb_1 %>% 
  mutate(epistaxis= epistaxis_1==epistaxis_2) %>% dplyr::count(epistaxis) #%>% View()
sevdb_1 %>% 
  mutate(epistaxis= epistaxis_1==epistaxis_2) %>%
  dplyr::select(epistaxis, epistaxis_1, epistaxis_2) %>% summarise_all(funs(sum(!is.na(.))))

sevdb_1 %>% 
  mutate(sang_intest= sang_intest_1==sang_intest_2) %>% dplyr::count(sang_intest) #%>% View()
sevdb_1 %>% 
  mutate(sang_intest= sang_intest_1==sang_intest_2) %>%
  dplyr::select(sang_intest, sang_intest_1, sang_intest_2) %>% summarise_all(funs(sum(!is.na(.))))

#- **TROMBOCITOPENIA**
#    + `trombocitopenia_2` y `cont_plaquetas_2` con correcciones
#    + `trombocitopenia_2` con menos NA's
#    + __OJO:__ `cont_plaquetas_1` con menos NA's

sevdb_1 %>% 
  mutate(trombocitopenia= trombocitopenia_1==trombocitopenia_2) %>% dplyr::count(trombocitopenia) #%>% View()
sevdb_1 %>% 
  mutate(trombocitopenia= trombocitopenia_1==trombocitopenia_2) %>%
  dplyr::select(trombocitopenia, trombocitopenia_1, trombocitopenia_2) %>% summarise_all(funs(sum(!is.na(.))))

sevdb_1 %>% 
  mutate(cont_plaquetas= cont_plaquetas_1==cont_plaquetas_2) %>% dplyr::count(cont_plaquetas) #%>% View()
sevdb_1 %>% 
  mutate(cont_plaquetas= cont_plaquetas_1==cont_plaquetas_2) %>%
  dplyr::select(cont_plaquetas, cont_plaquetas_1, cont_plaquetas_2) %>% summarise_all(funs(sum(!is.na(.))))
```


##### conditional {#eval-cond}

- **TIEMPO (¿AÑOS VIVIENDO?) EN LOCALIDAD**
    + `tiempo_local_1` y `tiempo_local_2` tienen elementos diferentes
    + `tiempo_local_2` con menor número de NA's
```{r, fig.width=8, fig.height=4}
sevdb_1 %>% 
  mutate(tiempo_local= tiempo_local_1==tiempo_local_2) %>% dplyr::count(tiempo_local) #%>% View()
sevdb_1 %>% 
  mutate(tiempo_local= tiempo_local_1==tiempo_local_2) %>%
  dplyr::select(tiempo_local, tiempo_local_1, tiempo_local_2) %>% summarise_all(funs(sum(!is.na(.))))
sevdb_1 %>% 
  mutate(tiempo_local= tiempo_local_1==tiempo_local_2) %>%
  mutate(control_1= tiempo_local_1>edad_1,
         control_2= tiempo_local_2>edad_2) %>%
  group_by(control_1, control_2) %>% dplyr::count()
#group_by(control_1, edad_1, tiempo_local_1, control_2, edad_2, tiempo_local_2) %>% dplyr::count()
a <- sevdb_1 %>% 
  ggplot(aes(edad_1, tiempo_local_1)) + 
  geom_point() + 
  scale_color_viridis() +
  labs(title="edad vs años vivendo en localidad",
       subtitle= "nov2014")
b <- sevdb_1 %>% 
  ggplot(aes(edad_2, tiempo_local_2)) + 
  geom_point() + 
  scale_color_viridis() +
  labs(title="edad vs años vivendo en localidad",
       subtitle= "sep2016")
Rmisc::multiplot(a,b, cols = 2)
```

- **MALARIA_PASADO_NUM**
    + `malaria_pasado_num_1` y `malaria_pasado_num_2` no tienen elementos diferentes
    + `malaria_pasado_num_2` con menor número de NA's
```{r}
sevdb_1 %>% 
  mutate(malaria_pasado_num= malaria_pasado_num_1==malaria_pasado_num_2) %>% dplyr::count(malaria_pasado_num) #%>% View()
sevdb_1 %>% 
  mutate(malaria_pasado_num= malaria_pasado_num_1==malaria_pasado_num_2) %>%
  dplyr::select(malaria_pasado_num, malaria_pasado_num_1, malaria_pasado_num_2) %>% summarise_all(funs(sum(!is.na(.))))
```

- **COMORBILITIES** 
    + ¿`presenta_comorb` distinta a `enf_cronica`?

- **ENFERMEDAD CRONICA**
    + `enf_renal_1` y `enf_renal_2` tienen elementos diferentes
    + `enf_cardiac_1` y `enf_cardiac_2` tienen elementos diferentes
    + `diabetes_1` y `diabetes_2` no tienen elementos diferentes
    + `enf_pulm_1` y `enf_pulm_2` no tienen elementos diferentes
    + `enf_snc_1` y `enf_snc_2` no tienen elementos diferentes
    + `diabetes_2`, `enf_cardiac_2`, `enf_pulm_2`, `enf_renal_2`, `enf_snc_2` con menos NA's
    
```{r}
#presenta_comorb, diabetes, enf_cardiac, enf_pulm, enf_renal, enf_snc
sevdb_1 %>% 
  mutate(presenta_comorb= presenta_comorb_1==presenta_comorb_2) %>% dplyr::count(presenta_comorb) #%>% View()
sevdb_1 %>% 
  mutate(presenta_comorb= presenta_comorb_1==presenta_comorb_2) %>%
  dplyr::select(presenta_comorb, presenta_comorb_1, presenta_comorb_2) %>% summarise_all(funs(sum(!is.na(.))))

sevdb_1 %>% 
  mutate(diabetes= diabetes_1==diabetes_2) %>% dplyr::count(diabetes) #%>% View()
sevdb_1 %>% 
  mutate(diabetes= diabetes_1==diabetes_2) %>%
  dplyr::select(diabetes, diabetes_1, diabetes_2) %>% summarise_all(funs(sum(!is.na(.))))

sevdb_1 %>% 
  mutate(enf_snc= enf_snc_1==enf_snc_2) %>% dplyr::count(enf_snc) #%>% View()
sevdb_1 %>% 
  mutate(enf_snc= enf_snc_1==enf_snc_2) %>%
  dplyr::select(enf_snc, enf_snc_1, enf_snc_2) %>% summarise_all(funs(sum(!is.na(.))))

sevdb_1 %>% 
  mutate(enf_cardiac= enf_cardiac_1==enf_cardiac_2) %>% dplyr::count(enf_cardiac) #%>% View()
sevdb_1 %>% 
  mutate(enf_cardiac= enf_cardiac_1==enf_cardiac_2) %>%
  dplyr::select(enf_cardiac, enf_cardiac_1, enf_cardiac_2) %>% summarise_all(funs(sum(!is.na(.))))

sevdb_1 %>% 
  mutate(enf_pulm= enf_pulm_1==enf_pulm_2) %>% dplyr::count(enf_pulm) #%>% View()
sevdb_1 %>% 
  mutate(enf_pulm= enf_pulm_1==enf_pulm_2) %>%
  dplyr::select(enf_pulm, enf_pulm_1, enf_pulm_2) %>% summarise_all(funs(sum(!is.na(.))))

sevdb_1 %>% 
  mutate(enf_renal= enf_renal_1==enf_renal_2) %>% dplyr::count(enf_renal) #%>% View()
sevdb_1 %>% 
  mutate(enf_renal= enf_renal_1==enf_renal_2) %>%
  dplyr::select(enf_renal, enf_renal_1, enf_renal_2) %>% summarise_all(funs(sum(!is.na(.))))
```

#### Execute Consensus covariates

- los resultados de las funciones ejecutadas en esta sección están replicados en el [anexo](#show-consensus-covariates)

```{r}
 sevdb_1c <- sevdb_1 %>% 
 
  tidyr::unite(edad_c, edad_1, edad_2) %>% #count()
  mutate(edad_c=str_replace(edad_c,"(\\d{2})_\\1", "\\1")) %>%
  mutate(edad_c=str_replace(edad_c,"NA_(\\d+)", "\\1")) %>% 
  mutate(edad_c=str_replace(edad_c,"(\\d+)_NA", "\\1")) %>% 
  mutate(edad_c=str_replace(edad_c,"(\\d+)_(\\d+)", "\\2")) %>% #count() #preferencia a _2
  mutate(edad_c=str_replace(edad_c,"NA", replacement = NA_character_)) %>%  #%>% count()
  mutate(edad_c= as.numeric(edad_c)) %>%
  
  tidyr::unite(sexo_c, sexo_1, sexo_2) %>% #count()
  mutate(sexo_c=str_replace(sexo_c,"(\\d+)_\\1", "\\1")) %>%
  mutate(sexo_c=str_replace(sexo_c,"NA_(\\d+)", "\\1")) %>% 
  mutate(sexo_c=str_replace(sexo_c,"(\\d+)_NA", "\\1")) %>% 
  mutate(sexo_c=str_replace(sexo_c,"(\\d+)_(\\d+)", "\\2")) %>% #count() #preferencia a _2
  mutate(sexo_c=str_replace(sexo_c,"NA", replacement = NA_character_)) %>%  #%>% count()
  mutate(sexo_c= as.numeric(sexo_c)) %>%
  
  tidyr::unite(shock_c, shock_1, shock_2) %>% #count()
  mutate(shock_c=str_replace(shock_c,"(\\d+)_\\1", "\\1")) %>%
  mutate(shock_c=str_replace(shock_c,"NA_(\\d+)", "\\1")) %>% 
  mutate(shock_c=str_replace(shock_c,"(\\d+)_NA", "\\1")) %>% 
  mutate(shock_c=str_replace(shock_c,"(\\d+)_(\\d+)", "\\2")) %>% #count() #preferencia a _2
  mutate(shock_c=str_replace(shock_c,"NA", replacement = NA_character_)) %>%  #%>% count()
  mutate(shock_c= as.numeric(shock_c)) %>%
  
  tidyr::unite(pa_sistolica_c, pa_sistolica_1, pa_sistolica_2) %>% #count()
  mutate(pa_sistolica_c=str_replace(pa_sistolica_c,"(\\d+)_\\1", "\\1")) %>%
  mutate(pa_sistolica_c=str_replace(pa_sistolica_c,"NA_(\\d+)", "\\1")) %>% 
  mutate(pa_sistolica_c=str_replace(pa_sistolica_c,"(\\d+)_NA", "\\1")) %>% 
  mutate(pa_sistolica_c=str_replace(pa_sistolica_c,"(\\d+)_(\\d+)", "\\2")) %>% #count() #preferencia a _2
  mutate(pa_sistolica_c=str_replace(pa_sistolica_c,"NA", replacement = NA_character_)) %>%  #%>% count()
  mutate(pa_sistolica_c= as.numeric(pa_sistolica_c)) %>%
  
  tidyr::unite(glasgow_c, glasgow_1, glasgow_2) %>% #count()
  mutate(glasgow_c=str_replace(glasgow_c,"(\\d+)_\\1", "\\1")) %>%
  mutate(glasgow_c=str_replace(glasgow_c,"NA_(\\d+)", "\\1")) %>% 
  mutate(glasgow_c=str_replace(glasgow_c,"(\\d+)_NA", "\\1")) %>% 
  mutate(glasgow_c=str_replace(glasgow_c,"(\\d+)_(\\d+)", "\\2")) %>% #count() #preferencia a _2
  mutate(glasgow_c=str_replace(glasgow_c,"NA", replacement = NA_character_)) %>%  #%>% count()
  mutate(glasgow_c= as.numeric(glasgow_c)) %>%
  
  tidyr::unite(glas_ojos_c, glas_ojos_1, glas_ojos_2) %>% #count()
  mutate(glas_ojos_c=str_replace(glas_ojos_c,"(\\d+)_\\1", "\\1")) %>%
  mutate(glas_ojos_c=str_replace(glas_ojos_c,"NA_(\\d+)", "\\1")) %>% 
  mutate(glas_ojos_c=str_replace(glas_ojos_c,"(\\d+)_NA", "\\1")) %>% 
  mutate(glas_ojos_c=str_replace(glas_ojos_c,"(\\d+)_(\\d+)", "\\2")) %>% #count() #preferencia a _2
  mutate(glas_ojos_c=str_replace(glas_ojos_c,"NA", replacement = NA_character_)) %>%  #%>% count()
  mutate(glas_ojos_c= as.numeric(glas_ojos_c)) %>%
  
  tidyr::unite(glas_verbal_c, glas_verbal_1, glas_verbal_2) %>% #count()
  mutate(glas_verbal_c=str_replace(glas_verbal_c,"(\\d+)_\\1", "\\1")) %>%
  mutate(glas_verbal_c=str_replace(glas_verbal_c,"NA_(\\d+)", "\\1")) %>% 
  mutate(glas_verbal_c=str_replace(glas_verbal_c,"(\\d+)_NA", "\\1")) %>% 
  mutate(glas_verbal_c=str_replace(glas_verbal_c,"(\\d+)_(\\d+)", "\\2")) %>% #count() #preferencia a _2
  mutate(glas_verbal_c=str_replace(glas_verbal_c,"NA", replacement = NA_character_)) %>%  #%>% count()
  mutate(glas_verbal_c= as.numeric(glas_verbal_c)) %>%
  
  tidyr::unite(glas_motor_c, glas_motor_1, glas_motor_2) %>% #count()
  mutate(glas_motor_c=str_replace(glas_motor_c,"(\\d+)_\\1", "\\1")) %>%
  mutate(glas_motor_c=str_replace(glas_motor_c,"NA_(\\d+)", "\\1")) %>% 
  mutate(glas_motor_c=str_replace(glas_motor_c,"(\\d+)_NA", "\\1")) %>% 
  mutate(glas_motor_c=str_replace(glas_motor_c,"(\\d+)_(\\d+)", "\\2")) %>% #count() #preferencia a _2
  mutate(glas_motor_c=str_replace(glas_motor_c,"NA", replacement = NA_character_)) %>%  #%>% count()
  mutate(glas_motor_c= as.numeric(glas_motor_c)) %>%
  
  tidyr::unite(postracion_c, postracion_1, postracion_2) %>% #count()
  mutate(postracion_c=str_replace(postracion_c,"(\\d+)_\\1", "\\1")) %>%
  mutate(postracion_c=str_replace(postracion_c,"NA_(\\d+)", "\\1")) %>% 
  mutate(postracion_c=str_replace(postracion_c,"(\\d+)_NA", "\\1")) %>% 
  mutate(postracion_c=str_replace(postracion_c,"(\\d+)_(\\d+)", "\\2")) %>% #count() #preferencia a _2
  mutate(postracion_c=str_replace(postracion_c,"NA", replacement = NA_character_)) %>%  #%>% count()
  mutate(postracion_c= as.numeric(postracion_c)) %>%
  
  tidyr::unite(confusion_c, confusion_1, confusion_2) %>% #count()
  mutate(confusion_c=str_replace(confusion_c,"(\\d+)_\\1", "\\1")) %>%
  mutate(confusion_c=str_replace(confusion_c,"NA_(\\d+)", "\\1")) %>% 
  mutate(confusion_c=str_replace(confusion_c,"(\\d+)_NA", "\\1")) %>% 
  mutate(confusion_c=str_replace(confusion_c,"(\\d+)_(\\d+)", "\\2")) %>% #count() #preferencia a _2
  mutate(confusion_c=str_replace(confusion_c,"NA", replacement = NA_character_)) %>%  #%>% count()
  mutate(confusion_c= as.numeric(confusion_c)) %>%
  
  tidyr::unite(coma_c, coma_1, coma_2) %>% #count()
  mutate(coma_c=str_replace(coma_c,"(\\d+)_\\1", "\\1")) %>%
  mutate(coma_c=str_replace(coma_c,"NA_(\\d+)", "\\1")) %>% 
  mutate(coma_c=str_replace(coma_c,"(\\d+)_NA", "\\1")) %>% 
  mutate(coma_c=str_replace(coma_c,"(\\d+)_(\\d+)", "\\2")) %>% #count() #preferencia a _2
  mutate(coma_c=str_replace(coma_c,"NA", replacement = NA_character_)) %>%  #%>% count()
  mutate(coma_c= as.numeric(coma_c)) %>%
  
  tidyr::unite(convulsion_c, convulsion_1, convulsion_2) %>% #count()
  mutate(convulsion_c=str_replace(convulsion_c,"(\\d+)_\\1", "\\1")) %>%
  mutate(convulsion_c=str_replace(convulsion_c,"NA_(\\d+)", "\\1")) %>% 
  mutate(convulsion_c=str_replace(convulsion_c,"(\\d+)_NA", "\\1")) %>% 
  mutate(convulsion_c=str_replace(convulsion_c,"(\\d+)_(\\d+)", "\\2")) %>% #count() #preferencia a _2
  mutate(convulsion_c=str_replace(convulsion_c,"NA", replacement = NA_character_)) %>%  #%>% count()
  mutate(convulsion_c= as.numeric(convulsion_c)) %>%
  
  tidyr::unite(disnea_c, disnea_1, disnea_2) %>% #count()
  mutate(disnea_c=str_replace(disnea_c,"(\\d+)_\\1", "\\1")) %>%
  mutate(disnea_c=str_replace(disnea_c,"NA_(\\d+)", "\\1")) %>% 
  mutate(disnea_c=str_replace(disnea_c,"(\\d+)_NA", "\\1")) %>% 
  mutate(disnea_c=str_replace(disnea_c,"(\\d+)_(\\d+)", "\\2")) %>% #count() #preferencia a _2
  mutate(disnea_c=str_replace(disnea_c,"NA", replacement = NA_character_)) %>%  #%>% count()
  mutate(disnea_c= as.numeric(disnea_c)) %>%
  
  tidyr::unite(taquipnea_c, taquipnea_1, taquipnea_2) %>% #count()
  mutate(taquipnea_c=str_replace(taquipnea_c,"(\\d+)_\\1", "\\1")) %>%
  mutate(taquipnea_c=str_replace(taquipnea_c,"NA_(\\d+)", "\\1")) %>% 
  mutate(taquipnea_c=str_replace(taquipnea_c,"(\\d+)_NA", "\\1")) %>% 
  mutate(taquipnea_c=str_replace(taquipnea_c,"(\\d+)_(\\d+)", "\\2")) %>% #count() #preferencia a _2
  mutate(taquipnea_c=str_replace(taquipnea_c,"NA", replacement = NA_character_)) %>%  #%>% count()
  mutate(taquipnea_c= as.numeric(taquipnea_c)) %>%
  
  tidyr::unite(infiltracion_c, infiltracion_1, infiltracion_2) %>% #count()
  mutate(infiltracion_c=str_replace(infiltracion_c,"(\\d+)_\\1", "\\1")) %>%
  mutate(infiltracion_c=str_replace(infiltracion_c,"NA_(\\d+)", "\\1")) %>% 
  mutate(infiltracion_c=str_replace(infiltracion_c,"(\\d+)_NA", "\\1")) %>% 
  mutate(infiltracion_c=str_replace(infiltracion_c,"(\\d+)_(\\d+)", "\\2")) %>% #count() #preferencia a _2
  mutate(infiltracion_c=str_replace(infiltracion_c,"NA", replacement = NA_character_)) %>%  #%>% count()
  mutate(infiltracion_c= as.numeric(infiltracion_c)) %>%
  
  tidyr::unite(edema_pulm_c, edema_pulm_1, edema_pulm_2) %>% #count()
  mutate(edema_pulm_c=str_replace(edema_pulm_c,"(\\d+)_\\1", "\\1")) %>%
  mutate(edema_pulm_c=str_replace(edema_pulm_c,"NA_(\\d+)", "\\1")) %>% 
  mutate(edema_pulm_c=str_replace(edema_pulm_c,"(\\d+)_NA", "\\1")) %>% 
  mutate(edema_pulm_c=str_replace(edema_pulm_c,"(\\d+)_(\\d+)", "\\2")) %>% #count() #preferencia a _2
  mutate(edema_pulm_c=str_replace(edema_pulm_c,"NA", replacement = NA_character_)) %>%  #%>% count()
  mutate(edema_pulm_c= as.numeric(edema_pulm_c)) %>%
  
  tidyr::unite(sdra_c, sdra_1, sdra_2) %>% #count()
  mutate(sdra_c=str_replace(sdra_c,"(\\d+)_\\1", "\\1")) %>%
  mutate(sdra_c=str_replace(sdra_c,"NA_(\\d+)", "\\1")) %>% 
  mutate(sdra_c=str_replace(sdra_c,"(\\d+)_NA", "\\1")) %>% 
  mutate(sdra_c=str_replace(sdra_c,"(\\d+)_(\\d+)", "\\2")) %>% #count() #preferencia a _2
  mutate(sdra_c=str_replace(sdra_c,"NA", replacement = NA_character_)) %>%  #%>% count()
  mutate(sdra_c= as.numeric(sdra_c)) %>%
  
  tidyr::unite(anemia_severa_c, anemia_severa_1, anemia_severa_2) %>% #count()
  mutate(anemia_severa_c=str_replace(anemia_severa_c,"(\\d+)_\\1", "\\1")) %>%
  mutate(anemia_severa_c=str_replace(anemia_severa_c,"NA_(\\d+)", "\\1")) %>% 
  mutate(anemia_severa_c=str_replace(anemia_severa_c,"(\\d+)_NA", "\\1")) %>% 
  mutate(anemia_severa_c=str_replace(anemia_severa_c,"(\\d+)_(\\d+)", "\\2")) %>% #count() #preferencia a _2
  mutate(anemia_severa_c=str_replace(anemia_severa_c,"NA", replacement = NA_character_)) %>%  #%>% count()
  mutate(anemia_severa_c= as.numeric(anemia_severa_c)) %>%

# SIN EVALUACION por que bilitot incluye la mayor cantidad de data sin aproximación de decimales
# ADEMÁS, por errores en la union de dos números decimales iguales
#  tidyr::unite(bilirrubina_c, bilirrubina_1, bilirrubina_2, remove=F, sep = "_") %>% #count()
#  mutate(bilirrubina_c=str_replace(bilirrubina_c,"(\\d+)_\\1", "\\1")) %>%
#  mutate(bilirrubina_c=str_replace(bilirrubina_c,"NA_(\\d+)", "\\1")) %>% 
#  mutate(bilirrubina_c=str_replace(bilirrubina_c,"(\\d+)_NA", "\\1")) %>% 
#  mutate(bilirrubina_c=str_replace(bilirrubina_c,"(\\d+\\.\\d+)_(\\d+\\.\\d+)", "\\1")) %>% #count() #preferencia a _2
#  mutate(bilirrubina_c=str_replace(bilirrubina_c,"NA", replacement = NA_character_)) %>%  #%>% count()
#  mutate(bilirrubina_c= as.numeric(bilirrubina_c)) %>%
  
  tidyr::unite(sangrado_c, sangrado_1, sangrado_2) %>% #count()
  mutate(sangrado_c=str_replace(sangrado_c,"(\\d+)_\\1", "\\1")) %>%
  mutate(sangrado_c=str_replace(sangrado_c,"NA_(\\d+)", "\\1")) %>% 
  mutate(sangrado_c=str_replace(sangrado_c,"(\\d+)_NA", "\\1")) %>% 
  mutate(sangrado_c=str_replace(sangrado_c,"(\\d+)_(\\d+)", "\\2")) %>% #count() #preferencia a _2
  mutate(sangrado_c=str_replace(sangrado_c,"NA", replacement = NA_character_)) %>%  #%>% count()
  mutate(sangrado_c= as.numeric(sangrado_c)) %>%
  
  tidyr::unite(gingivorragia_c, gingivorragia_1, gingivorragia_2) %>% #count()
  mutate(gingivorragia_c=str_replace(gingivorragia_c,"(\\d+)_\\1", "\\1")) %>%
  mutate(gingivorragia_c=str_replace(gingivorragia_c,"NA_(\\d+)", "\\1")) %>% 
  mutate(gingivorragia_c=str_replace(gingivorragia_c,"(\\d+)_NA", "\\1")) %>% 
  mutate(gingivorragia_c=str_replace(gingivorragia_c,"(\\d+)_(\\d+)", "\\2")) %>% #count() #preferencia a _2
  mutate(gingivorragia_c=str_replace(gingivorragia_c,"NA", replacement = NA_character_)) %>%  #%>% count()
  mutate(gingivorragia_c= as.numeric(gingivorragia_c)) %>%
  
  tidyr::unite(epistaxis_c, epistaxis_1, epistaxis_2) %>% #count()
  mutate(epistaxis_c=str_replace(epistaxis_c,"(\\d+)_\\1", "\\1")) %>%
  mutate(epistaxis_c=str_replace(epistaxis_c,"NA_(\\d+)", "\\1")) %>% 
  mutate(epistaxis_c=str_replace(epistaxis_c,"(\\d+)_NA", "\\1")) %>% 
  mutate(epistaxis_c=str_replace(epistaxis_c,"(\\d+)_(\\d+)", "\\2")) %>% #count() #preferencia a _2
  mutate(epistaxis_c=str_replace(epistaxis_c,"NA", replacement = NA_character_)) %>%  #%>% count()
  mutate(epistaxis_c= as.numeric(epistaxis_c)) %>%
  
  tidyr::unite(sang_intest_c, sang_intest_1, sang_intest_2) %>% #count()
  mutate(sang_intest_c=str_replace(sang_intest_c,"(\\d+)_\\1", "\\1")) %>%
  mutate(sang_intest_c=str_replace(sang_intest_c,"NA_(\\d+)", "\\1")) %>% 
  mutate(sang_intest_c=str_replace(sang_intest_c,"(\\d+)_NA", "\\1")) %>% 
  mutate(sang_intest_c=str_replace(sang_intest_c,"(\\d+)_(\\d+)", "\\2")) %>% #count() #preferencia a _2
  mutate(sang_intest_c=str_replace(sang_intest_c,"NA", replacement = NA_character_)) %>%  #%>% count()
  mutate(sang_intest_c= as.numeric(sang_intest_c)) %>%
  
  tidyr::unite(trombocitopenia_c, trombocitopenia_1, trombocitopenia_2) %>% #count()
  mutate(trombocitopenia_c=str_replace(trombocitopenia_c,"(\\d+)_\\1", "\\1")) %>%
  mutate(trombocitopenia_c=str_replace(trombocitopenia_c,"NA_(\\d+)", "\\1")) %>% 
  mutate(trombocitopenia_c=str_replace(trombocitopenia_c,"(\\d+)_NA", "\\1")) %>% 
  mutate(trombocitopenia_c=str_replace(trombocitopenia_c,"(\\d+)_(\\d+)", "\\2")) %>% #count() #preferencia a _2
  mutate(trombocitopenia_c=str_replace(trombocitopenia_c,"NA", replacement = NA_character_)) %>%  #%>% count()
  mutate(trombocitopenia_c= as.numeric(trombocitopenia_c)) %>%
  
  tidyr::unite(cont_plaquetas_c, cont_plaquetas_1, cont_plaquetas_2) %>% #count()
  mutate(cont_plaquetas_c=str_replace(cont_plaquetas_c,"(\\d{4,})_\\1", "\\1")) %>%
  mutate(cont_plaquetas_c=str_replace(cont_plaquetas_c,"NA_(\\d{4,})", "\\1")) %>% 
  mutate(cont_plaquetas_c=str_replace(cont_plaquetas_c,"(\\d{4,})_NA", "\\1")) %>% 
  mutate(cont_plaquetas_c=str_replace(cont_plaquetas_c,"(\\d{4,})_(\\d{4,})", "\\2")) %>% #count() #preferencia a _2
  mutate(cont_plaquetas_c=str_replace(cont_plaquetas_c,"(\\d{4,})_(\\d{1,3})", "\\1")) %>% #count() #problemas en _2
  mutate(cont_plaquetas_c=str_replace(cont_plaquetas_c,"NA", replacement = NA_character_)) %>%  #%>% count()
  mutate(cont_plaquetas_c=str_replace(cont_plaquetas_c,"(.....)_\\1", "\\1")) %>% #count()
  mutate(cont_plaquetas_c= as.numeric(cont_plaquetas_c)) %>%
  
  tidyr::unite(presenta_comorb_c, presenta_comorb_1, presenta_comorb_2) %>% #count()
  mutate(presenta_comorb_c=str_replace(presenta_comorb_c,"(\\d+)_\\1", "\\1")) %>%
  mutate(presenta_comorb_c=str_replace(presenta_comorb_c,"NA_(\\d+)", "\\1")) %>% 
  mutate(presenta_comorb_c=str_replace(presenta_comorb_c,"(\\d+)_NA", "\\1")) %>% 
  mutate(presenta_comorb_c=str_replace(presenta_comorb_c,"(\\d+)_(\\d+)", "\\2")) %>% #count() #preferencia a _2
  mutate(presenta_comorb_c=str_replace(presenta_comorb_c,"NA", replacement = NA_character_)) %>%  #%>% count()
  mutate(presenta_comorb_c= as.numeric(presenta_comorb_c)) %>%
  
  tidyr::unite(diabetes_c, diabetes_1, diabetes_2) %>% #count()
  mutate(diabetes_c=str_replace(diabetes_c,"(\\d+)_\\1", "\\1")) %>%
  mutate(diabetes_c=str_replace(diabetes_c,"NA_(\\d+)", "\\1")) %>% 
  mutate(diabetes_c=str_replace(diabetes_c,"(\\d+)_NA", "\\1")) %>% 
  mutate(diabetes_c=str_replace(diabetes_c,"(\\d+)_(\\d+)", "\\2")) %>% #count() #preferencia a _2
  mutate(diabetes_c=str_replace(diabetes_c,"NA", replacement = NA_character_)) %>%  #%>% count()
  mutate(diabetes_c= as.numeric(diabetes_c)) %>%
  
  tidyr::unite(enf_snc_c, enf_snc_1, enf_snc_2) %>% #count()
  mutate(enf_snc_c=str_replace(enf_snc_c,"(\\d+)_\\1", "\\1")) %>%
  mutate(enf_snc_c=str_replace(enf_snc_c,"NA_(\\d+)", "\\1")) %>% 
  mutate(enf_snc_c=str_replace(enf_snc_c,"(\\d+)_NA", "\\1")) %>% 
  mutate(enf_snc_c=str_replace(enf_snc_c,"(\\d+)_(\\d+)", "\\2")) %>% #count() #preferencia a _2
  mutate(enf_snc_c=str_replace(enf_snc_c,"NA", replacement = NA_character_)) %>%  #%>% count()
  mutate(enf_snc_c= as.numeric(enf_snc_c)) %>%
  
  tidyr::unite(enf_cardiac_c, enf_cardiac_1, enf_cardiac_2) %>% #count()
  mutate(enf_cardiac_c=str_replace(enf_cardiac_c,"(\\d+)_\\1", "\\1")) %>%
  mutate(enf_cardiac_c=str_replace(enf_cardiac_c,"NA_(\\d+)", "\\1")) %>% 
  mutate(enf_cardiac_c=str_replace(enf_cardiac_c,"(\\d+)_NA", "\\1")) %>% 
  mutate(enf_cardiac_c=str_replace(enf_cardiac_c,"(\\d+)_(\\d+)", "\\2")) %>% #count() #preferencia a _2
  mutate(enf_cardiac_c=str_replace(enf_cardiac_c,"NA", replacement = NA_character_)) %>%  #%>% count()
  mutate(enf_cardiac_c= as.numeric(enf_cardiac_c)) %>%
  
  tidyr::unite(enf_pulm_c, enf_pulm_1, enf_pulm_2) %>% #count()
  mutate(enf_pulm_c=str_replace(enf_pulm_c,"(\\d+)_\\1", "\\1")) %>%
  mutate(enf_pulm_c=str_replace(enf_pulm_c,"NA_(\\d+)", "\\1")) %>% 
  mutate(enf_pulm_c=str_replace(enf_pulm_c,"(\\d+)_NA", "\\1")) %>% 
  mutate(enf_pulm_c=str_replace(enf_pulm_c,"(\\d+)_(\\d+)", "\\2")) %>% #count() #preferencia a _2
  mutate(enf_pulm_c=str_replace(enf_pulm_c,"NA", replacement = NA_character_)) %>%  #%>% count()
  mutate(enf_pulm_c= as.numeric(enf_pulm_c)) %>%
  
  tidyr::unite(enf_renal_c, enf_renal_1, enf_renal_2) %>% #count()
  mutate(enf_renal_c=str_replace(enf_renal_c,"(\\d+)_\\1", "\\1")) %>%
  mutate(enf_renal_c=str_replace(enf_renal_c,"NA_(\\d+)", "\\1")) %>% 
  mutate(enf_renal_c=str_replace(enf_renal_c,"(\\d+)_NA", "\\1")) %>% 
  mutate(enf_renal_c=str_replace(enf_renal_c,"(\\d+)_(\\d+)", "\\2")) %>% #count() #preferencia a _2
  mutate(enf_renal_c=str_replace(enf_renal_c,"NA", replacement = NA_character_)) %>%  #%>% count()
  mutate(enf_renal_c= as.numeric(enf_renal_c)) %>%
  
  tidyr::unite(enf_otra_c, enf_otra_1, enf_otra_2) %>% #count()
  mutate(enf_otra_c=str_replace(enf_otra_c,"(\\d+)_\\1", "\\1")) %>%
  mutate(enf_otra_c=str_replace(enf_otra_c,"NA_(\\d+)", "\\1")) %>% 
  mutate(enf_otra_c=str_replace(enf_otra_c,"(\\d+)_NA", "\\1")) %>% 
  mutate(enf_otra_c=str_replace(enf_otra_c,"(\\d+)_(\\d+)", "\\2")) %>% #count() #preferencia a _2
  mutate(enf_otra_c=str_replace(enf_otra_c,"NA", replacement = NA_character_)) %>%  #%>% count()
  mutate(enf_otra_c= as.numeric(enf_otra_c)) %>%
  
  tidyr::unite(tiempo_local_c, tiempo_local_1, tiempo_local_2, remove = F) %>% #count()
  mutate(tiempo_local_c=str_replace(tiempo_local_c,"(\\d+)_\\1", "\\1")) %>%
  mutate(tiempo_local_c=str_replace(tiempo_local_c,"NA_(\\d+)", "\\1")) %>% 
  mutate(tiempo_local_c=str_replace(tiempo_local_c,"(\\d+)_NA", "\\1")) %>% 
  mutate(tiempo_local_c=str_replace(tiempo_local_c,"(\\d+)_(\\d+)", "\\2")) %>% #count() #preferencia a _2
  mutate(tiempo_local_c=str_replace(tiempo_local_c,"NA", replacement = NA_character_)) %>%  #%>% count()
  mutate(tiempo_local_c= as.numeric(tiempo_local_c)) %>% 

  tidyr::unite(malaria_pasado_num_c, malaria_pasado_num_1, malaria_pasado_num_2) %>% #count()
  mutate(malaria_pasado_num_c=str_replace(malaria_pasado_num_c,"(\\d+)_\\1", "\\1")) %>%
  mutate(malaria_pasado_num_c=str_replace(malaria_pasado_num_c,"NA_(\\d+)", "\\1")) %>% 
  mutate(malaria_pasado_num_c=str_replace(malaria_pasado_num_c,"(\\d+)_NA", "\\1")) %>% 
  mutate(malaria_pasado_num_c=str_replace(malaria_pasado_num_c,"(\\d+)_(\\d+)", "\\2")) %>% #count() #preferencia a _2
  mutate(malaria_pasado_num_c=str_replace(malaria_pasado_num_c,"NA", replacement = NA_character_)) %>%  #%>% count()
  mutate(malaria_pasado_num_c= as.numeric(malaria_pasado_num_c)) %>%  
  
  dplyr::select(
    codigo,
    
    edad_c,
    sexo_c,
    shock_c,
    pa_sistolica_c,
    glasgow_c,
    glas_ojos_c,
    glas_verbal_c,
    glas_motor_c,
    postracion_c,
    confusion_c,
    coma_c,
    convulsion_c,
    disnea_c,
    taquipnea_c,
    infiltracion_c,
    edema_pulm_c,
    sdra_c,
    anemia_severa_c,
    #bilirrubina_c,
    sangrado_c,
    gingivorragia_c,
    epistaxis_c,
    sang_intest_c,
    trombocitopenia_c,
    cont_plaquetas_c,
    presenta_comorb_c,
    diabetes_c,
    enf_snc_c,
    enf_cardiac_c,
    enf_pulm_c,
    enf_renal_c,
    enf_otra_c,
    tiempo_local_c,
    malaria_pasado_num_c,
    
    glucosa,
    hemoglobina,
    creatinina,
    bilitot, bilind, bildirect,
    parasitemia,tx_par_intest,coag_normal,casecontrol
  ) #%>% glimpse()
```

### Create WHO criteria

- en esta sección se:
    - __renombra__ a todas las variables consenso
    - __crea__ nuevas variables en base a los criterios OMS [@WHO2014severe]
    - __revisa__ detalles y frecuencias por variable creada

```{r}
sevdb_1oms <- sevdb_1c %>% 
  dplyr::rename(
    edad= edad_c,
    sexo= sexo_c,
    shock= shock_c,
    pa_sistolica= pa_sistolica_c,
    glasgow= glasgow_c,
    glas_ojos= glas_ojos_c,
    glas_verbal= glas_verbal_c,
    glas_motor= glas_motor_c,
    postracion= postracion_c,
    confusion= confusion_c,
    coma= coma_c,
    convulsion= convulsion_c,
    disnea= disnea_c,
    taquipnea= taquipnea_c,
    infiltracion= infiltracion_c,
    edema_pulm= edema_pulm_c,
    sdra= sdra_c,
    anemia_severa= anemia_severa_c,
    #bilirrubina= bilirrubina_c,
    sangrado= sangrado_c,
    gingivorragia= gingivorragia_c,
    epistaxis= epistaxis_c,
    sang_intest= sang_intest_c,
    trombocitopenia= trombocitopenia_c,
    cont_plaquetas= cont_plaquetas_c,
    presenta_comorb= presenta_comorb_c,
    diabetes= diabetes_c,
    enf_snc= enf_snc_c,
    enf_cardiac= enf_cardiac_c,
    enf_pulm= enf_pulm_c,
    enf_renal= enf_renal_c,
    enf_otra=enf_otra_c,
    tiempo_local=tiempo_local_c,
    malaria_pasado_num= malaria_pasado_num_c
  ) %>% #glimpse()
  mutate(shock_oms_pa= if_else(pa_sistolica<=80,1,0),# OR shock==1
         #snc_sum= rowSums(.[c(10,12,13)],na.rm = T),# retirados por sesgo
         #snc_oms=  if_else(snc_sum>1,1,0),
         glas_11= if_else(glasgow<=11 & glasgow>0,1,0),
         glas_9= if_else(glasgow<=9 & glasgow>0,1,0),
         pulmonar_sum= select(.,edema_pulm,sdra) %>% rowSums(na.rm = T),
         pulmonar_oms=  if_else(pulmonar_sum>0,1,0),
         #sdra=  if_else(sdra!=1,0,1),
         hipoglic_oms= if_else(glucosa<40,1,0),
         anemia_oms= if_else(hemoglobina<7,1,0),
         renal_oms= if_else(creatinina>3,1,0),
         bilsum= select(.,bilind,bildirect) %>% rowSums(na.rm = T),
         hiperbil_oms= if_else(bilsum>2.5,1,0),
         enf_sum= select(.,diabetes,enf_snc,enf_cardiac,enf_pulm,enf_renal,enf_otra) %>% rowSums(na.rm = T),#omite enf_otra
         enf_cronica= if_else(enf_sum>0,1,0),
         sangrado_sum= select(.,gingivorragia,epistaxis,sang_intest) %>% rowSums(na.rm = T),
         sangr_oms= if_else(sangrado_sum>0,1,0)
         #hiperbil_omsX= if_else(bilitot>2.5,1,0),
         #hiperbil_test= hiperbil_omsX==hiperbil_oms,
         ) %>% #glimpse() 
    mutate(
      shock_oms= select(.,shock,shock_oms_pa) %>% rowSums(na.rm = T)
      ) %>% #%>% glimpse()# OR shock==1
  mutate(trombo_sev= if_else(cont_plaquetas<50000,1,0),
         trombo_new= if_else(cont_plaquetas<150000,1,0))
```

##### who criteria

- **SHOCK**
    + menor a 80? o __menor e igual a 80__?
    + __inclusivo__ con casos definidos __sin__ `pa_sistolica` 
    (son los únicos en __Ab profile subset__)
    + en total, __solo 8 de 38 pacientes__ con `shock_oms` tienen `pa_sistolica` definida

```{r}
#SHOCK
sevdb_1oms %>% group_by(shock_oms, shock, shock_oms_pa) %>% dplyr::count() #%>% View()

#sevdb_1 %>% dplyr::select(pa_sistolica, pa_diastolica, shock_oms) %>%  
#  #mutate(shock_oms= as.factor(shock_oms)) %>% 
#  ggplot(aes(pa_sistolica, pa_diastolica, colour=shock_oms)) + 
#  geom_vline(aes(xintercept=80))+
#  #geom_histogram(alpha=.5, position = "identity") + 
#  geom_point() + 
#  scale_color_viridis() +
#  labs(title="presion arterial")
```

- **GLASGOW**
    + `coma` no era consistente con los valores reales de `glasgow`
    + `glas_oms` corrige error.    
    + __menor a 9__ o menor a 11?
    + heterogeneidad en data: 
        - valores finales sin valores de sub-componentes
        - y viseversa
        - **requiere revisión**

```{r}
#GLASGOW -------------------------------------------------- CON INCONSISTENCIAS!!!
sevdb_1oms %>% group_by(glas_9, glas_11, glas_ojos, glas_verbal, glas_motor, glasgow) %>% dplyr::count()
sevdb_1oms %>% group_by(glas_9, glas_11, coma) %>% dplyr::count()
```


- **CONVULSION**
```{r, eval=FALSE}
# CNS malaria = postracion, coma, convulsion, confusion
sevdb_1oms %>% dplyr::count(convulsion)
```

- **LUNG_INJURY**
    + definition:
        + `edema_pulm` or `sdra`
        + at least 1 criteria
    + lung related-injuries (disnea, taquipnea, infiltracion, edema)
        + not used
```{r}
# SDRA != LUNG_INJURY = at least 2 criteria of lung related-injuries (disnea, taquipnea, infiltracion, edema)
#sevdb_1oms %>% 
#  group_by(pulmonar_oms, pulmonar_sum, sdra, edema_pulm, disnea, taquipnea, infiltracion
           #, resp_normal
#           ) %>%  
#  dplyr::count()

# RESPIRACION NORMAL vs FRECUENCIA RESPIRATORIA? -------------------------------------------------- REDEFINIR!!
#sevdb_1oms %>% group_by(resp_normal, frespiratoria) %>% dplyr::count()

# SDRA vs LUNG INJURY -------------------------------------------------- CONSISTENTE?
sevdb_1oms %>% group_by(pulmonar_oms, sdra, edema_pulm) %>%  dplyr::count()
```

- ~~**BLEEDING**~~
    + ~~gingivorragia_corr, epistaxis_corr, sang_intest_corr~~
    + ~~at least one criteria~~
    + ~~`sangrado` con asignación incorrecta~~
```{r, eval=TRUE}
# BLEEDING -------------------------------------------------- REDEFINIR!!
#  = at least one criteria of bleeding (gingivorragia_corr, epistaxis_corr, sang_intest_corr)
sevdb_1oms %>% 
  group_by(sangr_oms, sangrado_sum, sangrado, gingivorragia, epistaxis, sang_intest) %>%  
  dplyr::count()
```
- ~~**TROMBOCITOPENIA**~~
    + ~~puntos de corte~~
        - ~~severe trombocitopenia: menor de 50 000~~
        - ~~trombocitopenia: menor de 150 000~~
    + ~~`trombocitopenia` con asignacion incorrecta~~
```{r, eval=TRUE}
#TROMBOCITOPENIA?? -------------------------------------------------- REDEFINIR!!
sevdb_1oms %>%
  #dplyr::select(trombocitopenia, cont_plaquetas, coag_normal) %>% #dplyr::count()
  dplyr::count(trombo_new,trombo_sev)
sevdb_1oms %>%
  group_by(trombo_new,trombo_sev) %>%  #glimpse()
  summarise_at(.vars = c("cont_plaquetas"),.funs = c("min","max"))
  #ggplot(aes(x=cont_plaquetas, fill=trombo_sev)) + 
  #geom_histogram(alpha=.5, position = "identity") + 
  #labs(title="trombocitopenia")
```

- **HIPOGLICEMIA**
    + ningún hipoglicémico!
```{r, eval=FALSE}
# HIPOGLICEMIA
# glic_normal = hipo_glic1 + hipo_glic2 ????????
sevdb_1oms %>% group_by(hipoglic_oms, glucosa) %>%  dplyr::count()
```

- **ANEMIA SEVERA**
    + `anemia_severa` no era consistente con los valores reales de `hemoglobina`
    + `anemia_oms` corrige error.

```{r, eval=FALSE}
# ANEMIA
sevdb_1 %>% group_by(anemia_oms, anemia_severa, hemoglobinuria, hemog_bajo, hto_bajo) %>%  dplyr::count()
sevdb_1 %>% group_by(anemia_oms, anemia_severa, hemoglobina) %>%  dplyr::count()
sevdb_1 %>% dplyr::select(hto, hemoglobina,anemia_oms) %>%  
  ggplot(aes(hto, hemoglobina, colour=anemia_oms)) + 
  geom_point() + 
  geom_hline(aes(yintercept=7))+
  scale_color_viridis() +
  labs(title="hematocrito vs hemoglobina")
```

- **DAÑO RENAL**
    + `renal_normal` y `hipercreat` no era consistente con los valores reales de `creatinina`
    + `renal_oms` corrige error

```{r, eval=FALSE}
# DAÑO RENAL
sevdb_1 %>% group_by(renal_oms, renal_normal, hipercreat, oliguria) %>%  dplyr::count()
sevdb_1 %>% group_by(renal_oms, renal_normal, hipercreat, creatinina) %>%  dplyr::count()
sevdb_1 %>% dplyr::select(creatin_valor, creatinina, renal_oms) %>%  
  ggplot(aes(creatin_valor, creatinina, colour=renal_oms)) + 
  geom_point() + 
  geom_hline(aes(yintercept=3))+
  scale_color_viridis() +
  labs(title="creatin_valor vs creatinina")
```

- **HIPERBILIRRUBINA**
    + `ictericia` diferente a `hiperbili`, no son consistentes con los valores reales de `bilind` y `bildirect`.
    + `hiperbil_oms` corrige error
    
```{r}
# HIPERBILIRRUBINA
#sevdb_1 %>% group_by(hiperbil_oms, hiperbil_omsX) %>% dplyr::count()

sevdb_1oms %>% dplyr::select(hiperbil_oms) %>% dplyr::count(hiperbil_oms)

sevdb_1oms %>% 
  group_by(hiperbil_oms, #hiperbil, 
           #bilirrubina, 
           bilitot, bilind, bildirect) %>%  dplyr::count() #%>% View()

#sevdb_1oms %>% 
#  group_by(hiperbil_oms, hiperbil, ictericia, hepato_normal) %>%  
#  dplyr::count() #%>% View()
```

##### conditional

- **COMORBILITIES** 
    + `presenta_comorb` y `enf_cronica` son diferentes
    
```{r, eval=FALSE}
# COMORBIDITIES
sevdb_1oms %>% 
  group_by(presenta_comorb, enf_cronica, diabetes, enf_cardiac, enf_pulm, enf_renal, enf_snc, enf_otra) %>% 
  dplyr::count()
```

### Classify wrt WHO criteria

- en esta sección se:
    - __clasifica__ a las observaciones en base a los criterios OMS [@WHO2014severe]
        - ___OJO:_ Severo con 1 o más criterios OMS__
        - __renombra__ a:
            - `malaria_pasado` por __`episodio_previo`__.
            - `tiempo_local` por __`en_zona_endemica`__.
    - __resume:__ 
        - las proporciones por criterio
        - la frecuencia de combinaciones de criterios
    - __reproduce__ una selección de 30 muestras con malaria vivax severa más representativas
    - __realiza__ inferencia estadística con respecto a las variables epidemiológicas.

_Deben observarse lineas completas._
___Reducir el zoom__ de la página de ser necesario._

```{r}
sevdb_2 <- sevdb_1oms %>% 
  dplyr::rename(#glas_oms=glas_11
                glas_oms=glas_9
                ) %>% #glimpse()
  dplyr::select(1:3,
         anemia_oms,
         hiperbil_oms, hipoglic_oms,
         pulmonar_oms, #sdra, #snc_sum, #snc_oms
         shock_oms, 
         glas_oms, #a.k.a coma
         renal_oms,
         convulsion,
         sangr_oms,trombo_sev,#trombo_new
         #trombocitopenia, #hiperparasitemia,
         presenta_comorb, enf_cronica, enf_sum, parasitemia, shock_oms_pa,
         tiempo_local, malaria_pasado_num,tx_par_intest,casecontrol
         ) %>% #glimpse()
  mutate(sev_WHO_num= select(.,anemia_oms:convulsion) %>% rowSums(na.rm = T),#,sev_SNC_num= rowSums(.[11:13],na.rm = T)
         sev_WHO_num_1= select(.,anemia_oms:sangr_oms) %>% rowSums(na.rm = T),
         sev_WHO_num_2= select(.,anemia_oms:trombo_sev) %>% rowSums(na.rm = T),
         sev_WHO_num_3= select(.,anemia_oms:convulsion,trombo_sev) %>% rowSums(na.rm = T),
         sev_WHO_num_4= select(.,anemia_oms:trombo_sev,-convulsion) %>% rowSums(na.rm = T)#,
         ) %>% #glimpse()
  dplyr::rename(episodio_previo_num=malaria_pasado_num,
                en_zona_endemica=tiempo_local) %>% 
  mutate(#sev_ALL_num= rowSums(.[c(18,19)],na.rm = T),
         sev_WHO= if_else(sev_WHO_num>0# | sev_ALL_num>1
                          ,"severo","no_severo"),# OJO: AQUI TRADUCCION!!!!
         sev_WHO_1= if_else(sev_WHO_num_1>0,"severo","no_severo"),
         sev_WHO_2= if_else(sev_WHO_num_2>0,"severo","no_severo"),
         sev_WHO_3= if_else(sev_WHO_num_3>0,"severo","no_severo"),
         sev_WHO_4= if_else(sev_WHO_num_4>0,"severo","no_severo"),
         episodio_previo=if_else(episodio_previo_num>0,"con","sin")# OJO: AQUI TRADUCCION!!!!
         ) %>% 
  mutate(sexo= as.factor(sexo),
         sev_WHO= as.factor(sev_WHO),
         sev_WHO_1= as.factor(sev_WHO_1),
         sev_WHO_2= as.factor(sev_WHO_2),
         sev_WHO_3= as.factor(sev_WHO_3),
         sev_WHO_4= as.factor(sev_WHO_4),
         episodio_previo= as.factor(episodio_previo)
         #,casecontrol= as.factor(casecontrol)
         ) %>% 
  mutate(sexo= fct_recode(sexo, "masculino"="1", "femenino"="0")) %>% # OJO: AQUI TRADUCCION!!!!
  #relevel(sexo, ref = "female") %>% 
  mutate(sexo= fct_relevel(sexo, "masculino"),
         episodio_previo= fct_relevel(episodio_previo, "sin")) %>% # OJO: AQUI TRADUCCION!!!!
  mutate(tx_par_intest= str_replace(tx_par_intest,"NaN|99|98",replacement = NA_character_) %>% as.double()) %>% 
  glimpse()
```

```{r}
###### TEST Hmisc::upData
sevdb_2x <- Hmisc::upData(sevdb_2,
                         labels = c(edad="Age (years)",
                                    sexo="Sex",
                                    shock_oms="Shock (systolic blood pressure < 80mmHg)",
                                    glas_oms="Coma (Glasgow score < 9/14)",
                                    pulmonar_oms="Lung injury (ARDS or pulmonary edema)",
                                    anemia_oms="Severe anemia (hemoglobin < 7mg/dL)",
                                    renal_oms="Acute renal failure (creatinin > 3mg/dL)",
                                    hiperbil_oms="Hyperbilirubinemia (bilirubin > 2.5mg/dL)", #Jaundice
                                    hipoglic_oms="Hypoglicemia (glucose < 40mg/dL)",
                                    convulsion="Convulsion",
                                    presenta_comorb="Comorbidity",
                                    enf_cronica="Chronic disease",
                                    #enf_sum="",
                                    parasitemia="Parasite density (par/uL)",
                                    #shock_oms_pa="",
                                    en_zona_endemica="Time in endemic area (years)",
                                    episodio_previo_num="Previous episodes (number)",
                                    sev_WHO_num="Cumulative WHO criteria",
                                    sev_WHO="WHO criteria",
                                    episodio_previo="Previous episodes"
                         ),
                         levels = list(sev_WHO=list("Non-severe"="no_severo",
                                                    "Severe"="severo"),
                                       episodio_previo=list("Without"="sin",
                                                            "With"="con"),
                                       sexo=list("Male"="masculino",
                                                 "Female"="femenino")
                                       )
                         )
```

```{r, eval=FALSE}
###### TEST Hmisc::upData
sevdb_2x <- Hmisc::upData(sevdb_2,
                         labels = c(edad="Age",
                                    sexo="Sex",
                                    shock_oms="Shock",
                                    glas_oms="Coma",
                                    pulmonar_oms="Lung injury",
                                    anemia_oms="Severe anemia",
                                    renal_oms="Acute renal failure",
                                    hiperbil_oms="Hyperbilirubinemia", #Jaundice
                                    hipoglic_oms="Hypoglicemia",
                                    convulsion="Convulsion",
                                    presenta_comorb="Comorbidity",
                                    enf_cronica="Chronic disease",
                                    #enf_sum="",
                                    parasitemia="Parasite density",
                                    #shock_oms_pa="",
                                    en_zona_endemica="Time in endemic area",
                                    episodio_previo_num="Previous episodes",
                                    sev_WHO_num="Cumulative WHO criteria",
                                    sev_WHO="WHO criteria",
                                    episodio_previo="Previous episodes"
                         ),
                         units = c(edad="(years)",
                                   shock_oms="(systolic blood pressure < 80mmHg)",
                                   glas_oms="(Glasgow score < 9/14)",
                                   pulmonar_oms="(ARDS or pulmonary edema)",
                                   anemia_oms="(hemoglobin < 7mg/dL)",
                                   renal_oms="(creatinin > 3mg/dL)",
                                   hiperbil_oms="(bilirubin > 2.5mg/dL)",
                                   hipoglic_oms="(glucose < 40mg/dL)",
                                   parasitemia="(par/uL)",
                                   en_zona_endemica="(years)",
                                   episodio_previo_num="(number)"
                                   #episodio_previo="prop"
                         ),
                         levels = list(sev_WHO=list("Non-severe"="no_severo",
                                                    "Severe"="severo"),
                                       episodio_previo=list("Without"="sin",
                                                            "With"="con"),
                                       sexo=list("Male"="masculino",
                                                 "Female"="femenino")
                                       )
                         )
```

<!-- reduced dictionary -->
```{r}
html(Hmisc::contents(sevdb_2x), maxlevels=10, levelType='table')
#glimpse(sevdb_2)
```

#### proportions per criteria {#prop-all}

```{r}
s1 <- Hmisc::summaryM(shock_oms + #clinicos
                        pulmonar_oms +
                        glas_oms +
                        hiperbil_oms + #lab
                        anemia_oms +
                        renal_oms +
                        hipoglic_oms +
                        #postracion +#extra
                        #coma +
                        convulsion +
                        #presenta_comorb+
                        #enf_cronica
                        #parasitemia + 
                        sev_WHO_num
                      ~ sev_WHO
                      ,
               data=sevdb_2x,
               overall=FALSE, test=FALSE)

latex(s1, caption='WHO classification: Clinical and Laboratory criteria',
      exclude1=TRUE, npct='both', 
      digits=2,
      #prmsd=TRUE, brmsd=TRUE, msdsize=mu$smaller2, #NOT-EVALUATE if PDF
      middle.bold=TRUE, long = TRUE,
      #legend.bottom = TRUE, #insert.bottom = TRUE, 
      what="%", html = TRUE, width="100%"
      ) #change here for LaTeX PDF
```

#### combination of criteria

__check `sev_WHO`__

- frecuencia por criterios acumulados

```{r}
sevdb_2 %>% group_by(sev_WHO_num, sev_WHO#, sev_SNC_num, sev_ALL_num
                     ) %>% dplyr::count()
```

```{r}
# SUM SEVERITY CRITERIA
#sevdb_o <- sevdb_2 %>% 
#  dplyr::rename(glas=glas_11,
#         #sdra,
#         lung=lung_injury,
#         icter=ictericia_corr,
#         bili=hiperbil,
#         #shock,
#         bleed=bleeding,
#         glic=glic_normal,
#         paras=hiperparasitemia,
#         anemia=anemia_severa#, 
#         #hemo=hemoglobinuria_corr
#         ) 
sevdb_2 %>% 
  group_by(shock_oms, glas_oms, pulmonar_oms, #sdra,snc_oms,
           hipoglic_oms, anemia_oms, renal_oms, hiperbil_oms,
           convulsion,
           #coma, postracion,
           sev_WHO_num, sev_WHO
           ) %>% 
  dplyr::count() %>% 
  dplyr::arrange(desc(#n
                      sev_WHO_num))
```

```{r}
sevdb_2 %>% 
  dplyr::select(shock_oms, glas_oms, pulmonar_oms, #sdra,snc_oms,
         hipoglic_oms, anemia_oms, renal_oms, hiperbil_oms,
         convulsion#,
         #coma, postracion
           ) %>% 
  summarise_all(sum,na.rm=T) #%>% t()
```

__RECLASIFICACIÓN en base a criterio OMS__

- `shock_oms` principal criterio de re-clasificación
- pero __solo 2 pacientes__ de este grupo poseen `pa_sistolica` definida

```{r}
sevdb_2 %>% 
  dplyr::select(codigo, sev_WHO) %>% 
  inner_join(sevdb_pre %>% 
               dplyr::select(codigo, casecontrol) %>% 
               mutate(codigo= str_replace(codigo,"LIM0","LIM")),
             by="codigo") %>% 
  group_by(casecontrol,sev_WHO) %>% dplyr::count()

sevdb_2 %>% 
  dplyr::select(codigo, starts_with("sev_WHO"), ends_with("_oms"), convulsion) %>% 
  inner_join(sevdb_pre_1 %>% 
               dplyr::select(codigo, casecontrol
                      #, pa_sistolica # solo dos pacientes reclasificados poseen esta variable definida
                      ) %>% 
               mutate(codigo= str_replace(codigo,"LIM0","LIM")),
             by="codigo") %>% 
  filter(casecontrol=="0" & sev_WHO=="severo") %>% 
  dplyr::select(anemia_oms:convulsion) %>% summarise_all(sum,na.rm=T)

sevdb_2 %>% 
  dplyr::select(codigo, sev_WHO, shock_oms) %>% 
  inner_join(sevdb_pre_1 %>% 
               dplyr::select(codigo, casecontrol
                      , pa_sistolica # solo dos pacientes reclasificados poseen esta variable definida
                      ) %>% 
               mutate(codigo= str_replace(codigo,"LIM0","LIM")),
             by="codigo") %>% 
  #filter(casecontrol=="0" & sev_WHO=="severo") %>% 
  filter(pa_sistolica<=80)
```

##### reproducible subsetting

__Elección de 30 pacientes severos más representativos:__

- __14 con `sev_WHO_num`>1__
- Al tener 57 pacientes con un solo criterio, ¿cómo priorizar entre criterios para la virtual elección de 
pacientes con un solo criterio y completar una muestra final de 30? 
- __condiciones propuestas:__
    + auscencia de __comorbilidades__ y __enfermedades crónicas__, y
    + `shock_oms` con `pa_sistolica` definida (__+ 7 casos__), o
    + `pulmonar_oms`= `edema`+`sdra` (__+5 casos__), o
    + `hiperbil_oms` (__+ 11 casos__)

```{r}
sevdb_30 <- sevdb_2 %>% 
  dplyr::filter(sev_WHO_num>1 | sev_WHO_num>0 & enf_sum<1 & presenta_comorb!=1
                  #& hiperbil_oms>0
                  #& (hiperbil_oms>0 | shock_oms>0)
                  & (shock_oms_pa>0 | pulmonar_oms>0 | hiperbil_oms>0
                     )
                  )
sevdb_30 %>% dplyr::select(codigo, sev_WHO_num, enf_cronica, presenta_comorb
                    ,shock_oms_pa#,shock_oms
                    ,pulmonar_oms
                    ,hiperbil_oms
                    ) %>% dplyr::arrange(desc(sev_WHO_num))
```

```{r, fig.height=4, fig.width=6, fig.align='center', eval=FALSE}
### symptoms x exposure

#plot(sevdb_2$episodio_previo_num,sevdb_2$sev_WHO_num)
sevdb_2 %>% 
  group_by(episodio_previo_num, sev_WHO_num) %>%
  dplyr::summarise(n=n()) %>%
  mutate(prop= n/sum(n)) %>%
  ggplot(aes(episodio_previo_num, sev_WHO_num, colour=n)) + 
  geom_point() + 
  scale_color_viridis() +
  labs(title="sevdb_2")
```

#### Inferential statistics {#sev-inf}

* la prueba de hipótesis se ejecutará sobre las variables evaluadas en:
    - @smith2013,
    - @baldevi2013, y
    - @quispe2014

```{r, eval=FALSE}
#- Joint `parasitemia` from `28Nov2014_Ed.dta` **[WARNING]**

sevdb_x <- haven::read_dta("data-raw/malaria severa 28Nov2014_Ed.dta") %>% 
  dplyr::select(codigo, parasitemia)%>% 
  mutate(codigo= str_replace(codigo,"LIM0","LIM"))

sevdb_2x <- sevdb_2 %>% 
  haven::as_factor() %>% 
  dplyr::inner_join(sevdb_x,by="codigo")
```

***

**wrt severity**

- significant differences with respect to 
    + proportion of previous episodes of malaria,
    + proportion of women patients towards severe cases

```{r}
s1 <- Hmisc::summaryM(edad + sexo +
                        en_zona_endemica +
                        episodio_previo_num + episodio_previo +  
                        #sev_WHO + sev_WHO_num +
                        #presenta_comorb + enf_cronica +
                        parasitemia #+
                      ~ sev_WHO #+ 
                      #episodio_previo
                      ,
               data=sevdb_2x,
               overall=FALSE, test=TRUE)

latex(s1, caption='Epidemiological data',
      exclude1=TRUE, npct='both', 
      digits=2, eps = 0.01,
      #prmsd=TRUE, brmsd=TRUE, msdsize=mu$smaller2, #NOT-EVALUATE if PDF
      middle.bold=TRUE, long = FALSE,
      #legend.bottom = TRUE, #insert.bottom = TRUE, 
      what="%", html = TRUE
      ) #change here for LaTeX PDF

```

***

**wrt previouse episodes of malaria**

- significant differences with respect to 
    + mean age,
    + proportion of severe patients

```{r}
s1 <- Hmisc::summaryM(edad + sexo +
                        en_zona_endemica +
                        episodio_previo_num + #episodio_previo +  
                        sev_WHO + #sev_WHO_num +
                        presenta_comorb + enf_cronica +
                        parasitemia #+
                      ~ #sev_WHO #+ 
                      episodio_previo
                      ,
               data=sevdb_2x,
               overall=FALSE, test=TRUE)

latex(s1, caption='Epidemiological data',
      exclude1=TRUE, npct='both', 
      digits=2,
      #prmsd=TRUE, brmsd=TRUE, msdsize=mu$smaller2, #NOT-EVALUATE if PDF
      middle.bold=TRUE, long = FALSE,
      #legend.bottom = TRUE, #insert.bottom = TRUE, 
      what="%", html = TRUE
      ) #change here for LaTeX PDF
```

***

**wrt severity by previous episodes of malaria**

- no significant differences with respect to any variable

```{r}
s1 <- Hmisc::summaryM(edad + sexo +
                        en_zona_endemica +
                        #episodio_previo + #episodio_previo_num + 
                        #sev_WHO + sev_WHO_num +
                        presenta_comorb + enf_cronica +
                        parasitemia #+
                      ~ sev_WHO + 
                      episodio_previo
                      ,
               data=sevdb_2x,
               overall=FALSE, test=TRUE)

latex(s1, caption='Epidemiological data',
      exclude1=TRUE, npct='both', 
      digits=2,
      #prmsd=TRUE, brmsd=TRUE, msdsize=mu$smaller2, #NOT-EVALUATE if PDF
      middle.bold=TRUE, long = FALSE,
      #legend.bottom = TRUE, #insert.bottom = TRUE, 
      what="%", html = TRUE
      ) #change here for LaTeX PDF
```
***

```{r, eval=FALSE}
#**age wrt previous episodes among severe cases**
#- significant differences with respect to 
#    + mean age
s1 <- Hmisc::summaryM(edad + sexo +
                        en_zona_endemica +
                        episodio_previo + #episodio_previo_num + 
                        sev_WHO + sev_WHO_num +
                        presenta_comorb + enf_cronica +
                        parasitemia #+
                      ~ #sev_WHO + 
                      episodio_previo
                      ,
               data=sevdb_2 %>% filter(sev_WHO=="severo"),
               overall=FALSE, test=FALSE)

latex(s1, caption='Epidemiological data',
      exclude1=TRUE, npct='both', 
      digits=2,
      #prmsd=TRUE, brmsd=TRUE, msdsize=mu$smaller2, #NOT-EVALUATE if PDF
      middle.bold=TRUE, long = FALSE,
      #legend.bottom = TRUE, #insert.bottom = TRUE, 
      what="%", html = TRUE
      ) #change here for LaTeX PDF
```

```{r, fig.width=6, fig.height=5, fig.align='center'}
#plot(s1, which='categorical')
#Key(0,0)
```

```{r, fig.width=8, fig.height=4, fig.align='center'}
#par(mfrow=c(1,2))
#plot(s1, which='continuous')
#par(mfrow=c(1,1))
```


### Filter Ab profile samples {#abnomic}

- en esta sección se:
    - __filtra__ a las observaciones equivalentes a las 
    **60 muestras** empleadas en el estudio de microarreglos de proteínas
    - __corrige__ la data NA de `episodios_previos`
    - __resume:__ 
        - las proporciones por criterio
        - la frecuencia de combinaciones de criterios
    - __compara__ con la selección __reproducida__ de 30 muestras con malaria vivax severa más representativas.
    - __realiza__ descripción estadística de las variables epidemiológicas.

***

__FILTERING of **60 samples** from the Severe study dataset__

```{r, message=FALSE}
subsev <- readr::read_csv("data-raw/samples.csv") %>% 
  filter(Study=="Severe Malaria") %>% 
  #mutate(sample= Sample.ID==Subject.ID) %>% dplyr::count(sample)
  dplyr::select(Sample.ID, Study) %>% 
  dplyr::rename(codigo= Sample.ID)

# generate
sevdb_3 <- dplyr::inner_join(sevdb_2,subsev, by="codigo") 
```

```{r}
#html(Hmisc::contents(sevdb_3), maxlevels=10, levelType='table')
sevdb_3 %>% dplyr::count(sev_WHO)
sevdb_3 %>% dplyr::count(episodio_previo)
sevdb_3 %>% group_by(episodio_previo, sev_WHO) %>% dplyr::count()
```

__¿episodio_previo==`NA`?__:

- ¿podemos corregir `NA` con covariables relacionadas a exposición?, o
- corregimos en forma insesgada a `sin`

```{r}
sevdb_3 %>% 
  dplyr::select(codigo,episodio_previo) %>% 
  filter(is.na(episodio_previo)) %>% 
  dplyr::inner_join(sevdb_pre %>% 
                      dplyr::rename(episodio_previo_num=malaria_pasado_num)%>% 
                      dplyr::select(codigo,episodio_previo_num,
                             fuente_agua:trabajo_riesgo) %>% #glimpse()
                      mutate(codigo= str_replace(codigo,"LIM0","LIM"))
                    , by="codigo")
#  filter(codigo=="LIM02017")
```

__POR DEFAULT: `NA`__
```{r,eval=FALSE}
sevdb_3 <- sevdb_3 %>% 
  #filter(is.na(episodio_previo)) %>% 
  #select(codigo, episodio_previo) %>% #glimpse()
  #mutate(episodio_previo= as.character(episodio_previo)) %>% 
  #mutate(episodio_previo= fct_recode(episodio_previo,"sin"=NA_character_)) %>% 
  #mutate(episodio_previo= funs(ifelse(is.na(.),"sin",.)))
  replace_na(list(episodio_previo="sin")) #%>% 
  #dplyr::count(episodio_previo)

sevdb_3 %>% dplyr::count(episodio_previo)
```



```{r, eval=FALSE}
# - Joint `parasitemia` from `28Nov2014_Ed.dta` **[WARNING]**

sevdb_4 <- sevdb_3 %>% 
  dplyr::inner_join(sevdb_x, by="codigo")
#html(Hmisc::contents(sevdb_4), maxlevels=10, levelType='table')

d <- describe(sevdb_4)
html(d, size=90, scroll=TRUE)

# THIS IS IN eval=FALSE !!!!!!!
sevdb_4 <- sevdb_3 %>% 
  haven::as_factor() %>%
  mutate_at(vars(-c(1:7),-ends_with("_num"),-abnomic), as.factor) %>% 
  mutate(admin_uci_corr= fct_recode(admin_uci_corr,NULL="No sabe"),
         ictericia_corr= fct_recode(ictericia_corr,NULL="No sabe"),
         cns_malaria= fct_recode(cns_malaria,"Si"="1", "No"="0"),
         glas_11= fct_recode(glas_11,"Si"="1", "No"="0"),
         sdra= fct_recode(sdra,"Si"="1", "No"="0"),
         lung_injury= fct_recode(lung_injury,"Si"="1", "No"="0"),
         hiperbil= fct_recode(hiperbil,"Si"="1", "No"="0"),
         shock= fct_recode(shock,"Si"="1", "No"="0"),
         bleeding= fct_recode(bleeding,"Si"="1", "No"="0"),
         glic_normal= fct_recode(glic_normal,"Si"="1", "No"="0"),
         hiperparasitemia= fct_recode(hiperparasitemia,"Si"="1", "No"="0"),
         anemia_severa= fct_recode(anemia_severa,"Si"="1", "No"="0"),
         presenta_comorb= fct_recode(presenta_comorb,"Si"="1", "No"="0")) %>% 
  dplyr::inner_join(sevdb_x)

html(Hmisc::contents(sevdb_4), maxlevels=10, levelType='table')
```

__RECLASIFICACIÓN en base a criterio OMS__
```{r}
sevdb_3 %>% 
  dplyr::select(codigo, sev_WHO) %>% 
  inner_join(sevdb_pre %>% 
               dplyr::select(codigo, casecontrol) %>% 
               mutate(codigo= str_replace(codigo,"LIM0","LIM")),
             by="codigo") %>% 
  group_by(casecontrol,sev_WHO) %>% dplyr::count()
```

__GRABAR BASE DE DATOS COMO `data/03-sevrcov.rds`__

```{r}
subsev <- readr::read_csv("data-raw/samples.csv") %>% 
  filter(Study=="Severe Malaria") %>% 
  #mutate(sample= Sample.ID==Subject.ID) %>% dplyr::count(sample)
  select(Sample.ID, Study) %>% 
  mutate(Study="in-microarray") %>% 
  dplyr::rename(codigo= Sample.ID)
subsev
# generate
#sevdb_3 <- dplyr::inner_join(sevdb_2,subsev, by="codigo") 

#pre-dataset
sevdb_2 %>% naniar::miss_var_summary()
#whole-subset dataset
sevdb_2ws <- sevdb_2 %>% 
  left_join(subsev) %>% 
  mutate(Study=if_else(is.na(Study),"whole-db",Study),# %>% as.factor(),
         #Study=as.factor(Study),
         #Study=forcats::fct_relevel(Study,"whole-db"),
         casecontrol=as.factor(casecontrol),
         sev_WHO_num=as.factor(sev_WHO_num),
         sev_WHO_num_1=as.factor(sev_WHO_num_1),
         sev_WHO_num_2=as.factor(sev_WHO_num_2),
         sev_WHO_num_3=as.factor(sev_WHO_num_3),
         sev_WHO_num_4=as.factor(sev_WHO_num_4),
         shock_oms=as.factor(shock_oms),
         pulmonar_oms=as.factor(pulmonar_oms),
         glas_oms=as.factor(glas_oms),
         hiperbil_oms=as.factor(hiperbil_oms),
         anemia_oms=as.factor(anemia_oms),
         renal_oms=as.factor(renal_oms),
         hipoglic_oms=as.factor(hipoglic_oms),
         convulsion=as.factor(convulsion),
         sangr_oms=as.factor(sangr_oms),
         #trombo_new=as.factor(trombo_new),
         trombo_sev=as.factor(trombo_sev),
         tx_par_intest=as.factor(tx_par_intest),
         presenta_comorb=as.factor(presenta_comorb),
         enf_cronica=as.factor(enf_cronica)) %>% 
  mutate(Study=forcats::fct_relevel(Study,"whole-db"),
         parasitemia_mil=parasitemia/1000) %>% 
  mutate(en_zona_endemica_4 = cut(en_zona_endemica,c(0,6,15,26,72)) %>% as.factor(),
         edad_4 = cut(edad,c(4,22,32,52,88)) %>% as.factor(),
         parasitemia_mil_4 = cut(parasitemia_mil,c(0,1.330,3.338,6.038,42.683)) %>% as.factor()
         )
sevdb_2ws %>% dplyr::count(Study)
sevdb_2ws %>% readr::write_rds("data/03-sevrcov-228obs_21var.rds")
# output
sevdb_3 %>% naniar::miss_var_summary()
sevdb_3sv <- sevdb_3 %>% 
  mutate(casecontrol=as.factor(casecontrol),
         sev_WHO_num=as.factor(sev_WHO_num),
         sev_WHO_num_1=as.factor(sev_WHO_num_1),
         sev_WHO_num_2=as.factor(sev_WHO_num_2),
         sev_WHO_num_3=as.factor(sev_WHO_num_3),
         sev_WHO_num_4=as.factor(sev_WHO_num_4),
         shock_oms=as.factor(shock_oms),
         pulmonar_oms=as.factor(pulmonar_oms),
         glas_oms=as.factor(glas_oms),
         hiperbil_oms=as.factor(hiperbil_oms),
         anemia_oms=as.factor(anemia_oms),
         renal_oms=as.factor(renal_oms),
         hipoglic_oms=as.factor(hipoglic_oms),
         convulsion=as.factor(convulsion),
         sangr_oms=as.factor(sangr_oms),
         #trombo_new=as.factor(trombo_new),
         trombo_sev=as.factor(trombo_sev),
         tx_par_intest=as.factor(tx_par_intest),
         presenta_comorb=as.factor(presenta_comorb),
         enf_cronica=as.factor(enf_cronica),
         parasitemia_mil=parasitemia/1000,
         en_zona_endemica_4 = cut(en_zona_endemica,c(0,6,15,26,72)) %>% as.factor(),
         edad_4 = cut(edad,c(4,22,32,52,88)) %>% as.factor(),
         parasitemia_mil_4 = cut(parasitemia_mil,c(0,1.330,3.338,6.038,42.683)) %>% as.factor()) %>% 
  readr::write_rds("data/03-sevrcov.rds")
```

```{r}
#inspeccion de comorbilidades y diagnosticos!!!
haven::read_dta("data-raw/malaria severa 28Nov2014_Ed.dta") %>% dplyr::count(presenta_comorb,sin_enf_cronica)
haven::read_dta("data-raw/malaria severa 28Nov2014_Ed.dta") %>% 
  select(starts_with("dx_"),presenta_comorb,sin_enf_cronica) %>% print(n=Inf)
haven::read_dta("data-raw/malaria severa 28Nov2014_Ed.dta") %>% 
  select(presenta_comorb:tx_par_intest) %>% print(n=Inf)
haven::read_dta("data-raw/malaria severa 28Nov2014_Ed.dta") %>% 
  select(codigo, parasitemia) %>% naniar::miss_var_summary()

#D/C: diagnostico clinico -> recomienda acción de descartar por x condición (indicación sin ver laboratorio)
haven::read_dta("data-raw/severe_malaria_vivax SEP2016.dta") %>% 
  select(starts_with("dx_"),presenta_comorb) %>% print(n=Inf)
#presenta_comorb == cronicas? rpta: sí!
haven::read_dta("data-raw/severe_malaria_vivax SEP2016.dta") %>% 
  select(presenta_comorb:tx_par_intest) %>% print(n=Inf)
haven::read_dta("data-raw/severe_malaria_vivax SEP2016.dta") %>% dplyr::count(presenta_comorb) #,sin_enf_cronica
haven::read_dta("data-raw/severe_malaria_vivax SEP2016.dta") %>% glimpse()
```

#### proportions per criteria {#abnomic-prop}

```{r}
###### TEST Hmisc::upData
sevdb_3x <- Hmisc::upData(sevdb_3,
                         labels = c(edad="Age (years)",
                                    sexo="Sex",
                                    shock_oms="Shock (systolic blood pressure < 80mmHg)",
                                    glas_oms="Coma (Glasgow score < 9/14)",
                                    pulmonar_oms="Lung injury (ARDS or pulmonary edema)",
                                    anemia_oms="Severe anemia (hemoglobin < 7mg/dL)",
                                    renal_oms="Acute renal failure (creatinin > 3mg/dL)",
                                    hiperbil_oms="Hyperbilirubinemia (bilirubin > 2.5mg/dL)", #Jaundice
                                    hipoglic_oms="Hypoglicemia (glucose < 40mg/dL)",
                                    convulsion="Convulsion",
                                    presenta_comorb="Comorbidity",
                                    enf_cronica="Chronic disease",
                                    #enf_sum="",
                                    parasitemia="Parasite density (par/uL)",
                                    #shock_oms_pa="",
                                    en_zona_endemica="Time in endemic area (years)",
                                    episodio_previo_num="Previous episodes (number)",
                                    sev_WHO_num="Cumulative WHO criteria",
                                    sev_WHO="WHO criteria",
                                    episodio_previo="Previous episodes"
                         ),
                         levels = list(sev_WHO=list("Non-severe"="no_severo",
                                                    "Severe"="severo"),
                                       episodio_previo=list("Without"="sin",
                                                            "With"="con"),
                                       sexo=list("Male"="masculino",
                                                 "Female"="femenino")
                                       )
                         )
```

```{r, eval=FALSE}
###### TEST Hmisc::upData
sevdb_3x <- Hmisc::upData(sevdb_3,
                         labels = c(edad="Age",
                                    sexo="Sex",
                                    shock_oms="Shock",
                                    glas_oms="Coma",
                                    pulmonar_oms="Lung injury",
                                    anemia_oms="Severe anemia",
                                    renal_oms="Acute renal failure",
                                    hiperbil_oms="Hyperbilirubinemia", #Jaundice
                                    hipoglic_oms="Hypoglicemia",
                                    convulsion="Convulsion",
                                    presenta_comorb="Comorbidity",
                                    enf_cronica="Chronic disease",
                                    #enf_sum="",
                                    parasitemia="Parasite density",
                                    #shock_oms_pa="",
                                    en_zona_endemica="Time in endemic area",
                                    episodio_previo_num="Previous episodes",
                                    sev_WHO_num="Cumulative WHO criteria",
                                    sev_WHO="WHO criteria",
                                    episodio_previo="Previous episodes"
                         ),
                         units = c(edad="(years)",
                                   shock_oms="(systolic blood pressure < 80mmHg)",
                                   glas_oms="(Glasgow score < 9/14)",
                                   pulmonar_oms="(ARDS or pulmonary edema)",
                                   anemia_oms="(hemoglobin < 7mg/dL)",
                                   renal_oms="(creatinin > 3mg/dL)",
                                   hiperbil_oms="(bilirubin > 2.5mg/dL)",
                                   hipoglic_oms="(glucose < 40mg/dL)",
                                   parasitemia="(par/uL)",
                                   en_zona_endemica="(years)",
                                   episodio_previo_num="(number)"
                                   #episodio_previo="prop"
                         ),
                         levels = list(sev_WHO=list("Non-severe"="no_severo",
                                                    "Severe"="severo"),
                                       episodio_previo=list("Without"="sin",
                                                            "With"="con"),
                                       sexo=list("Male"="masculino",
                                                 "Female"="femenino")
                                       )
                         )
```

```{r}
s1 <- Hmisc::summaryM(shock_oms + #clinicos
                        pulmonar_oms +
                        glas_oms +
                        hiperbil_oms + #lab
                        anemia_oms +
                        renal_oms +
                        hipoglic_oms +
                        #postracion +#extra
                        #coma +
                        convulsion +
                        #presenta_comorb+
                        #enf_cronica
                        #parasitemia + 
                        sev_WHO_num
                      ~ sev_WHO
                      ,
               data=sevdb_3x,
               overall=FALSE, test=FALSE)

latex(s1, caption='WHO classification: Clinical and Laboratory criteria',
      exclude1=TRUE, npct='both', 
      digits=2,
      #prmsd=TRUE, brmsd=TRUE, msdsize=mu$smaller2, #NOT-EVALUATE if PDF
      middle.bold=TRUE, long = TRUE,
      #legend.bottom = TRUE, #insert.bottom = TRUE, 
      what="%", html = TRUE
      ) #change here for LaTeX PDF
```

#### combinations of criteria

__check `sev_WHO`__

- frecuencia por criterios acumulados

```{r}
sevdb_3 %>% group_by(sev_WHO_num, sev_WHO#, sev_SNC_num, sev_ALL_num
                     ) %>% dplyr::count()
```

- frecuencia por combinación de criterios

```{r}
# SUM SEVERITY CRITERIA
#sevdb_o <- sevdb_2 %>% 
#  dplyr::rename(glas=glas_11,
#         #sdra,
#         lung=lung_injury,
#         icter=ictericia_corr,
#         bili=hiperbil,
#         #shock,
#         bleed=bleeding,
#         glic=glic_normal,
#         paras=hiperparasitemia,
#         anemia=anemia_severa#, 
#         #hemo=hemoglobinuria_corr
#         ) 
sevdb_3 %>% 
  group_by(shock_oms, glas_oms, pulmonar_oms, #sdra,snc_oms,
           hipoglic_oms, anemia_oms, renal_oms, hiperbil_oms,
           convulsion,
           #coma, postracion,
           sev_WHO_num, sev_WHO
           ) %>% 
  dplyr::count() %>% 
  dplyr::arrange(desc(#n
                      sev_WHO_num))
```
```{r}
sevdb_3 %>% 
  dplyr::select(shock_oms, glas_oms, pulmonar_oms, #sdra,snc_oms,
         hipoglic_oms, anemia_oms, renal_oms, hiperbil_oms,
         convulsion#,
         #coma, postracion
           ) %>% 
  summarise_all(sum,na.rm=T) #%>% t()
```

__ningún paciente tiene `pa_sistolica` definida:__

```{r}
sevdb_3 %>% 
  dplyr::select(codigo, starts_with("sev_WHO"), shock_oms) %>% 
  inner_join(sevdb_pos_1 %>% 
               dplyr::select(codigo, pa_sistolica) %>% 
               mutate(codigo= str_replace(codigo,"LIM0","LIM")),
             by="codigo") %>% 
  dplyr::count(pa_sistolica)
```
```{r, eval=FALSE}
sevdb_3 %>% dplyr::select(codigo, sev_WHO, sev_WHO_num) %>% 
  inner_join(sevdb_1oms, by="codigo") %>% 
  group_by(sev_WHO, sev_WHO_num, shock_oms, 
           #snc_oms, snc_sum, 
           convulsion#,
           #coma, postracion
           #, snc_normal
           ) %>% dplyr::count()
```
```{r, eval=FALSE}
sevdb_3 %>% dplyr::select(codigo, sev_WHO, sev_WHO_num) %>% 
  inner_join(sevdb_1oms, by="codigo") %>% 
  group_by(sev_WHO, sev_WHO_num, 
           pulmonar_oms, pulmonar_sum, sdra, edema_pulm, disnea, taquipnea, infiltracion
           ) %>% dplyr::count()
```

##### reproducible subset comparison

__Comparación con la elección de 30 muestras:__

- coinciden en 10 pacientes: incluye al paciente con más criterios
- excluye a 08 pacientes con dos o más criterios, y
- excluye a 19 pacientes con un criterio y las características dadas [previamente](#reproducible-subsetting)

```{r, fig.align='center', fig.width=5, fig.height=5}
g1_id <- sevdb_30 %>% dplyr::select(codigo)      # 960
g2_id <- sevdb_3 %>% dplyr::filter(sev_WHO=="severo") %>% dplyr::select(codigo)     # 963

g1_idl= unlist(as.list(g1_id))
g2_idl= unlist(as.list(g2_id))

tmp <- gplots::venn(list(sevdb_30=g1_idl, sevdb_3=g2_idl))
```

```{r}
dplyr::inner_join(sevdb_30,
                  sevdb_3 %>% 
                    dplyr::filter(sev_WHO=="severo"),
                  by="codigo") %>% dplyr::count(sev_WHO_num.x)
dplyr::anti_join(sevdb_30,
                  sevdb_3 %>% 
                    dplyr::filter(sev_WHO=="severo"),
                  by="codigo") %>% dplyr::count(sev_WHO_num)
dplyr::anti_join(sevdb_3 %>% 
                    dplyr::filter(sev_WHO=="severo"),
                 sevdb_30,
                  by="codigo") %>% dplyr::count(sev_WHO_num)
```

#### Descriptive statistics {#abnomic-descrip}

* se describen las variables evaluadas en:
    - @smith2013,
    - @baldevi2013, y
    - @quispe2014


```{r, eval=FALSE}
#- Joint `parasitemia` from `28Nov2014_Ed.dta` **[WARNING]**

sevdb_x <- haven::read_dta("data-raw/malaria severa 28Nov2014_Ed.dta") %>% 
  dplyr::select(codigo, parasitemia)%>% 
  mutate(codigo= str_replace(codigo,"LIM0","LIM"))

sevdb_3x <- sevdb_3 %>% 
  haven::as_factor() %>% 
  dplyr::inner_join(sevdb_x,by="codigo")
```

***

**wrt severity**

- significant differences with respect to 
    + proportion of previous episodes of malaria,
    + proportion of women patients towards severe cases

- observaciones:
    + a diferencia de la [muestra total](#sev-inf) (n=228), 
        - menor diferencia en la proporción de mujeres severas (37 vs 55%)
        - menor diferencia en la proporción de casos severos con episodio previo (66 vs 46%)
    + ~~PROBLEMA: alta proporción de pacientes con comorbilidades (?) (20 vs 30%)~~
        - comorbilidad: 14% (5/37) vs 43% (10/23)
        - enf_cronica: 8% (3/37) vs 17% (4/23)

```{r}
s1 <- Hmisc::summaryM(edad + sexo +
                        en_zona_endemica +
                        episodio_previo_num + episodio_previo + 
                        #sev_WHO + sev_WHO_num +
                        #presenta_comorb + enf_cronica +
                        parasitemia #+
                      ~ sev_WHO #+ 
                      #episodio_previo
                      ,
               data=sevdb_3x,
               overall=FALSE, test=TRUE)

latex(s1, caption='Epidemiological data',
      exclude1=TRUE, npct='both', 
      digits=2, eps = 0.01,
      #prmsd=TRUE, brmsd=TRUE, msdsize=mu$smaller2, #NOT-EVALUATE if PDF
      middle.bold=TRUE, long = FALSE,
      #legend.bottom = TRUE, #insert.bottom = TRUE, 
      what="%", html = TRUE
      ) #change here for LaTeX PDF

```

***

**wrt previous episodes of malaria**

- significant differences with respect to 
    + mean age,
    + proportion of severe patients

```{r}
s1 <- Hmisc::summaryM(edad + sexo +
                        en_zona_endemica +
                        episodio_previo_num + #episodio_previo +
                        sev_WHO + #sev_WHO_num +
                        presenta_comorb + enf_cronica +
                        parasitemia #+
                      ~ #sev_WHO #+ 
                      episodio_previo
                      ,
               data=sevdb_3x,
               overall=FALSE, test=FALSE)

latex(s1, caption='Epidemiological data',
      exclude1=TRUE, npct='both', 
      digits=2,
      #prmsd=TRUE, brmsd=TRUE, msdsize=mu$smaller2, #NOT-EVALUATE if PDF
      middle.bold=TRUE, long = FALSE,
      #legend.bottom = TRUE, #insert.bottom = TRUE, 
      what="%", html = TRUE
      ) #change here for LaTeX PDF
```

***

**wrt severity by previous episodes of malaria**

- no significant differences with respect to any variable

```{r}
s1 <- Hmisc::summaryM(edad + sexo +
                        en_zona_endemica +
                        #episodio_previo + #episodio_previo_num + 
                        #sev_WHO + #sev_WHO_num +
                        presenta_comorb + enf_cronica +
                        parasitemia #+
                      ~ sev_WHO + 
                      episodio_previo
                      ,
               data=sevdb_3x,
               overall=FALSE, test=FALSE)

latex(s1, caption='Epidemiological data',
      exclude1=TRUE, npct='both', 
      digits=2,
      #prmsd=TRUE, brmsd=TRUE, msdsize=mu$smaller2, #NOT-EVALUATE if PDF
      middle.bold=TRUE, long = FALSE,
      #legend.bottom = TRUE, #insert.bottom = TRUE, 
      what="%", html = TRUE
      ) #change here for LaTeX PDF
```

```{r, fig.width=6, fig.height=5, fig.align='center'}
#plot(s1, which='categorical')
#Key(0,0)
```

```{r, fig.width=8, fig.height=4, fig.align='center'}
#par(mfrow=c(1,2))
#plot(s1, which='continuous')
#par(mfrow=c(1,1))
```

## PRESENTATION

```{r,fig.height=3,fig.width=4}
data_s <- sevdb_2ws %>% 
  mutate(edad_10=cut(edad,c(quantile(edad,probs = seq(0,1,0.1),na.rm = T)[1],
                                       quantile(edad,probs = seq(0,1,0.1),na.rm = T)[2],
                                       quantile(edad,probs = seq(0,1,0.1),na.rm = T)[3],
                                       quantile(edad,probs = seq(0,1,0.1),na.rm = T)[4],
                                       quantile(edad,probs = seq(0,1,0.1),na.rm = T)[5],
                                       quantile(edad,probs = seq(0,1,0.1),na.rm = T)[6],
                                       quantile(edad,probs = seq(0,1,0.1),na.rm = T)[7],
                                       quantile(edad,probs = seq(0,1,0.1),na.rm = T)[8],
                                       quantile(edad,probs = seq(0,1,0.1),na.rm = T)[9],
                                       quantile(edad,probs = seq(0,1,0.1),na.rm = T)[10],
                                       quantile(edad,probs = seq(0,1,0.1),na.rm = T)[11]),
                           include.lowest = T),
         edad_4=cut(edad,c(quantile(edad,probs = seq(0,1,0.25),na.rm = T)[1],
                                       quantile(edad,probs = seq(0,1,0.25),na.rm = T)[2],
                                       quantile(edad,probs = seq(0,1,0.25),na.rm = T)[3],
                                       quantile(edad,probs = seq(0,1,0.25),na.rm = T)[4],
                                       quantile(edad,probs = seq(0,1,0.25),na.rm = T)[5]),
                           include.lowest = T)) 

#data_s %>% dplyr::count(edad)
data_s %>% dplyr::count(edad_4)

data_s %>% filter(!is.na(edad)) %>% 
  ggplot(aes(edad_4)) +
  geom_bar(aes(fill=sev_WHO),position = position_fill()) + 
  theme(axis.text.x = element_text(angle = 45, hjust = 1))
```

### table 1: whole-subset covariates

```{r}
#hist(sevdb_2ws$episodio_previo_num)
#sevdb_2ws %>% dplyr::count(episodio_previo_num)
tab_ws <- compareGroups(Study ~ 
                          edad + sexo +
                          en_zona_endemica +
                          enf_cronica + #presenta_comorb + 
                          #tx_par_intest + 
                          parasitemia_mil + 
                          episodio_previo + #episodio_previo_num + 
                          sev_WHO_2 #+ #sev_WHO_num +
                ,
              data = sevdb_2ws, byrow=F, method=c(edad=2,en_zona_endemica=2,parasitemia_mil=2)
              ) 
tab_ws %>% createTable(show.all = T,show.p.overall = F)
tab_ws %>% createTable(show.all = T,show.p.overall = F) %>% export2xls("table/tab1-in_out.xls")
```

### table 2: whole-subset criterias

```{r}
#hist(sevdb_2ws$episodio_previo_num)
#sevdb_2ws %>% dplyr::count(episodio_previo_num)
tab_cr <- compareGroups(Study ~ 
                          sev_WHO_2 +
                          anemia_oms +
                          shock_oms + #clinicos
                          pulmonar_oms +
                          renal_oms +
                          glas_oms +
                          convulsion +
                          hiperbil_oms + #lab
                          hipoglic_oms +
                          #postracion +#extra
                          #coma +
                          trombo_sev +
                          sangr_oms + #trombo_new
                          #presenta_comorb+
                          #enf_cronica
                          #parasitemia + 
                          sev_WHO_num_2
                ,
              data = sevdb_2ws#, byrow=T#, method=c(edad=2,en_zona_endemica=2,episodio_previo_num=2,parasitemia=2)
              ) 
#tab_cr %>% createTable(show.all = T,show.n = T,hide.no = "0") #%>% export2xls("table/tab1.xls")
tab_cr %>% createTable(show.all = T,show.n = F,show.p.overall = F,hide.no = "0") #%>% export2xls("table/tab1.xls")
tab_cr %>% createTable(show.all = T,show.n = F,show.p.overall = F,hide.no = "0") %>% export2xls("table/tab2-in_out_oms.xls")
```

```{r,eval=FALSE}
sevdb_2ws %>% 
  filter(sev_WHO_num==1 & Study == "in-microarray") %>% 
  select(anemia_oms:trombo_sev,sev_WHO,sev_WHO_2)

hist(sevdb_2ws$en_zona_endemica)
summary(sevdb_2ws$en_zona_endemica)
hist(sevdb_2ws$edad)
summary(sevdb_2ws$edad)
hist(sevdb_2ws$parasitemia_mil)
summary(sevdb_2ws$parasitemia_mil)
hist(sevdb_2ws$parasitemia)
summary(sevdb_2ws$parasitemia)
```

### table 3: whole counfunders

```{r}
#hist(sevdb_2ws$episodio_previo_num)
#sevdb_2ws %>% dplyr::count(episodio_previo_num)
tab_sv <- compareGroups(sev_WHO_2 ~ 
                          edad + #edad_4 + 
                          sexo +
                          en_zona_endemica + #en_zona_endemica_4 +
                          enf_cronica + #presenta_comorb + 
                          tx_par_intest + 
                          parasitemia_mil + #parasitemia_mil_4 +
                          episodio_previo #+ #episodio_previo_num + 
                          #sev_WHO #+ #sev_WHO_num +
                ,
              data = sevdb_2ws, byrow=T, method=c(edad=2,en_zona_endemica=2,episodio_previo_num=2,parasitemia_mil=2)
              ) 
tab_sv %>% createTable(show.all = T,show.n = F) #%>% export2xls("table/tab3-all-severe.xls")
tab_sv %>% createTable(show.all = T,show.n = F) %>% export2xls("table/tab3-all-severe.xls")
#tab_sv %>% createTable() #%>% export2xls("table/tab1.xls")
```


### table 3: subset counfunders

```{r}
#hist(sevdb_2ws$episodio_previo_num)
#sevdb_2ws %>% dplyr::count(episodio_previo_num)
tab_sv <- compareGroups(sev_WHO_2 ~ 
                          edad + #edad_4 + 
                          sexo +
                          en_zona_endemica + #en_zona_endemica_4 +
                          enf_cronica + #presenta_comorb + 
                          tx_par_intest + 
                          parasitemia_mil + #parasitemia_mil_4 +
                          episodio_previo #+ #episodio_previo_num + 
                          #sev_WHO #+ #sev_WHO_num +
                ,
              data = sevdb_3sv, byrow=T, method=c(edad=2,en_zona_endemica=2,episodio_previo_num=2,parasitemia_mil=2)
              ) 
tab_sv %>% createTable(show.all = T,show.n = F) #%>% export2xls("table/tab1.xls")
tab_sv %>% createTable(show.all = T,show.n = F) %>% export2xls("table/tab3-in_severe.xls")
```

## ANEXO

### Additional plots

- __TROMBOCITOPENIA__
```{r, message=FALSE, fig.align="center", fig.width=12, fig.height=4}
a <- sevdb_1oms %>%
  dplyr::select(trombocitopenia, cont_plaquetas) %>% #dplyr::count()
  mutate(trombo_sev= if_else(cont_plaquetas<50000,1,0)) %>% 
  mutate(trombo_sev= as.factor(trombo_sev)) %>% #glimpse()
  ggplot(aes(x=cont_plaquetas, fill=trombo_sev)) + 
  geom_histogram(alpha=.5, position = "identity") + 
  labs(title="trombocitopenia")

b <- sevdb_1oms %>%
  dplyr::select(codigo, trombocitopenia, cont_plaquetas) %>% #dplyr::count()
  mutate(trombo_sev= if_else(cont_plaquetas<50000,1,0)) %>% 
  mutate(trombo_sev= as.factor(trombo_sev)) %>% #glimpse()
  inner_join(sevdb_2 %>% 
               dplyr::select(codigo, sev_WHO),
             by="codigo") %>% 
  filter(sev_WHO=="severo") %>% 
  ggplot(aes(x=cont_plaquetas, fill=trombo_sev)) + 
  geom_histogram(alpha=.5, position = "identity") + 
  labs(title="trombocitopenia",
       subtitle="filter severe samples")

Rmisc::multiplot(a,b,cols = 2)
```

- __EDAD__

```{r, message=FALSE, fig.align="center", fig.width=4, fig.height=4}
ggplot(sevdb_2, aes(x=edad)) + 
  geom_histogram(
    #aes(y=..density..), 
                 breaks=seq(0, 85, by = 5), 
                 alpha=.5 ,position = "identity") + 
  labs(title="Histograma de Edad") +
  #geom_density() #+
  scale_x_continuous(breaks=seq(0, 85, by = 5)) #+
  #labs(x="Age", y="Count") + 
  #xlim(c(18,52)) + 
  #ylim(c(0,30))
#scale_y_continuous(limits = c(0,0.05))
```

```{r, message=FALSE, fig.align="center", fig.width=10, fig.height=7}
a <- ggplot(sevdb_2, aes(x=edad, fill=sev_WHO)) + 
  geom_histogram(
    #aes(y=..density..), 
                 breaks=seq(0, 85, by = 5), 
                 alpha=.5 ,position = "identity") + 
  labs(title="Histograma de Edad") +
  #geom_density() #+
  scale_x_continuous(breaks=seq(0, 85, by = 5)) #+
  #labs(x="Age", y="Count") + 
  #xlim(c(18,52)) + 
  #ylim(c(0,30))
#scale_y_continuous(limits = c(0,0.05))

b <- ggplot(sevdb_2, aes(x=edad, fill=episodio_previo)) + 
  geom_histogram(
    #aes(y=..density..), 
                 breaks=seq(0, 85, by = 5), 
                 alpha=.5 ,position = "identity") + 
  labs(title="Histograma de Edad") +
  #geom_density() #+
  scale_x_continuous(breaks=seq(0, 85, by = 5)) #+
  #labs(x="Age", y="Count") + 
  #xlim(c(18,52)) + 
  #ylim(c(0,30))
  #scale_y_continuous(limits = c(0,0.05))

c <- sevdb_2 %>% 
  filter(sev_WHO=="no_severo") %>% 
  ggplot(aes(x=edad, fill=episodio_previo)) + 
  geom_histogram(
    #aes(y=..density..), 
                 breaks=seq(0, 85, by = 5), 
                 alpha=.5 ,position = "identity") + 
  labs(title="Histograma de Edad",
       subtitle="filter no_severe samples") +
  #geom_density() #+
  scale_x_continuous(breaks=seq(0, 85, by = 5))

d <- sevdb_2 %>% 
  filter(sev_WHO=="severo") %>% 
  ggplot(aes(x=edad, fill=episodio_previo)) + 
  geom_histogram(
    #aes(y=..density..), 
                 breaks=seq(0, 85, by = 5), 
                 alpha=.5 ,position = "identity") + 
  labs(title="Histograma de Edad",
       subtitle="filter severe samples") +
  #geom_density() #+
  scale_x_continuous(breaks=seq(0, 85, by = 5))

#c <- ggplot(sampleSEV, aes(x=edad, fill=casecontrol)) + 
  #geom_histogram(
   # aes(y=..density..), 
    #             breaks=seq(0, 85, by = 5), position = "dodge") + 
  #labs(title="Histograma de Edad") +
  #geom_density(alpha=.3) #+
  #scale_x_continuous(breaks=seq(0, 85, by = 5))
  #labs(x="Age", y="Count") + 
  #xlim(c(18,52)) + 
  #ylim(c(0,30))

#d <- ggplot(sampleSEV, aes(x=edad, fill=con_exp)) + 
  ##geom_histogram(
    ##aes(y=..density..), 
                 ##breaks=seq(0, 85, by = 5), position = "dodge") + 
  ##labs(title="Histograma de Edad") +
  #geom_density(alpha=.3) +
  ##scale_x_continuous(breaks=seq(0, 85, by = 5)) +
  ##labs(x="Age", y="Count") + 
  ##xlim(c(18,52)) + 
  #scale_y_continuous(limits = c(0,0.025))


multiplot(a,c,b,d,cols=2)
```

- __TIEMPO VIVENDO EN ZONA ENDÉMICA__
```{r, message=FALSE, fig.align="center", fig.width=12, fig.height=4}
a <- sevdb_2 %>% 
  ggplot() + 
  geom_point(aes(edad, en_zona_endemica, colour=sev_WHO)) + 
  #scale_color_viridis() +
  labs(title="edad vs años vivendo en localidad",
       subtitle= "sep2016")

b <- sevdb_2 %>% 
  ggplot() + 
  geom_point(aes(edad, en_zona_endemica, colour=episodio_previo)) + 
  #scale_color_viridis() +
  labs(title="edad vs años vivendo en localidad",
       subtitle= "sep2016")

Rmisc::multiplot(a,b,cols = 2)
```

### Show Consensus covariates

- se __replica__ las variables consenso obtenidas por variable

##### general

```{r}
sevdb_1 %>% 
  mutate(edad= edad_1==edad_2) %>%
  dplyr::select(edad, edad_1, edad_2) %>%
  tidyr::unite(edad_c, edad_1, edad_2, remove=F) %>% #count()
  mutate(edad_c=str_replace(edad_c,"(\\d{2})_\\1", "\\1")) %>%
  mutate(edad_c=str_replace(edad_c,"NA_(\\d+)", "\\1")) %>% 
  mutate(edad_c=str_replace(edad_c,"(\\d+)_NA", "\\1")) %>% 
  mutate(edad_c=str_replace(edad_c,"(\\d+)_(\\d+)", "\\2")) %>% #count() #preferencia a _2
  mutate(edad_c=str_replace(edad_c,"NA", replacement = NA_character_)) %>%  #%>% count()
  mutate(edad_c= as.numeric(edad_c))
```

```{r}
sevdb_1 %>% 
  mutate(sexo= sexo_1==sexo_2) %>%
  dplyr::select(sexo, sexo_1, sexo_2) %>%
  tidyr::unite(sexo_c, sexo_1, sexo_2, remove=F) %>% #count()
  mutate(sexo_c=str_replace(sexo_c,"(\\d+)_\\1", "\\1")) %>%
  mutate(sexo_c=str_replace(sexo_c,"NA_(\\d+)", "\\1")) %>% 
  mutate(sexo_c=str_replace(sexo_c,"(\\d+)_NA", "\\1")) %>% 
  mutate(sexo_c=str_replace(sexo_c,"(\\d+)_(\\d+)", "\\2")) %>% #count() #preferencia a _2
  mutate(sexo_c=str_replace(sexo_c,"NA", replacement = NA_character_)) %>%  #%>% count()
  mutate(sexo_c= as.numeric(sexo_c))
```

##### who criteria

```{r}
sevdb_1 %>% 
  mutate(shock= shock_1==shock_2) %>%
  dplyr::select(shock, shock_1, shock_2) %>%
  tidyr::unite(shock_c, shock_1, shock_2, remove=F) %>% #count()
  mutate(shock_c=str_replace(shock_c,"(\\d+)_\\1", "\\1")) %>%
  mutate(shock_c=str_replace(shock_c,"NA_(\\d+)", "\\1")) %>% 
  mutate(shock_c=str_replace(shock_c,"(\\d+)_NA", "\\1")) %>% 
  mutate(shock_c=str_replace(shock_c,"(\\d+)_(\\d+)", "\\2")) %>% #count() #preferencia a _2
  mutate(shock_c=str_replace(shock_c,"NA", replacement = NA_character_)) %>%  #%>% count()
  mutate(shock_c= as.numeric(shock_c))
```

```{r}
sevdb_1 %>% 
  mutate(pa_sistolica= pa_sistolica_1==pa_sistolica_2) %>%
  dplyr::select(pa_sistolica, pa_sistolica_1, pa_sistolica_2) %>%
  tidyr::unite(pa_sistolica_c, pa_sistolica_1, pa_sistolica_2, remove=F) %>% #count()
  mutate(pa_sistolica_c=str_replace(pa_sistolica_c,"(\\d+)_\\1", "\\1")) %>%
  mutate(pa_sistolica_c=str_replace(pa_sistolica_c,"NA_(\\d+)", "\\1")) %>% 
  mutate(pa_sistolica_c=str_replace(pa_sistolica_c,"(\\d+)_NA", "\\1")) %>% 
  mutate(pa_sistolica_c=str_replace(pa_sistolica_c,"(\\d+)_(\\d+)", "\\2")) %>% #count() #preferencia a _2
  mutate(pa_sistolica_c=str_replace(pa_sistolica_c,"NA", replacement = NA_character_)) %>%  #%>% count()
  mutate(pa_sistolica_c= as.numeric(pa_sistolica_c))
```

```{r}
sevdb_1 %>% 
  mutate(glasgow= glasgow_1==glasgow_2) %>%
  dplyr::select(glasgow, glasgow_1, glasgow_2) %>%
  tidyr::unite(glasgow_c, glasgow_1, glasgow_2, remove=F) %>% #count()
  mutate(glasgow_c=str_replace(glasgow_c,"(\\d+)_\\1", "\\1")) %>%
  mutate(glasgow_c=str_replace(glasgow_c,"NA_(\\d+)", "\\1")) %>% 
  mutate(glasgow_c=str_replace(glasgow_c,"(\\d+)_NA", "\\1")) %>% 
  mutate(glasgow_c=str_replace(glasgow_c,"(\\d+)_(\\d+)", "\\2")) %>% #count() #preferencia a _2
  mutate(glasgow_c=str_replace(glasgow_c,"NA", replacement = NA_character_)) %>%  #%>% count()
  mutate(glasgow_c= as.numeric(glasgow_c))
```

```{r}
sevdb_1 %>% 
  mutate(glas_ojos= glas_ojos_1==glas_ojos_2) %>%
  dplyr::select(glas_ojos, glas_ojos_1, glas_ojos_2) %>%
  tidyr::unite(glas_ojos_c, glas_ojos_1, glas_ojos_2, remove=F) %>% #count()
  mutate(glas_ojos_c=str_replace(glas_ojos_c,"(\\d+)_\\1", "\\1")) %>%
  mutate(glas_ojos_c=str_replace(glas_ojos_c,"NA_(\\d+)", "\\1")) %>% 
  mutate(glas_ojos_c=str_replace(glas_ojos_c,"(\\d+)_NA", "\\1")) %>% 
  mutate(glas_ojos_c=str_replace(glas_ojos_c,"(\\d+)_(\\d+)", "\\2")) %>% #count() #preferencia a _2
  mutate(glas_ojos_c=str_replace(glas_ojos_c,"NA", replacement = NA_character_)) %>%  #%>% count()
  mutate(glas_ojos_c= as.numeric(glas_ojos_c))
```

```{r}
sevdb_1 %>% 
  mutate(glas_verbal= glas_verbal_1==glas_verbal_2) %>%
  dplyr::select(glas_verbal, glas_verbal_1, glas_verbal_2) %>%
  tidyr::unite(glas_verbal_c, glas_verbal_1, glas_verbal_2, remove=F) %>% #count()
  mutate(glas_verbal_c=str_replace(glas_verbal_c,"(\\d+)_\\1", "\\1")) %>%
  mutate(glas_verbal_c=str_replace(glas_verbal_c,"NA_(\\d+)", "\\1")) %>% 
  mutate(glas_verbal_c=str_replace(glas_verbal_c,"(\\d+)_NA", "\\1")) %>% 
  mutate(glas_verbal_c=str_replace(glas_verbal_c,"(\\d+)_(\\d+)", "\\2")) %>% #count() #preferencia a _2
  mutate(glas_verbal_c=str_replace(glas_verbal_c,"NA", replacement = NA_character_)) %>%  #%>% count()
  mutate(glas_verbal_c= as.numeric(glas_verbal_c))
```

```{r}
sevdb_1 %>% 
  mutate(glas_motor= glas_motor_1==glas_motor_2) %>%
  dplyr::select(glas_motor, glas_motor_1, glas_motor_2) %>%
  tidyr::unite(glas_motor_c, glas_motor_1, glas_motor_2, remove=F) %>% #count()
  mutate(glas_motor_c=str_replace(glas_motor_c,"(\\d+)_\\1", "\\1")) %>%
  mutate(glas_motor_c=str_replace(glas_motor_c,"NA_(\\d+)", "\\1")) %>% 
  mutate(glas_motor_c=str_replace(glas_motor_c,"(\\d+)_NA", "\\1")) %>% 
  mutate(glas_motor_c=str_replace(glas_motor_c,"(\\d+)_(\\d+)", "\\2")) %>% #count() #preferencia a _2
  mutate(glas_motor_c=str_replace(glas_motor_c,"NA", replacement = NA_character_)) %>%  #%>% count()
  mutate(glas_motor_c= as.numeric(glas_motor_c))
```

```{r}
sevdb_1 %>% 
  mutate(coma= coma_1==coma_2) %>%
  dplyr::select(coma, coma_1, coma_2) %>%
  tidyr::unite(coma_c, coma_1, coma_2, remove=F) %>% #count()
  mutate(coma_c=str_replace(coma_c,"(\\d+)_\\1", "\\1")) %>%
  mutate(coma_c=str_replace(coma_c,"NA_(\\d+)", "\\1")) %>% 
  mutate(coma_c=str_replace(coma_c,"(\\d+)_NA", "\\1")) %>% 
  mutate(coma_c=str_replace(coma_c,"(\\d+)_(\\d+)", "\\2")) %>% #count() #preferencia a _2
  mutate(coma_c=str_replace(coma_c,"NA", replacement = NA_character_)) %>%  #%>% count()
  mutate(coma_c= as.numeric(coma_c))
```

```{r}
sevdb_1 %>% 
  mutate(convulsion= convulsion_1==convulsion_2) %>%
  dplyr::select(convulsion, convulsion_1, convulsion_2) %>%
  tidyr::unite(convulsion_c, convulsion_1, convulsion_2, remove=F) %>% #count()
  mutate(convulsion_c=str_replace(convulsion_c,"(\\d+)_\\1", "\\1")) %>%
  mutate(convulsion_c=str_replace(convulsion_c,"NA_(\\d+)", "\\1")) %>% 
  mutate(convulsion_c=str_replace(convulsion_c,"(\\d+)_NA", "\\1")) %>% 
  mutate(convulsion_c=str_replace(convulsion_c,"(\\d+)_(\\d+)", "\\2")) %>% #count() #preferencia a _2
  mutate(convulsion_c=str_replace(convulsion_c,"NA", replacement = NA_character_)) %>%  #%>% count()
  mutate(convulsion_c= as.numeric(convulsion_c))
```

```{r}
sevdb_1 %>% 
  mutate(edema_pulm= edema_pulm_1==edema_pulm_2) %>%
  dplyr::select(edema_pulm, edema_pulm_1, edema_pulm_2) %>%
  tidyr::unite(edema_pulm_c, edema_pulm_1, edema_pulm_2, remove=F) %>% #count()
  mutate(edema_pulm_c=str_replace(edema_pulm_c,"(\\d+)_\\1", "\\1")) %>%
  mutate(edema_pulm_c=str_replace(edema_pulm_c,"NA_(\\d+)", "\\1")) %>% 
  mutate(edema_pulm_c=str_replace(edema_pulm_c,"(\\d+)_NA", "\\1")) %>% 
  mutate(edema_pulm_c=str_replace(edema_pulm_c,"(\\d+)_(\\d+)", "\\2")) %>% #count() #preferencia a _2
  mutate(edema_pulm_c=str_replace(edema_pulm_c,"NA", replacement = NA_character_)) %>%  #%>% count()
  mutate(edema_pulm_c= as.numeric(edema_pulm_c))
```

```{r}
sevdb_1 %>% 
  mutate(sdra= sdra_1==sdra_2) %>%
  dplyr::select(sdra, sdra_1, sdra_2) %>%
  tidyr::unite(sdra_c, sdra_1, sdra_2, remove=F) %>% #count()
  mutate(sdra_c=str_replace(sdra_c,"(\\d+)_\\1", "\\1")) %>%
  mutate(sdra_c=str_replace(sdra_c,"NA_(\\d+)", "\\1")) %>% 
  mutate(sdra_c=str_replace(sdra_c,"(\\d+)_NA", "\\1")) %>% 
  mutate(sdra_c=str_replace(sdra_c,"(\\d+)_(\\d+)", "\\2")) %>% #count() #preferencia a _2
  mutate(sdra_c=str_replace(sdra_c,"NA", replacement = NA_character_)) %>%  #%>% count()
  mutate(sdra_c= as.numeric(sdra_c))
```

```{r, eval=FALSE}
sevdb_1 %>% 
  mutate(cont_plaquetas= cont_plaquetas_1==cont_plaquetas_2) %>%
  dplyr::select(cont_plaquetas, cont_plaquetas_1, cont_plaquetas_2) %>% 
  tidyr::unite(cont_plaquetas_c, cont_plaquetas_1, cont_plaquetas_2, remove = F) %>% #count()
  mutate(cont_plaquetas_c=str_replace(cont_plaquetas_c,"(\\d{4,})_\\1", "\\1")) %>%
  mutate(cont_plaquetas_c=str_replace(cont_plaquetas_c,"NA_(\\d{4,})", "\\1")) %>% 
  mutate(cont_plaquetas_c=str_replace(cont_plaquetas_c,"(\\d{4,})_NA", "\\1")) %>% 
  mutate(cont_plaquetas_c=str_replace(cont_plaquetas_c,"(\\d{4,})_(\\d{4,})", "\\2")) %>% #count() #preferencia a _2
  mutate(cont_plaquetas_c=str_replace(cont_plaquetas_c,"(\\d{4,})_(\\d{1,3})", "\\1")) %>% #count() #problemas en _2
  mutate(cont_plaquetas_c=str_replace(cont_plaquetas_c,"NA", replacement = NA_character_)) %>%  #%>% count()
  mutate(cont_plaquetas_c=str_replace(cont_plaquetas_c,"(.....)_\\1", "\\1")) %>% #count()
  mutate(cont_plaquetas_c= as.numeric(cont_plaquetas_c))
```

##### conditional

```{r}
sevdb_1 %>% 
  mutate(presenta_comorb= presenta_comorb_1==presenta_comorb_2) %>%
  dplyr::select(presenta_comorb, presenta_comorb_1, presenta_comorb_2) %>%
  tidyr::unite(presenta_comorb_c, presenta_comorb_1, presenta_comorb_2, remove=F) %>% #count()
  mutate(presenta_comorb_c=str_replace(presenta_comorb_c,"(\\d+)_\\1", "\\1")) %>%
  mutate(presenta_comorb_c=str_replace(presenta_comorb_c,"NA_(\\d+)", "\\1")) %>% 
  mutate(presenta_comorb_c=str_replace(presenta_comorb_c,"(\\d+)_NA", "\\1")) %>% 
  mutate(presenta_comorb_c=str_replace(presenta_comorb_c,"(\\d+)_(\\d+)", "\\2")) %>% #count() #preferencia a _2
  mutate(presenta_comorb_c=str_replace(presenta_comorb_c,"NA", replacement = NA_character_)) %>%  #%>% count()
  mutate(presenta_comorb_c= as.numeric(presenta_comorb_c))
```

```{r}
sevdb_1 %>% 
  mutate(diabetes= diabetes_1==diabetes_2) %>%
  dplyr::select(diabetes, diabetes_1, diabetes_2) %>%
  tidyr::unite(diabetes_c, diabetes_1, diabetes_2, remove=F) %>% #count()
  mutate(diabetes_c=str_replace(diabetes_c,"(\\d+)_\\1", "\\1")) %>%
  mutate(diabetes_c=str_replace(diabetes_c,"NA_(\\d+)", "\\1")) %>% 
  mutate(diabetes_c=str_replace(diabetes_c,"(\\d+)_NA", "\\1")) %>% 
  mutate(diabetes_c=str_replace(diabetes_c,"(\\d+)_(\\d+)", "\\2")) %>% #count() #preferencia a _2
  mutate(diabetes_c=str_replace(diabetes_c,"NA", replacement = NA_character_)) %>%  #%>% count()
  mutate(diabetes_c= as.numeric(diabetes_c))
```

```{r}
sevdb_1 %>% 
  mutate(enf_snc= enf_snc_1==enf_snc_2) %>%
  dplyr::select(enf_snc, enf_snc_1, enf_snc_2) %>%
  tidyr::unite(enf_snc_c, enf_snc_1, enf_snc_2, remove=F) %>% #count()
  mutate(enf_snc_c=str_replace(enf_snc_c,"(\\d+)_\\1", "\\1")) %>%
  mutate(enf_snc_c=str_replace(enf_snc_c,"NA_(\\d+)", "\\1")) %>% 
  mutate(enf_snc_c=str_replace(enf_snc_c,"(\\d+)_NA", "\\1")) %>% 
  mutate(enf_snc_c=str_replace(enf_snc_c,"(\\d+)_(\\d+)", "\\2")) %>% #count() #preferencia a _2
  mutate(enf_snc_c=str_replace(enf_snc_c,"NA", replacement = NA_character_)) %>%  #%>% count()
  mutate(enf_snc_c= as.numeric(enf_snc_c))
```

```{r}
sevdb_1 %>% 
  mutate(enf_cardiac= enf_cardiac_1==enf_cardiac_2) %>%
  dplyr::select(enf_cardiac, enf_cardiac_1, enf_cardiac_2) %>%
  tidyr::unite(enf_cardiac_c, enf_cardiac_1, enf_cardiac_2, remove=F) %>% #count()
  mutate(enf_cardiac_c=str_replace(enf_cardiac_c,"(\\d+)_\\1", "\\1")) %>%
  mutate(enf_cardiac_c=str_replace(enf_cardiac_c,"NA_(\\d+)", "\\1")) %>% 
  mutate(enf_cardiac_c=str_replace(enf_cardiac_c,"(\\d+)_NA", "\\1")) %>% 
  mutate(enf_cardiac_c=str_replace(enf_cardiac_c,"(\\d+)_(\\d+)", "\\2")) %>% #count() #preferencia a _2
  mutate(enf_cardiac_c=str_replace(enf_cardiac_c,"NA", replacement = NA_character_)) %>%  #%>% count()
  mutate(enf_cardiac_c= as.numeric(enf_cardiac_c))
```

```{r}
sevdb_1 %>% 
  mutate(enf_pulm= enf_pulm_1==enf_pulm_2) %>%
  dplyr::select(enf_pulm, enf_pulm_1, enf_pulm_2) %>%
  tidyr::unite(enf_pulm_c, enf_pulm_1, enf_pulm_2, remove=F) %>% #count()
  mutate(enf_pulm_c=str_replace(enf_pulm_c,"(\\d+)_\\1", "\\1")) %>%
  mutate(enf_pulm_c=str_replace(enf_pulm_c,"NA_(\\d+)", "\\1")) %>% 
  mutate(enf_pulm_c=str_replace(enf_pulm_c,"(\\d+)_NA", "\\1")) %>% 
  mutate(enf_pulm_c=str_replace(enf_pulm_c,"(\\d+)_(\\d+)", "\\2")) %>% #count() #preferencia a _2
  mutate(enf_pulm_c=str_replace(enf_pulm_c,"NA", replacement = NA_character_)) %>%  #%>% count()
  mutate(enf_pulm_c= as.numeric(enf_pulm_c))
```

```{r}
sevdb_1 %>% 
  mutate(enf_renal= enf_renal_1==enf_renal_2) %>%
  dplyr::select(enf_renal, enf_renal_1, enf_renal_2) %>%
  tidyr::unite(enf_renal_c, enf_renal_1, enf_renal_2, remove=F) %>% #count()
  mutate(enf_renal_c=str_replace(enf_renal_c,"(\\d+)_\\1", "\\1")) %>%
  mutate(enf_renal_c=str_replace(enf_renal_c,"NA_(\\d+)", "\\1")) %>% 
  mutate(enf_renal_c=str_replace(enf_renal_c,"(\\d+)_NA", "\\1")) %>% 
  mutate(enf_renal_c=str_replace(enf_renal_c,"(\\d+)_(\\d+)", "\\2")) %>% #count() #preferencia a _2
  mutate(enf_renal_c=str_replace(enf_renal_c,"NA", replacement = NA_character_)) %>%  #%>% count()
  mutate(enf_renal_c= as.numeric(enf_renal_c))
```
```{r}
sevdb_1 %>% 
  mutate(tiempo_local= tiempo_local_1==tiempo_local_2) %>%
  dplyr::select(tiempo_local, tiempo_local_1, tiempo_local_2)%>%
  tidyr::unite(tiempo_local_c, tiempo_local_1, tiempo_local_2, remove = F) %>% #count()
  mutate(tiempo_local_c=str_replace(tiempo_local_c,"(\\d+)_\\1", "\\1")) %>%
  mutate(tiempo_local_c=str_replace(tiempo_local_c,"NA_(\\d+)", "\\1")) %>% 
  mutate(tiempo_local_c=str_replace(tiempo_local_c,"(\\d+)_NA", "\\1")) %>% 
  mutate(tiempo_local_c=str_replace(tiempo_local_c,"(\\d+)_(\\d+)", "\\2")) %>% #count() #preferencia a _2
  mutate(tiempo_local_c=str_replace(tiempo_local_c,"NA", replacement = NA_character_)) %>%  #%>% count()
  mutate(tiempo_local_c= as.numeric(tiempo_local_c))
```

```{r}
sevdb_1 %>% 
  mutate(malaria_pasado_num= malaria_pasado_num_1==malaria_pasado_num_2) %>%
  dplyr::select(malaria_pasado_num, malaria_pasado_num_1, malaria_pasado_num_2) %>%
  tidyr::unite(malaria_pasado_num_c, malaria_pasado_num_1, malaria_pasado_num_2, remove = F) %>% #count()
  mutate(malaria_pasado_num_c=str_replace(malaria_pasado_num_c,"(\\d+)_\\1", "\\1")) %>%
  mutate(malaria_pasado_num_c=str_replace(malaria_pasado_num_c,"NA_(\\d+)", "\\1")) %>% 
  mutate(malaria_pasado_num_c=str_replace(malaria_pasado_num_c,"(\\d+)_NA", "\\1")) %>% 
  mutate(malaria_pasado_num_c=str_replace(malaria_pasado_num_c,"(\\d+)_(\\d+)", "\\2")) %>% #count() #preferencia a _2
  mutate(malaria_pasado_num_c=str_replace(malaria_pasado_num_c,"NA", replacement = NA_character_)) %>%  #%>% count()
  mutate(malaria_pasado_num_c= as.numeric(malaria_pasado_num_c))
```
### Glimpse of all covariates

_Deben observarse lineas completas._
___Reducir el zoom__ de la página de ser necesario._

#### nov2014
```{r}
sevdb_pre %>% glimpse()
```
#### sep2016
```{r}
sevdb_pos %>% glimpse()
```

## Computer environment

```{r}
devtools::session_info()
```

```{r}
file.remove("s1.tex")
```


## References
