==> 05-sevrout.Rmd <==
---
title: "Covariates: table 1 + table 2"
author: "Andree Valle Campos"
date: '`r Sys.Date()`'
output:
  html_document:
  #pdf_document:
  # html_notebook:
    toc: yes
    toc_depth: 6
    toc_float:
      collapsed: yes
    #theme: united
    code_folding: "hide"
    #fig_caption: TRUE
    #number_sections: TRUE
bibliography: malaria.bib
link-citations: yes
#csl: american-medical-association.csl
editor_options: 
  chunk_output_type: console
params:
  spanish: false
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE, #fig.path = "01-",
                      warning = FALSE,message = FALSE)
# knitr::opts_knit$set(root.dir = '../.')
options(width = 90) # expand limits of CONSOLE output
```

## To-do

- (x) tabla 1 - copy to google sheets
- (o) tabla 2 - copy to google sheets

## Packages

```{r}
library(tidyverse)
library(compareGroups)
library(ggupset)
library(patchwork)
library(dagitty)
library(ggdag)
```

## Data input

```{r}
sevdb_2ws <- read_rds("data/03-sevrcov-228obs_21var.rds") %>% janitor::clean_names()
# sevdb_2ws %>% glimpse()
```

```{r,eval=FALSE}
sevdb_3sv <- read_rds("data/03-sevrcov.rds") %>% janitor::clean_names()
# sevdb_3sv %>% glimpse()
```

## Select variables

```{r,warning=FALSE,message=FALSE}
sevdb_2ws_notranslation <- sevdb_2ws %>% 
  mutate(edad_4=Hmisc::cut2(edad,g = 4),
         enf_sum=as.factor(enf_sum)) %>% 
  cdcper::cdc_edades_peru(variable_edad = edad) %>% 
  mutate(episodio_previo_cat = Hmisc::cut2(episodio_previo_num,g = 4)) %>% 
  #start select
  select(-(edad_grupo_x:edad_inei_grupos_labels),
         -edad_etapas_de_vida_t,
         -(sev_who_num_1:sev_who_num_4),
         -(sev_who_1:sev_who_4),
         -shock_oms_pa) %>% 
  select(starts_with("edad"),
         sexo:enf_sum,
         tx_par_intest,
         parasitemia_mil,parasitemia_mil_4,
         en_zona_endemica,en_zona_endemica_4,
         episodio_previo_num,
         episodio_previo_cat,
         episodio_previo,
         everything()) %>% 
  select(-parasitemia) %>% 
  select(starts_with("edad"),sexo,
         glas_oms,convulsion,
         shock_oms,pulmonar_oms,
         renal_oms,hipoglic_oms,
         anemia_oms,hiperbil_oms,
         everything()) %>% 
  select(-tx_par_intest) %>% 
  select(-sangr_oms,-trombo_sev,-presenta_comorb,-casecontrol) %>% 
  mutate(edad_etapas_de_vida_c=fct_recode(edad_etapas_de_vida_c,
                                          "0-11"="0_11a",
                                          "12-17"="12_17a",
                                          "18-29"="18_29a",
                                          "30-59"="30_59a",
                                          "60+"="60a_mas"))
```

```{r}
sevdb_2ws_clean <- sevdb_2ws_notranslation %>% 
  # per language
  # add variable labels
  labelled::set_variable_labels(
    edad="Age (years)",
    edad_etapas_de_vida_c="Age groups",
    sexo="Sex",
    shock_oms="Shock (systolic blood pressure < 80 mmHg)",
    glas_oms="Coma (Glasgow score <= 9/14)",
    pulmonar_oms="Lung injury (ARDS or pulmonary edema)",
    anemia_oms="Severe anemia (hemoglobin < 7 g/dL)",
    renal_oms="Acute renal failure (creatinine > 3 mg/dL)",
    hiperbil_oms="Hyperbilirubinemia (bilirubin > 2.5 mg/dL)", #Jaundice
    hipoglic_oms="Hypoglycemia (glucose < 40 mg/dL)",
    convulsion="Convulsion",
    # presenta_comorb="Comorbidity",
    enf_cronica="Chronic disease",
    enf_sum = "Chronic diseases (number)",
    parasitemia_mil="Parasite density (x 1000 par/uL)",
    en_zona_endemica="Time in endemic area (years)",
    episodio_previo_num="Previous episodes (number)",
    episodio_previo_cat="Previous episodes (range)",
    sev_who_num="Severe malaria criteria (number)",
    sev_who="Severe malaria",
    study="Protein microarray study",
    episodio_previo="Previous episodes") %>% 
  # # add value levels
  # # problem: do not work on factor variables
  # # additionally: only variable labels work, not variable values
  # labelled::set_value_labels(
  #   study = c("No"="whole-db","Yes"="in-microarray")
  # ) %>% 
  mutate(across(.cols = c(glas_oms:hiperbil_oms,
                          # presenta_comorb,
                          enf_cronica),
                .fns = ~fct_recode(.,"No"="0","Yes"="1"))) %>% 
  mutate(episodio_previo=fct_recode(episodio_previo,
                                    "Without"="sin",
                                    "With"="con")) %>% 
  mutate(sexo=fct_recode(sexo,
                         "Male"="masculino",
                         "Female"="femenino")) %>% 
  mutate(sev_who=fct_recode(sev_who,
                         "No"="no_severo",
                         "Yes"="severo")) #%>% 
  # mutate(study=fct_recode(study,
  #                        "No"="whole-db",
  #                        "Yes"="in-microarray"))
```

```{r,eval=params$spanish}
sevdb_2ws_clean <- sevdb_2ws_notranslation %>% 
  # per language
  # add labels
  labelled::set_variable_labels(
    edad="Edad (años)",
    edad_etapas_de_vida_c="Edad (grupos)",
    sexo="Sexo",
    shock_oms="Shock (presión arterial sistólica < 80 mmHg)",
    glas_oms="Coma (puntaje Glasgow <= 9/14)",
    pulmonar_oms="Lesión pulmonar (SDRA o edema pulmonar)",
    anemia_oms="Anemia grave (hemoglobina < 7 g/dL)",
    renal_oms="Insuficiencia renal aguda (creatinina > 3 mg/dL)",
    hiperbil_oms="Hiperbilirrubinemia (bilirrubina > 2,5 mg/dL)", #Jaundice
    hipoglic_oms="Hipoglicemia (glucosa < 40mg/dL)",
    convulsion="Convulsión",
    #presenta_comorb="Comorbilidad",
    enf_cronica="Enfermedad crónica",
    enf_sum="Enfermedades crónicas (número)",
    parasitemia_mil="Densidad de parásitos (x 1000 par/uL)",
    en_zona_endemica="Tiempo en zona endémica (años)",
    episodio_previo_num="Episodios anteriores (número)",
    episodio_previo_cat="Episodios anteriores (rangos)",
    sev_who_num="Criterios de Malaria severa (número)",
    sev_who="Malaria severa",
    study="Estudio con microarreglos de proteinas",
    episodio_previo="Episodios anteriores") %>% 
  # replace levels
  mutate(across(.cols = c(glas_oms:hiperbil_oms,
                          # presenta_comorb,
                          enf_cronica),
                .fns = ~fct_recode(.,"No"="0","Sí"="1"))) %>% 
  mutate(episodio_previo=fct_recode(episodio_previo,
                                    "Sin"="sin",
                                    "Con"="con")) %>% 
  mutate(sexo=fct_recode(sexo,
                         "Masculino"="masculino",
                         "Femenino"="femenino")) %>% 
  mutate(sev_who=fct_recode(sev_who,
                         "No"="no_severo",
                         "Sí"="severo")) #%>% 
  # mutate(study=fct_recode(study,
  #                        "No"="whole-db",
  #                        "Sí"="in-microarray"))
```

```{r}
sevdb_2ws_clean %>% glimpse()
```


## Table 1 - all covariates

```{r}
tab01all <- sevdb_2ws_clean %>% 
  # extract edad quartiles
  select(-edad_4) %>% 
  select(-parasitemia_mil_4,
         -en_zona_endemica_4) %>% 
  compareGroups(study~.,data = .,
                method = c(edad=2,
                           en_zona_endemica=2,
                           episodio_previo_num=2,
                           parasitemia_mil=2)) %>% 
  createTable(digits = 1,
              sd.type = 2,
              show.n = T,
              show.all = T)

tab01all %>% export2md() 
tab01all %>% export2xls("table/05-tab01-all_covariates.xls")
  
```

## Table 1 - oms criteria

```{r}
tab01oms <- sevdb_2ws_clean %>% 
  # extract edad quartiles
  select(-edad_4) %>% 
  select(-parasitemia_mil_4,
         -en_zona_endemica_4) %>% 
  # extract variables
  select(glas_oms:hiperbil_oms,
         sev_who_num,
         study) %>% 
  compareGroups(study~.,data = .,
                chisq.test.perm = TRUE,
                # method = c(edad=2,
                #            en_zona_endemica=2,
                #            episodio_previo_num=2,
                #            parasitemia_mil=2)
                ) %>% 
  createTable(digits = 1,
              sd.type = 2,
              show.n = T,
              show.all = T)

tab01oms %>% export2md() 
tab01oms %>% export2xls("table/05-tab01-oms_covariates.xls")
```

## Table 2 - all observations

```{r}
tab02all <- sevdb_2ws_clean %>% 
  # extract edad quartiles
  select(-edad_4) %>% 
  select(-parasitemia_mil_4,
         -en_zona_endemica_4) %>% 
  # extract variables
  select(-(glas_oms:hiperbil_oms),
         # -(sangr_oms:trombo_sev),
         # -enf_sum,
         -sev_who_num,
         # -casecontrol,
         -study) %>% 
  compareGroups(sev_who~.,data = .,
                byrow = F,
                method = c(edad=2,
                           en_zona_endemica=2,
                           episodio_previo_num=2,
                           parasitemia_mil=2)) %>% 
  createTable(digits = 1,
              sd.type = 2,
              show.n = T,
              show.all = T)

tab02all %>% export2md() 
tab02all %>% export2xls("table/05-tab02-all_observations.xls")
  
```

### conditional association

```{r,eval=FALSE,echo=FALSE}
dagify(y ~ x,
       y ~ conf + x,
       x ~ conf,
       labels = c("y" = "cáncer piel", 
                  "x" = "exp. solar",
                  "conf" = "café"
       ),
       exposure = "x",
       outcome = "y") %>% 
  tidy_dagitty() %>% 
  ggdag_adjustment_set()
```


```{r}

# episode [pos="0.363,0.464"]
# severe [pos="0.575,0.460"]
# age [pos="0.461,0.157"]
# sex [pos="0.462,0.286"]
# time_endemic [pos="0.311,0.250"]
# chronic [pos="0.618,0.244"]

dag_loc <- tribble(
  ~name,~x,~y,
  "episode",0.363,0.464,
  "severe",0.575,0.460,
  "age",0.461,0.157+0.45,
  "sex",0.462,0.286+0.25,
  "time_endemic",0.311,0.250+0.3,
  "chronic",0.618,0.244+0.3
  ) %>% 
  mutate(to=name,
         xend=x,
         yend=y)

dag <- dagitty("dag {
    chronic -> severe
    age -> chronic
    age -> episode
    age -> time_endemic
    age -> severe
    episode -> severe
    time_endemic -> episode
    sex -> episode
    sex -> severe
}")

dag

tidy_dag <- tidy_dagitty(dag)

tidy_dag$data <- tidy_dag$data %>% 
  select(-x,-y,-xend,-yend) %>% 
  left_join(dag_loc %>% select(name:y)) %>% 
  left_join(dag_loc %>% select(to:yend)) %>% 
  select(name,x,y,direction,to,xend,yend,circular) %>% 
  mutate(label=name)

# tidy_dag
```

```{r}
dag_gen_one <- 
  ggdag(tidy_dag,
      use_labels = "name",
      text = FALSE) +
  theme_dag()
# dag_gen_one
ggsave("figure/05-fig08-severe_compare-dag_only-epi.png",
       plot = dag_gen_one,
       height = 6,width = 6,dpi = "retina")
```

```{r}
dag_gen_two <- 
  ggdag_adjustment_set(tidy_dag,
                       exposure = "episode",
                       outcome = "severe",
                       use_labels = "name",
                       text = FALSE,
                       node_size = 14) + 
  theme_dag() +
  theme(legend.position = "bottom")
# dag_gen_two
ggsave("figure/05-fig09-severe_compare-dag_adjust-epi.png",
       plot = dag_gen_two,
       height = 6,width = 6,dpi = "retina")
```

```{r,eval=FALSE}
ggdag_paths(tidy_dag,
            use_labels = "name",
            from = "episode",
            to = "severe",
            text = FALSE,
            adjust_for = c("age","sex")) +
  theme_dag()
```


```{r}
dag_gen_both <- dag_gen_one +
  dag_gen_two +
  # plot_layout(ncol = 2,nrow = 2) +
  plot_annotation(tag_levels = "A")
dag_gen_both
ggsave("figure/05-fig07-severe_compare-dag_both-epi.png",
       plot = dag_gen_both,
       height = 5,width = 10,dpi = "retina")
```

### multiple linear model

- multiple logistic regression
- estimate the prevalence ratio 
of severe vivax malaria cases 
with respect to 
patients with and without previous episodes,
using a
minimum sufficient adjustment set of variables
(age and sex)
identified as confounders

```{r}

sevdb_2ws_clean_glm <- sevdb_2ws_clean %>%
  # select(episodio_previo,sev_who,sexo,edad_4,
  #        enf_cronica,en_zona_endemica) %>%
  # count(sev_who)
  mutate(sev_who_cat = as.numeric(sev_who)-1,
         # sev_who_cat = as.factor(sev_who_cat)
         ) #%>% 
  # filter(!is.na(episodio_previo))

sevdb_2ws_clean_glm %>% 
  # naniar::vis_miss()
  count(sev_who,sev_who_cat,episodio_previo)

sevdb_2ws_clean_glm %>% 
  # naniar::vis_miss()
  count(sev_who,sev_who_cat,episodio_previo_cat)

# glm(sev_who_cat ~ episodio_previo,
#     data = sevdb_2ws_clean_glm,
#     family = binomial(link = "log"),
#     na.action = na.exclude) %>%
#   epitidy::epi_tidymodel_pr()

model_epi_sev_01 <- 
  glm(sev_who_cat ~ episodio_previo + sexo + edad_4,
    data = sevdb_2ws_clean_glm,
    family = poisson(link = "log"),
    na.action = na.exclude) %>%
  epitidy::epi_tidymodel_pr() %>% 
  # potential function to epitidy
  # epitidy::epi_tidyoutput(
  # data,variable_dot,variable_low,variable_upp,variable_pval,
  # digits_dot = 2,digits_low = 2,digits_upp = 2,digits_pval=3)
  select(-log.pr,-se) %>% 
  # round numeric values
  mutate_at(.vars = vars(pr,conf.low,conf.high),
            .funs = round, digits=2) %>%
  mutate_at(.vars = vars(p.value),
            .funs = round, digits=3) %>%
  #join confidence intervals
  mutate(ci=str_c(conf.low," - ",conf.high)) %>% 
  select(-conf.low,-conf.high) %>% 
  relocate(p.value,.after = last_col())

model_epi_sev_01
model_epi_sev_01 %>% 
  writexl::write_xlsx("table/05-tab04-severe_episode_dic-multiple.xlsx")

model_epi_sev_02 <- 
  glm(sev_who_cat ~ episodio_previo_cat + sexo + edad_4,
    data = sevdb_2ws_clean_glm,
    family = poisson(link = "log"),
    na.action = na.exclude) %>%
  epitidy::epi_tidymodel_pr() %>% 
  select(-log.pr,-se) %>% 
  # round numeric values
  mutate_at(.vars = vars(pr,conf.low,conf.high),
            .funs = round, digits=2) %>%
  mutate_at(.vars = vars(p.value),
            .funs = round, digits=3) %>%
  #join confidence intervals
  mutate(ci=str_c(conf.low," - ",conf.high)) %>% 
  select(-conf.low,-conf.high) %>% 
  relocate(p.value,.after = last_col())

model_epi_sev_02
model_epi_sev_02 %>% 
  writexl::write_xlsx("table/05-tab04-severe_episode_cat-multiple.xlsx")

# glm(sev_who_cat ~ 
#       episodio_previo + sexo + edad_4 + 
#       enf_cronica + en_zona_endemica,
#     data = sevdb_2ws_clean_glm,
#     family = poisson(link = "log"),
#     na.action = na.exclude) %>%
#   epitidy::epi_tidymodel_pr()
```


```{r,eval=FALSE,echo=FALSE}
library(tidymodels)

# poisson_reg()

lm_model <- linear_reg() %>% 
  set_engine(engine = "lm") %>% 
  set_mode(mode = "regression")

lm_model %>% 
  fit(sev_who ~ episodio_previo, data = sevdb_2ws_clean)

sevdb_2ws_clean %>%
  select(episodio_previo,sev_who,sexo,edad_4,sev_who_num) %>% 
  count(sev_who_num)


```


## Table 2 - microarray study

```{r}
tab02abn <- sevdb_2ws_clean %>% 
  # filter observations
  filter(study=="in-microarray") %>% 
  # extract edad quartiles
  select(-edad_4) %>% 
  select(-parasitemia_mil_4,
         -en_zona_endemica_4) %>% 
  # extract variables
  select(-(glas_oms:hiperbil_oms),
         # -(sangr_oms:trombo_sev),
         # -enf_sum,
         -sev_who_num,
         # -casecontrol,
         -study) %>% 
  compareGroups(sev_who~.,data = .,
                byrow = F,
                method = c(edad=2,
                           en_zona_endemica=2,
                           episodio_previo_num=2,
                           parasitemia_mil=2)) %>% 
  createTable(digits = 1,
              sd.type = 2,
              show.n = T,
              show.all = T)

tab02abn %>% export2md() 
tab02abn %>% export2xls("table/05-tab02-abn_observations.xls")
  
```

### conditional association

```{r}
# dag {
# bb="-1.519,-1.953,1.998,1.641"
# age [pos="-0.454,-0.966"]
# antibody [outcome,pos="0.656,-0.124"]
# chronic [pos="-0.646,-0.426"]
# episode [adjusted,pos="0.199,-0.600"]
# severe [exposure,pos="-0.197,-0.124"]
# sex [pos="-0.111,-1.239"]
# time_endemic [pos="0.301,-1.198"]
# age -> chronic
# age -> episode
# age -> severe
# age -> time_endemic
# chronic -> severe
# episode -> antibody
# episode -> severe
# severe -> antibody
# sex -> episode
# sex -> severe
# time_endemic -> episode
# }


dag_loc <- tribble(
  ~name,~x,~y,
  "age",-0.454,-0.966+1.3,
  "antibody",0.656,-0.124,
  "chronic",-0.646,-0.426+0.45,
  "episode",0.199,-0.600+0.8,
  "severe",-0.197,-0.124,
  "sex",-0.111,-1.239+1.8,
  "time_endemic",0.301+0.2,-1.198+1.67
  ) %>% 
  mutate(to=name,
         xend=x,
         yend=y)

# dag_loc

dag <- dagitty("dag {
    severe [exposure]
    episode [adjusted]
    antibody [outcome]
    age -> chronic
    age -> episode
    age -> severe
    age -> time_endemic
    chronic -> severe
    episode -> antibody
    episode -> severe
    severe -> antibody
    sex -> episode
    sex -> severe
    time_endemic -> episode
}")

dag

tidy_dag <- tidy_dagitty(dag)

tidy_dag$data <- tidy_dag$data %>% 
  select(-x,-y,-xend,-yend) %>% 
  left_join(dag_loc %>% select(name:y)) %>% 
  left_join(dag_loc %>% select(to:yend)) %>% 
  select(name,x,y,direction,to,xend,yend,circular) %>% 
  mutate(label=name)

# tidy_dag
```

```{r}
dag_only <- 
  ggdag(tidy_dag,
      use_labels = "name",
      text = FALSE) +
  theme_dag()
# dag_only
ggsave("figure/05-fig04-severe_compare-dag.png",
       plot = dag_only,
       height = 6,width = 6,dpi = "retina")
```

```{r}
dag_adjust <- 
  ggdag_adjustment_set(tidy_dag,
                       use_labels = "name",
                       text = FALSE,
                       node_size = 14) + 
  theme_dag() +
  theme(legend.position = "bottom")
# dag_adjust
ggsave("figure/05-fig05-severe_compare-dag_adjust.png",
       plot = dag_adjust,
       height = 6,width = 6,dpi = "retina")
```

```{r}
dag_both <- dag_only +
  dag_adjust +
  # plot_layout(ncol = 2,nrow = 2) +
  plot_annotation(tag_levels = "A")
dag_both
ggsave("figure/05-fig06-severe_compare-dag_both.png",
       plot = dag_both,
       height = 5,width = 10,dpi = "retina")
```

### multiple linear regression

- multiple linear regression
- estimate the average change 
of antibody reactivity
with respect to 
severe and non-severe vivax malaria patients,
using a
minimum sufficient adjustment set of variables
(previous episodes)
identified as confounders
- results in `12-absevere.Rmd`

## Table 3 - microarray study and oms criteria

```{r}
tab03oms <- sevdb_2ws_clean %>% 
  # filter observations
  filter(study=="in-microarray") %>% 
  # extract edad quartiles
  select(-edad_4) %>% 
  select(-parasitemia_mil_4,
         -en_zona_endemica_4) %>% 
  # extract variables
  select(glas_oms:hiperbil_oms,
         sev_who_num,
         # study,
         sev_who) %>% 
  compareGroups(sev_who~.,data = .,
                byrow = F,
                # method = c(edad=2,
                #            en_zona_endemica=2,
                #            episodio_previo_num=2,
                #            parasitemia_mil=2)
                ) %>% 
  createTable(digits = 1,
              sd.type = 2,
              show.n = T,
              show.all = T)

tab03oms %>% export2md() 
tab03oms %>% export2xls("table/05-tab03-abn_oms_covariates.xls")
  
```



<!-- ```{r} -->
<!-- sevdb_2ws_clean %>%  -->
<!--   # extract edad quartiles -->
<!--   select(study,sev_who) %>%  -->
<!--   compareGroups(study~.,data = ., -->
<!--                 byrow = T) %>%  -->
<!--   createTable(digits = 1, -->
<!--               sd.type = 2, -->
<!--               show.n = T, -->
<!--               show.all = T) -->
<!-- ``` -->


## Figures

```{r}
sevdb_2ws_clean %>% 
  ggplot(aes(x = edad)) +
  geom_histogram(binwidth = 1)

sevdb_2ws_clean %>% 
  ggplot(aes(x = parasitemia_mil)) +
  geom_histogram(binwidth = 1)

sevdb_2ws_clean %>% 
  ggplot(aes(x = en_zona_endemica)) +
  geom_histogram(binwidth = 1)

sevdb_2ws_clean %>% 
  ggplot(aes(x = episodio_previo_num)) +
  geom_histogram(binwidth = 1)
```


```{r}
sevdb_2ws_clean %>% 
  filter(!is.na(edad)) %>%
  ggplot(aes(x = edad_4)) +
  geom_bar(aes(fill=sev_who),position = position_fill()) + 
  theme(axis.text.x = element_text(angle = 45, hjust = 1))

sevdb_2ws_clean %>% 
  # filter(!is.na(edad)) %>%
  ggplot(aes(x = parasitemia_mil_4)) +
  geom_bar(aes(fill=sev_who),position = position_fill()) + 
  theme(axis.text.x = element_text(angle = 45, hjust = 1))

sevdb_2ws_clean %>% 
  # filter(!is.na(edad)) %>%
  ggplot(aes(x = en_zona_endemica_4)) +
  geom_bar(aes(fill=sev_who),position = position_fill()) + 
  theme(axis.text.x = element_text(angle = 45, hjust = 1))
```

```{r}
sevdb_2ws_clean %>% 
  filter(!is.na(edad)) %>%
  count(edad_4,sev_who) %>% 
  pivot_wider(names_from = sev_who,values_from = n) %>% 
  rowwise() %>% 
  mutate(denominator=sum(c_across(cols = -edad_4))) %>% 
  ungroup() %>% 
  mutate(raw=pmap(.l = select(.,x=3,n=denominator), #cuidado
                    .f = binom.test),
           raw=map(.x = raw,.f = broom::tidy)) %>%
    unnest(raw) %>% 
  select(1:5,starts_with("conf")) %>% 
  ggplot(aes(x = edad_4,y = estimate)) +
  geom_point() +
  geom_errorbar(aes(max=conf.high, min = conf.low)) +
  ylim(0,1)
```


### upset plot

#### all observations

esta sección no se agrega a HTML automáticamente. generar en consola.

```{r,eval=FALSE}
upset_all <- sevdb_2ws_clean %>% 
  # filter observations
  # filter(study=="in-microarray") %>% 
  # extract edad quartiles
  select(-edad_4) %>% 
  select(-parasitemia_mil_4,
         -en_zona_endemica_4) %>% 
  # extract variables
  select(codigo,glas_oms:hiperbil_oms) %>% 
  pivot_longer(cols = -codigo,
               names_to = "symptom",
               values_to = "member") %>% 
  filter(member=="Sí") %>%
  select(-member) %>% 
  select(codigo,symptom) %>% 
  arrange(codigo) %>% 
  #create list of pathways per gene
  nest(cols=symptom) %>% 
  #the list should be a character not a dataframe, 
  #so pull them out
  mutate(pathways_c = map(cols,pull)) %>%
  #plot it
  ggplot(aes(x = pathways_c)) +
  geom_bar() +
  geom_text(stat='count', 
            aes(label=after_stat(count)), 
            vjust=-1) +
  scale_x_upset() +
  scale_y_continuous(lim = c(0, 30)) +
  labs(x=NULL,y="Frecuency")
upset_all
ggsave("figure/05-fig01-severe_symptoms-upset.png",
       plot = upset_all,
       height = 6,width = 6,dpi = "retina")
```

#### in-microarray

```{r,eval=FALSE}
upset_micro <- sevdb_2ws_clean %>% 
  # filter observations
  filter(study=="in-microarray") %>% 
  # extract edad quartiles
  select(-edad_4) %>% 
  select(-parasitemia_mil_4,
         -en_zona_endemica_4) %>% 
  # extract variables
  select(codigo,glas_oms:hiperbil_oms) %>% 
  pivot_longer(cols = -codigo,
               names_to = "symptom",
               values_to = "member") %>% 
  filter(member=="Sí") %>%
  select(-member) %>% 
  select(codigo,symptom) %>% 
  arrange(codigo) %>% 
  #create list of pathways per gene
  nest(cols=symptom) %>% 
  #the list should be a character not a dataframe, 
  #so pull them out
  mutate(pathways_c = map(cols,pull)) %>%
  #plot it
  ggplot(aes(x = pathways_c)) +
  geom_bar() +
  geom_text(stat='count', 
            aes(label=after_stat(count)), 
            vjust=-1) +
  scale_x_upset() +
  scale_y_continuous(lim = c(0, 7)) +
  labs(x=NULL,y="Frecuency")
upset_micro
ggsave("figure/05-fig02-severe_symptoms-upset-microarray.png",
       plot = upset_micro,
       height = 6,width = 6,dpi = "retina")
```

#### both

```{r,eval=FALSE}
upset_both <- upset_all +
  upset_micro +
  # plot_layout(ncol = 2,nrow = 2) +
  plot_annotation(tag_levels = "A")
ggsave("figure/05-fig03-severe_symptoms-upset-sample_microarray.png",
       plot = upset_both,
       height = 3.5,width = 7,dpi = "retina")
```

```{r}
knitr::include_graphics("figure/05-fig03-severe_symptoms-upset-sample_microarray.png")
```


## Additional 

```{r}
read.csv("data-raw/RawData.csv") %>% 
  as_tibble() %>% 
  pivot_longer(cols = -(Index:Description),
               names_to = "samples",
               values_to = "array_read") %>% 
  select(samples) %>% 
  distinct() %>% 
  mutate(precode=str_replace(samples,"(...).+","\\1")) %>% 
  dplyr::count(precode)
```

```{r}
readr::read_csv("data-raw/ADi-NAMRU6_Data-samples.csv") %>% 
  # glimpse()
  dplyr::count(Study,Sample.Type)

readr::read_csv("data-raw/ADi-NAMRU6_Data-samples.csv") %>% 
  filter(Study=="Controls")

readr::read_csv("data-raw/samples.csv") %>% 
  dplyr::count(Sample.Type)
```
