==> 06-uscntrl.Rmd <==
---
title: "non-exposed US control samples - GCRC"
output: html_document
editor_options: 
  chunk_output_type: console
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
```

```{r}
library(tidyverse)
library(readxl)
library(Biobase)
library(limma)
library(patchwork)
```

## import useful data

```{r}
all_listgen <- readr::read_rds("data/04-listgen-microarray_chip-clean.rds")
```

## data cleaning

negative controls are non-malaria infected individuals from the USA controls (GCRC, General Clinical Research Center)

### extract data from excel

```{r}
rute <- "data-raw/Microarray results - First analysis wDNAfree controls.xlsx"
data_uscntrl <- read_xlsx(path = rute) %>% 
  janitor::clean_names() %>% 
  select(x1:x6,unexposed_controls:x56) %>% 
  filter(!is.na(x56)) %>% 
  filter(!is.na(x1)) %>% 
  rename(x50=unexposed_controls) %>% 
  mutate(across(.cols = x50:x56,.fns = as.numeric)) %>% 
  # tail(n=100)
  mutate(x6=snakecase::to_snake_case(x6)) %>%
  filter(str_starts(string = x6,
                    pattern = "PF|pf|PV|pv|NODNA|nodna")) %>%
  select(-(x1:x5)) %>% 
  mutate(x6=janitor::make_clean_names(x6))
  #epihelper::print_inf()
```

### transform to dataframe format

```{r}
anti <- data_uscntrl %>% 
  filter(str_starts(string = x6,
                    pattern = "PF|pf|PV|pv")) %>% 
  column_to_rownames(var = "x6")
```

### median of array-control

```{r}
ctrl <- data_uscntrl %>% 
  filter(str_starts(string = x6,pattern = "nodna")) %>% 
  summarise_if(is.numeric,funs(median)) %>% 
  # MEDIAN to NORMALIZE against noDNA controls
  mutate(x6="nodna") %>% 
  column_to_rownames(var = "x6")
```

### crear normalization array

```{r}
ctrm <- NULL # crear objeto vacío
# loop para generar una matriz de las dimensiones de `anti`
for(i in 1:nrow(anti)){
  ctrm <- rbind(ctrm,ctrl)
}
# dimensiones de la matriz con la mediana de ctrls `noDNA` creada
dim(ctrm)
dim(anti)
# matriz con la mediana de ctrls `noDNA` para las 8 primeras muestras
# ctrm[1:8,] #%>% class()
```

### aplicar normalización

```{r}
# definir resultado para los valores menores o iguales a cero:
# log2.NA = function(x) {log2(ifelse(x>0, x, NA))}

# normalización con respecto a los controles `noDNA` 
# c/ transformación log2
norm <- log2(anti/ctrm) %>% 
  rownames_to_column(var = "x6") %>% 
  as_tibble() %>% 
  mutate(across(.cols = x50:x56,
                .fns = ~if_else(condition = .x==-Inf,
                               true = 0,
                                # true = NA_real_,
                               false = .x))) %>% 
  mutate(x6specie=case_when(
    str_starts(x6,"pf") ~ "pf",
    str_starts(x6,"pv") ~ "pv",
    TRUE ~ NA_character_
  )) #%>% 
  #count(x6specie)
```

### visualizar distribución

```{r}
norm %>% 
  pivot_longer(cols = x50:x56,
               names_to = "usctrl",
               values_to = "usvalue") %>% 
  ggplot(aes(x = usvalue)) +
  geom_histogram()
```

```{r}
norm %>% 
  pivot_longer(cols = x50:x56,
               names_to = "usctrl",
               values_to = "usvalue") %>% 
  ggplot(aes(x = usvalue)) +
  geom_histogram() +
  facet_grid(~x6specie)
```

```{r,eval=FALSE}
norm %>% 
  filter(x6specie=="pv") %>% 
  pivot_longer(cols = x50:x56,
               names_to = "usctrl",
               values_to = "usvalue") %>% 
  ggplot(aes(x = usvalue,y = x6)) +
  geom_point()
```

## crear unico assay data (pvivax+ y us-control)

```{r}
usctrlpv <- norm %>% 
  filter(x6specie=="pv") %>% 
  filter(!str_detect(x6,"pvivax")) %>% 
  mutate(x6 = str_to_upper(x6)) %>% 
  mutate(x6=str_replace_all(x6,"_O_","o")) %>% 
  mutate(x6=str_replace_all(x6,"_S_","_S")) %>% 
  select(-x6specie) %>% 
  mutate(x6=case_when(
    str_detect(x6,"PVX_100680") ~ "PVX_100680_1o1",
    str_detect(x6,"PVX_003825_RENAMED") ~ "PVX_003825",
    str_detect(x6,"PVX_083240_RENAMED") ~ "PVX_083240",
    str_detect(x6,"PVX_085915_RENAMED") ~ "PVX_085915 (renamed)",
    str_detect(x6,"PVX_095185_RENAMED") ~ "PVX_095185",
    str_detect(x6,"PVX_097720_RENAMED") ~ "PVX_097720",
    str_detect(x6,"PVX_082700_RENAMED_0") ~ "PVX_082700 (renamed0",
    TRUE ~ x6
  ))
```

### importar eset vivax

```{r}
eset.VIVAX.SEV_import <- 
  readr::read_rds("data/04-eset_vivax_sev.rds")
eset.VIVAX.SEV.FILTER_import <- 
  readr::read_rds("data/04-eset_vivax_sev_filter.rds")
```

### extraer assay data

```{r}
preanalysis <- eset.VIVAX.SEV.FILTER_import@assayData$exprs %>% 
  as_tibble(rownames = "feature") %>% 
  mutate(num = str_replace(feature,"(.+)\\.(.+)","\\2"),
         orf = str_replace(feature,"(.+)\\.(.+)","\\1")) %>% 
  select(feature,orf,num,everything())
```

### unir us-control data

```{r}
expression_set <- preanalysis %>% 
  left_join(usctrlpv,by = c("orf" = "x6")) %>% 
  select(-orf,-num)
  # glimpse()
  # naniar::vis_miss()
  # filter(is.na(x50)) %>% 
  # select(feature:num) %>% 
  # arrange(orf)
```

```{r,eval=FALSE}
expression_set %>% 
  filter(str_detect(feature,"123520"))
preanalysis %>% 
  filter(str_detect(feature,"123520"))
usctrlpv %>% 
  filter(str_detect(x6,"123520"))
# filter(feature=="PVX_123520.241")
```


### control de calidad

```{r}
# usctrlpv %>% 
#   filter(str_detect(x6,"PVX_082700")) %>% 
#   arrange(x6) %>% 
#   mutate(x6=case_when(
#     # str_detect(x6,"PVX_100680") ~ "PVX_100680_1o1",
#     # str_detect(x6,"PVX_003825_RENAMED") ~ "PVX_003825",
#     # str_detect(x6,"PVX_083240_RENAMED") ~ "PVX_083240",
#     # str_detect(x6,"PVX_085915_RENAMED") ~ "PVX_085915 (renamed)",
#     # str_detect(x6,"PVX_095185_RENAMED") ~ "PVX_095185",
#     # str_detect(x6,"PVX_097720_RENAMED") ~ "PVX_097720",
#     # str_detect(x6,"PVX_082700_RENAMED_0") ~ "PVX_082700 (renamed0",
#     TRUE ~ x6
#   ))
# 
# preanalysis %>% 
#   filter(str_detect(orf,"PVX_082700")) %>% 
#   select(feature:num) %>% 
#   arrange(orf)
```

### transformar a long

```{r}
expression_set_long <- expression_set %>% 
  pivot_longer(cols = -feature,
               names_to = "samples",
               values_to = "values") %>% 
  mutate(sample_type=case_when(
    str_starts(samples,"LIM") ~ "sample",
    str_starts(samples,"x") ~ "control",
    TRUE ~ NA_character_
  ))

expression_pheno <- expression_set %>% 
  select(-feature) %>% 
  colnames() %>% 
  enframe(name = NULL,value = "samples") %>% 
  mutate(sample_type=case_when(
    str_starts(samples,"LIM") ~ "sample",
    str_starts(samples,"x") ~ "control",
    TRUE ~ NA_character_
  ))
```

### visualizar normalización

```{r}
expression_set_long %>% 
  ggplot(aes(x = values, fill = sample_type)) +
  geom_histogram(position = position_identity(),alpha = 0.5)
```

```{r}
expression_set_long %>% 
  ggplot(aes(x = values, y = feature, color = sample_type)) +
  geom_point(alpha = 0.5)
```

## build experessionSet

### assay data

```{r}
norx <- expression_set_long %>% # RULE TO EDIT a value of a column!
  select(rowname=feature,
         sample_name=samples,
         expression=values) %>% 
  #dplyr::filter(rowname=="PVX_111175.544",sample_name=="LIM2017") # 21
  reshape2::acast(rowname ~ sample_name,
                  value.var = "expression") #%>% class()
```

### pheno data

```{r}
pheno_d <- new("AnnotatedDataFrame", 
               data = expression_pheno %>% 
                 as.data.frame() %>% 
                 column_to_rownames(var="samples"))
```

### feature data

```{r}
expression_feature <- expression_set %>% 
  select(feature) %>% 
  mutate(feature_specie=case_when(
    str_starts(feature,"PF|pf") ~ "pf",
    str_starts(feature,"pv|PV") ~ "pv",
    TRUE ~ NA_character_
  )) %>% 
  as.data.frame() %>% 
  column_to_rownames(var="feature")

feat_d <- new("AnnotatedDataFrame", data=expression_feature)
```

### create eset

```{r}
eset <- Biobase::ExpressionSet(
  assayData = norx[feat_d %>% rownames(),], 
  phenoData = pheno_d,
  featureData = feat_d)
eset
```

## differential expression analysis

```{r}
eset.VIVAX.SEV.FILTER <- eset
```

#### 5.2 previous episode

```{r}
eset <- eset.VIVAX.SEV.FILTER
pData(eset) %>% glimpse()
```

##### Differential expression

###### Define the control group

```{r}
#eset$sample_type <- factor(eset$sample_type)

table(eset$sample_type) # debe ser: control (fct referencia) vs caso
# eset$sample_type <- relevel(eset$sample_type, ref = "sin")
table(eset$sample_type) # debe ser: control (fct referencia) vs caso
```

###### Matrix + Linear model + eBayes

```{r}
# related with manual imputation in 'pheno' object
eset <- eset[,(rowSums(is.na(pData(eset)["sample_type"]))==0)] # new
design <- model.matrix(~ eset$sample_type)
fit <- lmFit(eset, design)
eb <- eBayes(fit)
topTable(eb) %>% rownames_to_column() #%>% filter(adj.P.Val<0.05)#number=Inf
```

```{r}
# tidy toptable
td <- topTable(eb, coef = 2,               # default: slope
               sort.by =NULL, resort.by =NULL, 
               #genelist = NULL,                 # no fit$gene
               number = Inf,                     # show all the hipothesis
               confint=0.95) %>% rownames_to_column() %>% as.tbl() %>% 
  dplyr::rename(ID=rowname) %>% 
  arrange(P.Value) %>% 
  dplyr::mutate(p.order = seq(n())) %>% 
  mutate(significance=if_else(adj.P.Val<0.05,
                              "FDR<0.05","not sig.")) %>% 
  mutate(reactivity=if_else(adj.P.Val<0.05 & logFC>0,
                              "reactive","not_reactive")) %>% 
  #dplyr::count(significance)
  inner_join(biobroom::tidy.ExpressionSet(eset,addPheno = TRUE) %>% 
               #dplyr::count(sample_type)
               group_by(gene,sample_type) %>% 
               dplyr::summarise(mean_group=mean(value)) %>% 
               ungroup() %>% 
               spread(sample_type,mean_group) %>% 
               #arrange(desc(con)) %>% 
               dplyr::rename(ID=gene),
             by="ID") %>% 
  relocate(reactivity,.after = last_col())
```

###### missings?

```{r, eval=FALSE}
td %>% epihelper::print_inf()
td_missing_p <- td %>% 
  filter(is.na(P.Value)) %>% 
  arrange(desc(AveExpr)) %>% 
  select(ID) %>% 
  pull()
  # glimpse()
expression_set %>% 
  # filter(feature=="PVX_083560_2o2.1028") %>% 
  filter(magrittr::is_in(feature,td_missing_p)) %>% 
  naniar::vis_miss()
  # ggplot(aes(x = values, y = feature, color = sample_type)) +
  # geom_point(alpha = 0.5)
```

###### extreme negative?

```{r}
eb_negative <- biobroom::tidy.MArrayLM(eb) %>% 
  filter(estimate<0) %>% 
  pull(gene)
# biobroom::glance.MArrayLM(eb)
# biobroom::augment.MArrayLM(eb)

expression_set_long %>% 
  # filter(feature=="PVX_123520.241") %>%
  filter(magrittr::is_in(feature,eb_negative)) %>% 
  # filter(magrittr::is_in(feature,td_missing_p)) %>% 
  # naniar::vis_miss()
  ggplot(aes(x = values, y = feature, color = sample_type)) +
  geom_point(alpha=0.5)
```

##### top reactive

###### all reactives

```{r}
td %>% glimpse()
td %>% count(reactivity)
td %>% write_rds("data/06-reactive_compare-all_reactives.rds")
```


###### top10% reactives

```{r}
exp_t <- td %>% 
  # arrange(desc(AveExpr)) %>% 
  dplyr::arrange(desc(sample)) %>% 
  select(-feature_specie,-t,-B,-p.order,-significance,-AveExpr) %>% 
  # select(ID,#Gene.ID,Description,
  #        AveExpr,p.order,logFC) %>% 
  slice(1:16) #%>% 
  # inner_join(all, by="Gene.ID") %>% 
  # inner_join(lg %>% dplyr::rename(seg.num=n), by="Gene.ID")

exp_t
writexl::write_xlsx(x = exp_t,
                    path = "table/06-tab01-reactive_compare-top_reactives.xlsx")
```



##### Plots

###### p-value histogram

```{r, fig.height=3, fig.width=6}
reactive_antigen_a <- td %>% 
  ggplot(aes(P.Value)) +
  geom_histogram(breaks=seq(0, 1, by = .05)) +
  scale_x_continuous(breaks=seq(0, 1, by = .5)) +
  geom_vline(aes(xintercept=0.05), lty=3, col="red") +
  labs(#title="P-value histogram",
       #subtitle="Without multiple testing correction"
       ) + 
  xlab("p value") + 
  ylab("Frequency") +
  theme_bw()
# bx <-  td %>% ggplot(aes(adj.P.Val)) +
#   geom_histogram(breaks=seq(0, 1, by = .05)) +
#   scale_x_continuous(breaks=seq(0, 1, by = .5)) +
#   geom_vline(aes(xintercept=0.05), lty=3, col="red") +
#   coord_cartesian(xlim = c(0,1)) +
#   labs(title="P-value histogram",
#        subtitle="After multiple testing correction") + xlab("adjust p.value") + theme_bw()
# Rmisc::multiplot(ax,bx,cols = 2)
reactive_antigen_a#+bx
ggsave("figure/06-fig01-reactive_compare-pvalue_histogram.png",
       height = 3.5,width = 3.5,dpi = "retina")
```

###### volcano plot

```{r, fig.height=6, fig.width=3.4}
reactive_antigen_b <- td %>% 
  mutate(significance=fct_recode(significance,
                                 "<0.05"="FDR<0.05",
                                 "not\nsignificant"="not sig.")) %>%
  ggplot(aes(logFC,-log10(P.Value))) +
  geom_point(aes(colour=significance), alpha =0.75) + 
  colorspace::scale_color_discrete_diverging(palette = "Blue-Red 2",
                                             rev = TRUE) +
  # scale_color_manual(values=c("red","black")) +
  #geom_point(aes(colour=con)) + viridis::scale_color_viridis() +
  # ggrepel::geom_text_repel(
  #   data = td %>% 
  #     filter(logFC < -1),
  #   aes(label=ID)
  # ) +
  # ggrepel::geom_text_repel(
  #   data = td %>% 
  #     filter((adj.P.Val<0.05 & logFC>1 #& con>1
  #             )# | 
  #            #(adj.P.Val<0.001 & #logFC>0.9 & 
  #            #   con>2.7) 
  #     ) %>% 
  #     inner_join(all, by="Gene.ID") %>% 
  #     inner_join(lg %>% dplyr::rename(seg.num=n), by="Gene.ID") #%>% 
  #   #arrange(desc(con)) %>% 
  #   #top_n(10,con)
  #   ,
  #   aes(label=Gene.Name), #force = 15,
  #   direction = "x",
  #   box.padding = unit(4, "lines"),
  #   point.padding = unit(.5, "lines")
  # ) + 
  theme_bw() +
  #theme(legend.position="bottom") +
  # theme(#legend.position=c(.18, .44),
  #   legend.position=c(.17, .92),
  #   legend.margin = margin(2,2,2,2)) +
  labs(x = expression(log[2]~fold~change),
       y = expression(-log[10](p~value)), 
       color = "BH\nadjusted\np value") +
  # labs(
  #   # title="With vs Without previous episode",
  #   #    #title="Con vs Sin episodio previo",
  #   #    subtitle="Effect size against statistical significance"#,
  #   #    #subtitle="Tamaño del efecto y significancia estadística",
  #   #    #caption="annotated features filtered by\n(adj.P.Val<0.05 & con>1 & logFC>1.5)"
  # )+
  # or (adj.P.Val<0.001 & con>AveExpr(top1))
  geom_hline(aes(yintercept=-log10(0.05)), lty=3)+
  geom_vline(aes(xintercept=1), lty=3) #+
  #geom_vline(aes(xintercept=-1), lty=3) +
  # coord_cartesian(xlim = c(-2, 2), ylim = c(0,15)) 
reactive_antigen_b
ggsave("figure/06-fig02-reactive_compare-volcano_plot.png",
       height = 3.5,width = 3.5,dpi = "retina")
```

```{r}
reactive_antigen_a +
  reactive_antigen_b +
  # plot_layout(ncol = 2,nrow = 2) +
  plot_annotation(tag_levels = "A")
ggsave("figure/06-fig03-reactive_compare-histogram_volcano.png",
       height = 3.5,width = 6.5,dpi = "retina")
```


###### heatmap

```{r}
# tidy eset
eset_tidy <- biobroom::tidy.ExpressionSet(eset,addPheno = TRUE) %>% 
  rename(id.name = gene)

# summarize samples
eset_tidy_arrange <- eset_tidy %>%
  select(id.name,value,sample) %>% 
  group_by(sample) %>% 
  dplyr::summarise(sum_value = median(value,na.rm = TRUE)) %>% 
  ungroup() %>% 
  arrange(sum_value)

# summarize antigens
eset_tidy_antigen <- eset_tidy %>% 
  filter(sample_type=="sample") %>% 
  select(id.name,value) %>% 
  group_by(id.name) %>% 
  dplyr::summarise(sum_antigen = mean(value,na.rm = TRUE)) %>% 
  ungroup() %>% 
  arrange(sum_antigen)

eset_tidy_antigen %>% arrange(desc(sum_antigen))
```

```{r}
# join in one dataset
pre_heatmap_all <- eset_tidy %>% 
  # add labels for each gene code and common name
  mutate(id.name_ag = str_replace(id.name,"(.{10})(.+)","\\1")) %>% 
  left_join(all_listgen %>% select(id_name,gene_id),
            by = c("id.name_ag"="gene_id")) %>% 
  # join columns useful to reorder row and columns
  left_join(eset_tidy_arrange) %>% 
  left_join(eset_tidy_antigen) %>%
  # directly add values from summary table
  # left_join(td %>% select(id.name = ID, average_sample = sample)) %>% 
  mutate(sample = fct_reorder(.f = sample,
                               .x = sum_value,
                               .fun = mean),
         id.name = fct_reorder(.f = id.name,
                               .x = sum_antigen,
                               .fun = mean)) %>% 
  # extra step to add order to column with code-names
  mutate(
    id_name = fct_reorder(.f = id_name,
                          .x = sum_antigen,
                          #.x = average_sample,
                          .fun = mean)
  ) %>% 
  mutate(group =case_when(
    magrittr::is_in(id.name,exp_t %>% pull(ID)) ~ "top",
    TRUE ~ "rest"
  )) %>% 
  mutate(group = as.factor(group)) %>% 
  mutate(group = fct_rev(group)) %>% 
  #only plot the differentially reactive antigens
  filter(
    magrittr::is_in(id.name,
                    td %>% 
                      filter(reactivity!="not_reactive") %>% 
                      pull(ID))
  ) 

```

```{r}
# crear heatmap plot
reactive_heatmap <- pre_heatmap_all %>% 
  ggplot(aes(x = sample,y = id.name, fill = value)) +
  geom_tile() +
  facet_grid(group~sample_type,
             scale = 'free', space = 'free') +
  colorspace::scale_fill_continuous_diverging(
    name="Antigen\nreactivity",
    mid = 1,
    palette = "Blue-Red 3",
    # begin = 0.2
    ) +
  labs(y = NULL, x=NULL#,
       #caption="*BH adjusted p-value <0.05 and log2 fold-change >1"
       ) + 
  theme(axis.text.x = element_text(angle = 90, 
                                   vjust = 0.5, 
                                   hjust=1,
                                   size=5),
        axis.text.y = element_text(size=1))

reactive_heatmap
ggsave("figure/06-fig04-reactive_compare-heatmap-nocluster.png",
       height = 6,width = 6,dpi = "retina")
```

```{r}
# all_listgen %>% glimpse()
# crear heatmap plot (top subset)
reactive_heatmap_top <- pre_heatmap_all %>% 
  # only top antigens
  filter(group=="top") %>% 
  # replace one category
  mutate(group=fct_recode(group,"Top 10%"="top")) %>% 
  mutate(sample_type=fct_recode(sample_type,
                                "P. vivax malaria"="sample",
                                "US-control"="control")) %>% 
  # try to reorder
  mutate(
    id_name = fct_reorder(.f = id_name,
                          .x = sum_antigen,
                          #.x = average_sample,
                          .fun = mean)
  ) %>% 
  arrange(id_name) %>% 
  # make plot
  ggplot(aes(x = sample,y = id_name, fill = value)) +
  geom_tile() +
  facet_grid(group~sample_type,
             scale = 'free', space = 'free') +
  colorspace::scale_fill_continuous_diverging(
    name="Antigen\nreactivity",
    mid = 1,
    palette = "Blue-Red 3",
    # begin = 0.2
    ) +
  labs(y = NULL, x=NULL#,
       #caption="*BH adjusted p-value <0.05 and log2 fold-change >1"
       ) + 
  theme(axis.text.x = element_text(angle = 90, 
                                   vjust = 0.5, 
                                   hjust=1,
                                   size=5),
        axis.text.y = element_text(size=7))

reactive_heatmap_top
ggsave("figure/06-fig05-reactive_compare-heatmap-nocluster-top.png",
       height = 2.5,width = 8,dpi = "retina")
```

## contabilizar antigenos

```{r}
# 515 features originales
eset.VIVAX.SEV_import
# 394 features filtrados
eset.VIVAX.SEV.FILTER_import
# 340 significativos
td %>% count(reactivity)
```

## compare diversity - selection pressure

```{r, eval=FALSE}
# De forma transitoria, descargué la info de P01 para todos sus genes y 
# la empareje con genes de Sal1 usando grupos ortólogos con 17% de missings
# partir de aqui
# aprovechar para comparar genes con snp counts de sal1 y p01
# usar como denominador
# comparar
# - numero de snps
# - numero de snp sinonimos
# - numero de snp no sinonimos
pv_p01_sal1 <- readr::read_rds("data/20211223-pvivax-sal1-p01-snp_variant_solution.rds")

td %>% 
  select(id=ID,reactivity) %>% 
  mutate(gene_id_sal = str_replace(id,"(PVX_.{6}).+","\\1")) %>% 
  # count(str_length(gene_id_sal)) %>% 
  left_join(pv_p01_sal1) %>% 
  # naniar::vis_miss()
  # naniar::miss_var_summary()
  filter(is.na(gene_id_p01)) %>% 
  epihelper::print_inf()
```


```{r}
all_listgen_snp <- all_listgen %>% 
  # glimpse()
  select(id_name,gene_id,
         non_syn_syn_snp_ratio_all_strains,
         total_sn_ps_all_strains,
         transcript_length) %>% 
  mutate(
    non_syn_syn_snp_ratio_length_norm_e5 = (non_syn_syn_snp_ratio_all_strains/transcript_length)*100000,
    total_snps_length_norm_e3 = (total_sn_ps_all_strains/transcript_length)*1000
  ) %>% 
  mutate(non_syn_snp_number = 
           total_sn_ps_all_strains - 
           (total_sn_ps_all_strains / (1+non_syn_syn_snp_ratio_all_strains))) %>% 
  mutate(non_syn_snp_number_length_norm_e3 = 
           (non_syn_snp_number/transcript_length)*1000)

all_listgen_snp %>% 
  filter(str_starts(gene_id,"PV")) %>% 
  pivot_longer(
    cols = -c(id_name:gene_id)) %>% 
  ggplot(aes(x = value)) +
  geom_histogram() +
  facet_wrap(~name,scales = "free")

ag_selection <- td %>% 
  janitor::clean_names() %>% 
  select(id.name = id, sample, reactivity) %>% 
  mutate(id.name_ag = str_replace(id.name,"(.{10})(.+)","\\1")) %>% 
  left_join(all_listgen_snp %>% 
              select(-transcript_length),
            by = c("id.name_ag"="gene_id")) %>% 
  mutate(category = case_when(
    magrittr::is_in(id.name,exp_t %>% pull(ID)) ~ "reactive_top",
    reactivity =="reactive" ~ "reactive",
    TRUE ~ "non_reactive"
    ))
```

```{r}
ag_selection %>% 
  pivot_longer(
    cols = -c(id.name:id_name,category)) %>% 
  ggplot(aes(x = value, fill = reactivity)) +
  geom_histogram(alpha = 0.25, position = position_identity()) +
  facet_wrap(~name,scales = "free")

ag_selection %>% 
  pivot_longer(
    cols = -c(id.name:id_name,category)) %>% 
  ggplot(aes(x = value, fill = category)) +
  geom_histogram(alpha = 0.25, position = position_identity()) +
  facet_wrap(~name,scales = "free")
```


```{r}
ag_selection %>% 
  # glimpse()
  # select(id.name,reactivity,non_syn_syn_snp_ratio_all_strains) %>% 
  compareGroups::compareGroups(
    formula = 
      reactivity ~ non_syn_syn_snp_ratio_all_strains + 
      non_syn_syn_snp_ratio_length_norm_e5 +
      total_sn_ps_all_strains +
      total_snps_length_norm_e3 +
      non_syn_snp_number +
      non_syn_snp_number_length_norm_e3,
    data = .,
    method = list(non_syn_syn_snp_ratio_all_strains = 2,
                  non_syn_syn_snp_ratio_length_norm_e5 = 2,
                  total_sn_ps_all_strains = 2,
                  total_snps_length_norm_e3 = 2,
                  non_syn_snp_number = 2,
                  non_syn_snp_number_length_norm_e3 = 2)) %>% 
  compareGroups::createTable()

ag_selection %>% 
  # glimpse()
  # select(id.name,reactivity,non_syn_syn_snp_ratio_all_strains) %>% 
  compareGroups::compareGroups(
    formula = 
      category ~ non_syn_syn_snp_ratio_all_strains + 
      non_syn_syn_snp_ratio_length_norm_e5 +
      total_sn_ps_all_strains +
      total_snps_length_norm_e3 +
      non_syn_snp_number +
      non_syn_snp_number_length_norm_e3,
    data = .,
    method = list(non_syn_syn_snp_ratio_all_strains = 2,
                  non_syn_syn_snp_ratio_length_norm_e5 = 2,
                  total_sn_ps_all_strains = 2,
                  total_snps_length_norm_e3 = 2,
                  non_syn_snp_number = 2,
                  non_syn_snp_number_length_norm_e3 = 2)) %>% 
  compareGroups::createTable()
```

```{r}
ag_selection %>% 
  select(id.name,reactivity,total_snps_length_norm_e3) %>% 
  mutate(reactivity = as.factor(reactivity)) %>% 
  wilcox.test(total_snps_length_norm_e3 ~ reactivity, 
              data = .) %>% 
  broom::tidy()
```


```{r}

```


pendientes
  - comparar distribución de snp en reactivos vs no-reactivos (con respecto al genoma, no solo a los antigenos del chip)
  - restringir el análisis en 04-abnomics según esta lista de antígenos seleccionados (transformación -> filtrado -> diferencialmente rvo a control)
  

## end

```{r}

```
